### 2021-04-24 #3217 @whedon
[PRE REVIEW]: Collaborative Container Modules with Singularity Registry HPC
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://github.com/singularityhub/singularity-hpc" target ="_blank">https://github.com/singularityhub/singularity-hpc</a>
**Version:** 0.0.22
