+++ paper/paper.bib
@@ -68,6 +68,7 @@
   series    = "SC '11",
   month     =  nov,
   year      =  2011,
+  doi       = "10.1145/2063348.2063360",
   address   = "New York, NY, USA",
   location  = "Seattle, Washington"
 }
@@ -155,7 +156,8 @@
   booktitle={Docker for Data Science},
   pages={119--135},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi="10.1007/978-1-4842-3012-1_7"
 }
 
 
@@ -163,7 +165,8 @@
   title        = "Docker Hub",
   howpublished = "\url{https://hub.docker.com/}",
   note         = "Accessed: 2020-11-23",
-  year      =  2020
+  year      =  2020,
+  doi       = "10.1007/978-1-4842-3012-1_6"
 }
 
 
@@ -175,6 +178,7 @@
   number={3},
   pages={e1005209},
   year={2017},
+  doi="10.1371/journal.pcbi.1005209",
   publisher={Public Library of Science}
 }
 
@@ -277,7 +281,7 @@
   volume    =  2015,
   month     =  feb,
   year      =  2015,
-  doi       = "http://dx.doi.org/10.1155/2015/243180"
+  doi       = "10.1155/2015/243180"
 }
 
 @ARTICLE{Wandell2015-yt,
