+++ _library/bids-mrtrix3_connectome.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/mrtrix3_connectome"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/mrtrix3_connectome/container.yaml"
-updated_at: "2021-04-24 00:29:21.556684"
+updated_at: "2021-04-24 20:01:00.718402"
 container_url: ""
 
 versions:
