+++ _library/ghcr.io-autamus-proj.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/proj"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/proj/container.yaml"
-updated_at: "2021-04-24 00:28:59.743069"
+updated_at: "2021-04-24 20:00:37.925321"
 container_url: ""
 aliases:
  - "proj"
