+++ _library/ghcr.io-autamus-bismark.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bismark"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bismark/container.yaml"
-updated_at: "2021-04-24 00:28:56.475856"
+updated_at: "2021-04-24 20:00:34.518728"
 container_url: ""
 aliases:
  - "bismark"
