+++ _library/ghcr.io-autamus-snappy.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/snappy"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/snappy/container.yaml"
-updated_at: "2021-04-24 00:29:05.262721"
+updated_at: "2021-04-24 20:00:43.696056"
 container_url: ""
 
 versions:
