+++ _library/ghcr.io-autamus-glpk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/glpk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/glpk/container.yaml"
-updated_at: "2021-04-24 00:29:18.722393"
+updated_at: "2021-04-24 20:00:57.816970"
 container_url: ""
 aliases:
  - "glpsol"
