+++ _library/ghcr.io-autamus-hmmer.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/hmmer"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/hmmer/container.yaml"
-updated_at: "2021-04-24 00:29:12.034734"
+updated_at: "2021-04-24 20:00:50.764711"
 container_url: ""
 aliases:
  - "hmmalign"
