+++ _library/jupyter-datascience-notebook.md
@@ -3,7 +3,7 @@ layout: container
 name:  "jupyter/datascience-notebook"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/jupyter/datascience-notebook/container.yaml"
-updated_at: "2021-04-24 20:00:24.453364"
+updated_at: "2021-04-24 20:40:24.080631"
 container_url: ""
 aliases:
  - "run-notebook"
