+++ _library/biocontainers-samtools.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/samtools"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/samtools/container.yaml"
-updated_at: "2021-04-24 20:00:21.098630"
+updated_at: "2021-04-24 20:40:21.096545"
 container_url: ""
 aliases:
  - "samtools"
