+++ _library/ghcr.io-autamus-libxpm.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/libxpm"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/libxpm/container.yaml"
-updated_at: "2021-04-24 20:40:52.071163"
+updated_at: "2021-04-25 12:09:35.653416"
 container_url: ""
 
 versions:
