+++ _library/ghcr.io-autamus-eagle.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/eagle"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/eagle/container.yaml"
-updated_at: "2021-04-24 20:40:49.801025"
+updated_at: "2021-04-25 12:09:33.037885"
 container_url: ""
 aliases:
  - "eagle"
