+++ _library/singularityhub-singularity-deploy.md
@@ -3,13 +3,12 @@ layout: container
 name:  "singularityhub/singularity-deploy"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/singularityhub/singularity-deploy/container.yaml"
-updated_at: "2021-04-24 20:40:58.471108"
+updated_at: "2021-04-25 12:09:43.057412"
 container_url: ""
 aliases:
  - "salad"
 
 versions:
- - "0.0.12"
  - "salad"
 description: "Example shpc container using Singularity Deploy, build and serve from GitHub releases."
 ---
@@ -25,7 +24,7 @@ $ shpc install singularityhub/singularity-deploy
 Or a specific version:
 
 ```bash
-$ shpc install singularityhub/singularity-deploy:0.0.12
+$ shpc install singularityhub/singularity-deploy:salad
 ```
 
 And then you can tell lmod about your modules folder:
@@ -37,8 +36,8 @@ $ module use ./modules
 And load the module, and ask for help, or similar.
 
 ```bash
-$ module load singularityhub/singularity-deploy/0.0.12
-$ module help singularityhub/singularity-deploy/0.0.12
+$ module load singularityhub/singularity-deploy/salad
+$ module help singularityhub/singularity-deploy/salad
 ```
 
 You can use tab for auto-completion of module names or commands that are provided.
