+++ _library/nvcr.io-nvidia-hpc-benchmarks.md
@@ -3,7 +3,7 @@ layout: container
 name:  "nvcr.io/nvidia/hpc-benchmarks"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/nvcr.io/nvidia/hpc-benchmarks/container.yaml"
-updated_at: "2021-04-24 20:40:26.895176"
+updated_at: "2021-04-25 12:09:06.086981"
 container_url: ""
 
 versions:
