+++ _library/jupyter-r-notebook.md
@@ -3,7 +3,7 @@ layout: container
 name:  "jupyter/r-notebook"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/jupyter/r-notebook/container.yaml"
-updated_at: "2021-04-24 20:40:24.641701"
+updated_at: "2021-04-25 12:09:03.429228"
 container_url: ""
 aliases:
  - "run-notebook"
