+++ registry/rabbitmq/container.yaml
@@ -3,8 +3,9 @@ url: https://hub.docker.com/r/_/rabbitmq
 maintainer: '@vsoch'
 description: RabbitMQ is an open source multi-protocol messaging broker.
 latest:
-  latest: sha256:b8c2a55dd4af3db0affb1c89ac1ea9c89f4d71a1fe46cf53884dd35bc73db6c6
+  3.8.15-rc.2-alpine: sha256:d948ed2ff641ecc3eba7936343d3597fb2e72e1c4a665ab5e93625373d0ca585
 tags:
+  3.8.15-rc.2-alpine: sha256:d948ed2ff641ecc3eba7936343d3597fb2e72e1c4a665ab5e93625373d0ca585
   latest: sha256:b8c2a55dd4af3db0affb1c89ac1ea9c89f4d71a1fe46cf53884dd35bc73db6c6
 aliases:
   rabbitmq-defaults: /opt/rabbitmq/sbin/rabbitmq-defaults
