+++ _library/ghcr.io-autamus-ascent.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/ascent"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/ascent/container.yaml"
-updated_at: "2021-04-25 12:09:30.184220"
+updated_at: "2021-04-25 20:27:51.231514"
 container_url: ""
 
 versions:
