+++ _library/ghcr.io-autamus-bart.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bart"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bart/container.yaml"
-updated_at: "2021-04-25 12:09:17.383293"
+updated_at: "2021-04-25 20:27:40.744537"
 container_url: ""
 aliases:
  - "bart"
