+++ _library/ghcr.io-autamus-scons.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/scons"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/scons/container.yaml"
-updated_at: "2021-04-25 12:09:29.080814"
+updated_at: "2021-04-25 20:27:50.332604"
 container_url: ""
 aliases:
  - "scons"
