+++ _library/ghcr.io-autamus-gcc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/gcc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gcc/container.yaml"
-updated_at: "2021-04-25 12:09:14.280853"
+updated_at: "2021-04-25 20:27:38.156548"
 container_url: ""
 aliases:
  - "c++"
