+++ _library/ghcr.io-autamus-igraph.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/igraph"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/igraph/container.yaml"
-updated_at: "2021-04-25 12:09:21.791515"
+updated_at: "2021-04-25 20:27:44.404734"
 container_url: ""
 
 versions:
