+++ _library/ghcr.io-autamus-bowtie2.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bowtie2"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bowtie2/container.yaml"
-updated_at: "2021-04-25 12:09:33.480551"
+updated_at: "2021-04-25 20:27:53.945538"
 container_url: ""
 aliases:
  - "bowtie2"
