+++ registry/singularityhub/singularity-deploy/container.yaml
@@ -4,8 +4,9 @@ maintainer: '@vsoch'
 description: Example shpc container using Singularity Deploy, build and serve from
   GitHub releases.
 latest:
-  salad: 0.0.12
+  0.0.12: singularityhub/singularity-deploy
 tags:
+  0.0.12: singularityhub/singularity-deploy
   salad: 0.0.12
 aliases:
   salad: /code/salad
