+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:861e35707e91b99f261725e43f2d1c4931d6ee9e942cb43ecb5cb0fd6f3fe276
+  latest: sha256:8b12a331db4c8521a8f8c4d80d2cd827f1f8dd804d9fa31d621eccab932882db
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
