+++ registry/openjdk/container.yaml
@@ -7,7 +7,7 @@ latest:
   16.0.1-buster: sha256:a0168916eecee7337948e66604681a40c973715455cdca50ab519b8da3c20760
 tags:
   16.0.1-buster: sha256:a0168916eecee7337948e66604681a40c973715455cdca50ab519b8da3c20760
-  latest: sha256:3f032cf26beef3150eb2abb5144dfdbfdf9d25cef0580d8b68af9eb0035d47db
+  latest: sha256:fa5414c6c4f8f836308947ae46daf65f2bce39ac01e89edfbe730cd0703dca06
 aliases:
   jar: /usr/bin/jar
   jarsigner: /usr/bin/jarsigner
