+++ _library/ghcr.io-autamus-git.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/git"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/git/container.yaml"
-updated_at: "2021-04-25 20:27:37.790041"
+updated_at: "2021-04-26 20:25:42.177496"
 container_url: ""
 aliases:
  - "git"
