+++ _library/openjdk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "openjdk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/openjdk/container.yaml"
-updated_at: "2021-04-25 20:28:01.710016"
+updated_at: "2021-04-26 20:26:12.126259"
 container_url: ""
 aliases:
  - "jar"
