+++ _library/ghcr.io-autamus-r-seqlogo.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/r-seqlogo"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/r-seqlogo/container.yaml"
-updated_at: "2021-04-25 20:27:45.707447"
+updated_at: "2021-04-26 20:25:51.832739"
 container_url: ""
 
 versions:
