+++ registry/tomcat/container.yaml
@@ -4,7 +4,7 @@ maintainer: '@vsoch'
 description: Apache Tomcat is an open source implementation of the Java Servlet and
   JavaServer Pages technologies
 latest:
-  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:be78a79ee4bee2627cf91ecce42f0b4b5175b7147d5bc9184778accba730ac9d
+  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:ae64a2cc72052228acb020f8a188b7cd036c186f668af510ff0532435be2a64f
 tags:
-  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:be78a79ee4bee2627cf91ecce42f0b4b5175b7147d5bc9184778accba730ac9d
+  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:ae64a2cc72052228acb020f8a188b7cd036c186f668af510ff0532435be2a64f
   latest: sha256:0509684774ac53d8b152f4a50dd92889218a8633c0feddb4b35d092e55cd215d
