+++ registry/nvcr.io/nvidia/tensorflow/container.yaml
@@ -6,9 +6,10 @@ description: TensorFlow is an open-source software library for high-performance 
   a variety of platforms (CPUs, GPUs, TPUs), and from desktops to clusters of servers
   to mobile and edge devices.
 latest:
-  21.03-tf1-py3: sha256:addbbdccb193849d58083154458318b2aa9b7945d9975284291b08b4336415e1
+  21.04-tf1-py3: sha256:fb948e24013e079fbba954dc0ce025d3648d8b12d343cd07d24e7ca14b66ba5b
 tags:
-  21.02-tf1-py3: sha256:addbbdccb193849d58083154458318b2aa9b7945d9975284291b08b4336415e1
-  21.03-tf1-py3: sha256:addbbdccb193849d58083154458318b2aa9b7945d9975284291b08b4336415e1
+  21.02-tf1-py3: sha256:fb948e24013e079fbba954dc0ce025d3648d8b12d343cd07d24e7ca14b66ba5b
+  21.03-tf1-py3: sha256:fb948e24013e079fbba954dc0ce025d3648d8b12d343cd07d24e7ca14b66ba5b
+  21.04-tf1-py3: sha256:fb948e24013e079fbba954dc0ce025d3648d8b12d343cd07d24e7ca14b66ba5b
 filter:
 - 21*
