+++ registry/nvcr.io/hpc/gromacs/container.yaml
@@ -6,12 +6,12 @@ description: GROMACS is a popular molecular dynamics application used to simulat
 latest:
   "2021": sha256:aa095dcdb175e10132a5862204bf91e6f374f72ba9f2360d9ff5c45ae67785fd
 tags:
-  "2020.2": sha256:2b8c453a5fc709f19e48c76d2b8403d1a5248eef2f05c1429df23b77e1c06b33
+  "2020.2": sha256:2d51288e7b0cacc3c33fdde07a2b81a803c243411ccb805e03179fa2554b8daa
   2020.2-arm64: sha256:2b8c453a5fc709f19e48c76d2b8403d1a5248eef2f05c1429df23b77e1c06b33
   "2021": sha256:aa095dcdb175e10132a5862204bf91e6f374f72ba9f2360d9ff5c45ae67785fd
+filter:
+- ^((?!arm).)*$
 aliases:
 - name: python
   command: /usr/bin/python
   options: --nv
-filter: 
-  - "^((?!arm).)*$"
