+++ registry/php/container.yaml
@@ -6,7 +6,7 @@ description: While designed for web development, the PHP scripting language also
 latest:
   8.0.3-alpine: sha256:f32cd843417d9271d1500ed3867922175604a243d8033ad1572d617ad7954171
 tags:
-  7.3.27-zts-alpine3.12: sha256:f32cd843417d9271d1500ed3867922175604a243d8033ad1572d617ad7954171
+  7.3.27-zts-alpine3.12: sha256:33c32435b162a6aa427a506f4d3fd3616ab1897a0211c5787649e9e07b967223
   8.0.3-alpine: sha256:f32cd843417d9271d1500ed3867922175604a243d8033ad1572d617ad7954171
   latest: sha256:4b1016d76de017cb030009c358cad5eb28cfa5856b4f16e933b81605edade881
 aliases:
