+++ registry/ghcr.io/autamus/clingo/container.yaml
@@ -5,7 +5,7 @@ description: An ASP system to ground and solve logic programs.
 latest:
   5.5.0: sha256:8afc0abd02c5501f55d99f2af07e9e59a9b922f0a97172a81f2aab88db2d9960
 tags:
-  5.4.1: sha256:8afc0abd02c5501f55d99f2af07e9e59a9b922f0a97172a81f2aab88db2d9960
+  5.4.1: sha256:c98a480f85a68063dfb9c4cb6eedfbe0cf84a1af737b73ef2d94213fffd91ff5
   5.5.0: sha256:8afc0abd02c5501f55d99f2af07e9e59a9b922f0a97172a81f2aab88db2d9960
   latest: sha256:8afc0abd02c5501f55d99f2af07e9e59a9b922f0a97172a81f2aab88db2d9960
 aliases:
