+++ registry/consul/container.yaml
@@ -6,8 +6,8 @@ description: Consul is a datacenter runtime that provides service discovery, con
 latest:
   1.10.0-beta: sha256:d94fea53aa3d88db24ddfc20e5e100f761173d3c287947a5934f6646a2628944
 tags:
-  "1.7": sha256:d94fea53aa3d88db24ddfc20e5e100f761173d3c287947a5934f6646a2628944
-  1.7.14: sha256:d94fea53aa3d88db24ddfc20e5e100f761173d3c287947a5934f6646a2628944
+  "1.7": sha256:2d91cb12de631aeddcd35770fc9712d5b829dbd075d7e3e0d21e3bb459a2725a
+  1.7.14: sha256:2d91cb12de631aeddcd35770fc9712d5b829dbd075d7e3e0d21e3bb459a2725a
   1.10.0-beta: sha256:d94fea53aa3d88db24ddfc20e5e100f761173d3c287947a5934f6646a2628944
   latest: sha256:0dc30c8081ea3cb4a90c908a571cb4aa469ae9e3d2b27d9fb713031127360e42
 aliases:
