+++ registry/ghcr.io/autamus/cantera/container.yaml
@@ -6,6 +6,6 @@ description: Cantera is an open-source collection of object-oriented software to
 latest:
   2.5.1: sha256:d23871fef4e7ff81e40b5ae4c775fe9a4a3f3c81e4195820e6872e379a0c6237
 tags:
-  2.4.0: sha256:d23871fef4e7ff81e40b5ae4c775fe9a4a3f3c81e4195820e6872e379a0c6237
+  2.4.0: sha256:ad490652d6f83f71cb56b6195f92ec4837671ceafcdb897d66393e564cde9699
   2.5.1: sha256:d23871fef4e7ff81e40b5ae4c775fe9a4a3f3c81e4195820e6872e379a0c6237
   latest: sha256:d23871fef4e7ff81e40b5ae4c775fe9a4a3f3c81e4195820e6872e379a0c6237
