+++ _library/jupyter-r-notebook.md
@@ -3,7 +3,7 @@ layout: container
 name:  "jupyter/r-notebook"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/jupyter/r-notebook/container.yaml"
-updated_at: "2021-04-26 20:25:31.690570"
+updated_at: "2021-04-27 21:45:20.253078"
 container_url: ""
 aliases:
  - "run-notebook"
