+++ _library/biocontainers-bwa.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/bwa"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/bwa/container.yaml"
-updated_at: "2021-04-26 20:25:29.099724"
+updated_at: "2021-04-27 21:45:17.772831"
 container_url: ""
 aliases:
  - "bwa"
