+++ _library/ghcr.io-autamus-grass.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/grass"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/grass/container.yaml"
-updated_at: "2021-04-26 20:25:44.206972"
+updated_at: "2021-04-27 21:45:32.231351"
 container_url: ""
 aliases:
  - "grass78"
