+++ _library/ghcr.io-autamus-ior.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/ior"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/ior/container.yaml"
-updated_at: "2021-04-26 20:26:00.372399"
+updated_at: "2021-04-27 21:45:47.579319"
 container_url: ""
 aliases:
  - "ior"
