### 2021-04-28 #3217 @Kevin-Mattheus-Moerman
@vsoch thanks for this submission. I have a couple of comments/questions on this submission and the paper.

- [ ] JOSS submissions should have a very clear scientific research application. I have to admit this submission is not my domain of expertise but from reading the paper I struggle to find a clear description of the direct scientific research application. Page 3 refers to "Reproducible Science", is this perhaps the core scientific research relevance of this work? It would be best if your first sections directly spell out the relevance of this work to scientific research. After reading the summary and statement of need the reader should clearly/quickly understand how the software impacts scientific research, and why one would feel the need to cite this work when doing research in this domain. I suggest you work on the summary and statement of need section to more clearly reflect the relevance to scientific research. You may also comment here on this to help clarify this to the editors.

- [ ] This paper is rather long. JOSS papers are typically [between 250-1000 words](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Can you work on shortening the paper?
