### 2021-04-28 #3217 @vsoch
Thanks @Kevin-Mattheus-Moerman ! I can definitely shorten the paper, and let's chat about the statement of need first - I am definitely not great at writing these, but creating, managing, and running scientific software is the core of a researcher's work, so the tool is definitely valuable. There is also precedence in JoSS for this exact kind of tool, namely papers that have been published before:

 - [VCC: A framework for building containerizedreproducible cluster software environments](https://joss.theoj.org/papers/10.21105/joss.00208): this is spot on a similar idea, just with a different implementation strategy
 - [slurmR: A lightweight wrapper for HPC with Slurm](https://joss.theoj.org/papers/10.21105/joss.01493): A wrapper for the job manager on HPC, another similar idea (much more simple than what I'm presenting here).

And even a previous paper of mine that is a tool of the same nature, but instead of relying on deploying a docker-compose application, it's all done with technology native to HPC, and more accessible to users (and of course admins) to share and install scientific software:

 - [Singularity Registry: Open Source Registry for Singularity Images](https://joss.theoj.org/papers/10.21105/joss.00426)

And you can see that the project [has been consistently active](https://github.com/singularityhub/sregistry) since I created it, which is a clear reflection of community need. The tool here, shpc, grew out of both admins and researchers asking for something like this for a few years until I finally made some time to do it recently. So let's chat about how to fix this up. I do think the core of the tool is around reproducibility, and making the management, use, and design of scientific software on HPC _much_ easier for the user.  Here is another shot!

------------

> Portability and reproducibility of complex software stacks is essential for researchers to perform their work. High Performance Computing (HPC) environments add another level of complexity, where possibly conflicting dependencies must co-exist. Although container technologies like Singularity make it possible to "bring your own environment," without any form of central strategy to manage containers, researchers that seek reproducibility via using containers are tasked with managing their own container collection, often not taking care to ensure that a particular digest or version is used. The reproducibility of the work is at risk, as they cannot easily install and use containers, nor can they share their software with others.

> Singularity Registry HPC (shpc) is the first of its kind to provide an easy means for a researcher to add their research software for sharing and collaboration with other researchers to an existing collection of over 200 popular scientific libraries. The software installs these containers as environment modules that are easy to use and read documentation for, and exposes aliases for commands in the container that the researcher can add to his or her pipeline without thinking about complex interactions with a container. The simple addition of an entry to the registry maintained by shpc comes down to adding a yaml file, and after doing this, another researcher can easily install the same software, down to the digest, to reproduce the original work.

I'm pinging my co-author @alecbcs for his comments as well! And really sorry that it's too long - that should be easy to fix.
