+++ _library/ghcr.io-autamus-boost.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/boost"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/boost/container.yaml"
-updated_at: "2021-04-27 21:45:43.127629"
+updated_at: "2021-04-28 20:37:57.814791"
 container_url: ""
 
 versions:
