+++ _library/ghcr.io-autamus-gnuplot.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/gnuplot"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gnuplot/container.yaml"
-updated_at: "2021-04-27 21:45:44.589871"
+updated_at: "2021-04-28 20:37:59.223234"
 container_url: ""
 aliases:
  - "gnuplot"
