+++ _library/ghcr.io-autamus-node-js.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/node-js"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/node-js/container.yaml"
-updated_at: "2021-04-27 21:45:32.011683"
+updated_at: "2021-04-28 20:37:47.465823"
 container_url: ""
 aliases:
  - "node"
