+++ _library/ghcr.io-autamus-unixodbc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/unixodbc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/unixodbc/container.yaml"
-updated_at: "2021-04-27 21:45:42.072373"
+updated_at: "2021-04-28 20:37:56.822849"
 container_url: ""
 aliases:
  - "odbc_config"
