+++ _library/biocontainers-emboss.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/emboss"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/emboss/container.yaml"
-updated_at: "2021-04-27 21:45:16.088452"
+updated_at: "2021-04-28 20:37:32.544597"
 container_url: ""
 aliases:
  - "embossdata"
