+++ _library/ghcr.io-autamus-clingo.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/clingo"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/clingo/container.yaml"
-updated_at: "2021-04-27 21:45:39.929776"
+updated_at: "2021-04-28 20:37:54.839582"
 container_url: ""
 aliases:
  - "clingo"
