+++ _library/ghcr.io-autamus-scons.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/scons"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/scons/container.yaml"
-updated_at: "2021-04-27 21:45:45.018917"
+updated_at: "2021-04-28 20:37:59.626024"
 container_url: ""
 aliases:
  - "scons"
