+++ _library/ghcr.io-autamus-hisat2.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/hisat2"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/hisat2/container.yaml"
-updated_at: "2021-04-27 21:45:51.773775"
+updated_at: "2021-04-28 20:38:06.150937"
 container_url: ""
 aliases:
  - "hisat2"
