+++ _library/node.md
@@ -3,7 +3,7 @@ layout: container
 name:  "node"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/node/container.yaml"
-updated_at: "2021-04-28 20:37:37.051657"
+updated_at: "2021-04-29 15:05:10.014423"
 container_url: ""
 aliases:
  - "node"
