+++ _library/ghcr.io-autamus-picard.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/picard"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/picard/container.yaml"
-updated_at: "2021-04-28 20:37:47.265718"
+updated_at: "2021-04-29 15:05:21.889025"
 container_url: ""
 aliases:
  - "picard"
