+++ _library/ghcr.io-autamus-beast2.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/beast2"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/beast2/container.yaml"
-updated_at: "2021-04-28 20:37:57.612891"
+updated_at: "2021-04-29 15:05:34.303567"
 container_url: ""
 aliases:
  - "beast"
