+++ _library/ghcr.io-autamus-cloc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/cloc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/cloc/container.yaml"
-updated_at: "2021-04-28 20:37:57.017181"
+updated_at: "2021-04-29 15:05:33.622526"
 container_url: ""
 aliases:
  - "cloc"
