+++ _library/elasticsearch.md
@@ -3,7 +3,7 @@ layout: container
 name:  "elasticsearch"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/elasticsearch/container.yaml"
-updated_at: "2021-04-28 20:37:31.352591"
+updated_at: "2021-04-29 15:05:03.201465"
 container_url: ""
 aliases:
  - "elasticsearch"
