+++ _library/nvcr.io-nvidia-pytorch.md
@@ -3,7 +3,7 @@ layout: container
 name:  "nvcr.io/nvidia/pytorch"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/nvcr.io/nvidia/pytorch/container.yaml"
-updated_at: "2021-04-28 20:37:39.868270"
+updated_at: "2021-04-29 15:05:13.313148"
 container_url: ""
 
 versions:
