+++ _library/ghcr.io-autamus-fraggenescan.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/fraggenescan"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/fraggenescan/container.yaml"
-updated_at: "2021-04-28 20:37:46.052501"
+updated_at: "2021-04-29 15:05:20.461213"
 container_url: ""
 
 versions:
