+++ registry/mariadb/container.yaml
@@ -5,9 +5,10 @@ description: MariaDB Server is one of the most popular database servers in the w
   It’s made by the original developers of MySQL and guaranteed to stay open source.
   Notable users include Wikipedia, DBS Bank and ServiceNow.
 latest:
-  10.5.9-focal: sha256:36288c675a192bd0a8a99cd6ba0780e31df85f0bfd0cbb204837cd108be3d236
+  10.6.0-focal: sha256:718cb856543ae92a5460997ed57e8ec398f478beb0110fe5efe675f1f14d6bd8
 tags:
   10.5.9-focal: sha256:36288c675a192bd0a8a99cd6ba0780e31df85f0bfd0cbb204837cd108be3d236
+  10.6.0-focal: sha256:718cb856543ae92a5460997ed57e8ec398f478beb0110fe5efe675f1f14d6bd8
   latest: sha256:36288c675a192bd0a8a99cd6ba0780e31df85f0bfd0cbb204837cd108be3d236
 aliases:
   mariabackup: /usr/bin/mariabackup
