+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:7f0444a703db0081677cd13e3fdd92ee567f90e4f2460ac182102bfafab4500f
+  latest: sha256:c7460db8d1fa49b4c72b4670ab7cfe77d27c295f98b32cba8b495432b0ba2d07
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
