+++ _library/ghcr.io-autamus-cowsay.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/cowsay"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/cowsay/container.yaml"
-updated_at: "2021-04-29 15:05:38.980986"
+updated_at: "2021-04-30 08:37:21.674444"
 container_url: ""
 aliases:
  - "cowsay"
