+++ _library/openjdk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "openjdk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/openjdk/container.yaml"
-updated_at: "2021-04-29 15:05:51.242253"
+updated_at: "2021-04-30 08:37:30.590553"
 container_url: ""
 aliases:
  - "jar"
