+++ paper/paper.md
@@ -23,74 +23,30 @@ bibliography: paper.bib
 
 # Summary
 
-Singularity Registry HPC, "shpc" (https://singularity-hpc.readthedocs.io/), is a collaborative framework to easily provide containers as modules for high performance computing (HPC) centers.
-Whether you are a user or a cluster administrator, installing the software
-gives you an easy means to install Singularity containers [@Kurtzer2017-xj] as
-environment modules [@McLay2011-wu]. Unlike registry servers [@SRegistry] that expect a user to pull a container directly, shpc provides a simple command line client to manage 
-this interaction for you, and make container commands easily available
-as command line aliases. With over 200 containers readily avilable
-from biocontainers [@da2017biocontainers] to Nvidia's Container Catalog [@noauthor_undated-kp] 
-to brain imaging data structure apps [@gorgolewski2017bids], to core containers
-on Docker Hub [@noauthor_undated-xn], and software from package managers like Spack [@gamblin2015spack] made available via the Autamus registry [@autamus], with just a few clicks you can quickly assemble your own collection of custom containers modules for yourself or your user base.
+Portability and reproducibility of complex software stacks is essential for researchers to perform their work. High Performance Computing (HPC) environments add another level of complexity, where possibly conflicting dependencies must co-exist. Although container technologies like Singularity [@Kurtzer2017-xj] make it possible to "bring your own environment," without any form of central strategy to manage containers, researchers that seek reproducibility via using containers are tasked with managing their own container collection, often not taking care to ensure that a particular digest or version is used. The reproducibility of the work is at risk, as they cannot easily install and use containers, nor can they share their software with others.
 
-
-## Statement of Need
-
-To use software on a high performance computing (HPC) system, a user typically needs to request it
-to be installed by an administrator, or if they have expertise with container software,
-the user can pull a container from an external registry. Requiring each user to find,
-pull, and interact with containers not only creates redundancy in terms of filesystem space,
-but also is a risk to reproducibility. Although containers are considered reproducible, it's
-uncommon for a user to pull a specific digest, or an exact version of a container. Writing complex commands to control binds, the environment, or even find software in the container adds an additional level of complexity. Ideally, it would be easy for an administrator, a group, or an individual to maintain a _local_ container registry, and one that provides a community maintained collection of versioned containers and aliases for the most popular scientific software. Singularity Registry HPC (shpc) is this local registry, and the first of its kind to offer a means to manage containers natively on an HPC system.
+Singularity Registry HPC (shpc) is the first of its kind to provide an easy means for a researcher to add their research software for sharing and collaboration with other researchers to an existing collection of over 200 popular scientific libraries [@da2017biocontainers, @noauthor_undated-kp, @gorgolewski2017bids, @gamblin2015spack, @autamus]. The software installs these containers as environment modules [@McLay2011-wu] that are easy to use and read documentation for, and exposes aliases for commands in the container that the researcher can add to his or her pipeline without thinking about complex interactions with a container. The simple addition of an entry to the registry maintained by shpc comes down to adding a yaml file, and after doing this, another researcher can easily install the same software, down to the digest, to reproduce the original work.
 
 
 ## Background
 
-Using environment modules on high performance computing clusters is a common
-trend. Although writing the recipes can be complex and beyond the ability of the scientific
-user, it's a fairly common practice for cluster administrators to provide
-a set of natively installed recipes for their users [@noauthor_undated-bt]. The practice is so common
[…]
