+++ _library/ghcr.io-autamus-curl.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/curl"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/curl/container.yaml"
-updated_at: "2021-04-30 08:37:11.361846"
+updated_at: "2021-04-30 20:02:21.397968"
 container_url: ""
 aliases:
  - "c_rehash"
