+++ _library/ghcr.io-autamus-openjdk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/openjdk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/openjdk/container.yaml"
-updated_at: "2021-04-30 08:37:25.888521"
+updated_at: "2021-04-30 20:02:38.479894"
 container_url: ""
 aliases:
  - "jaotc"
