+++ _library/biocontainers-abyss.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/abyss"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/abyss/container.yaml"
-updated_at: "2021-04-30 08:36:58.278818"
+updated_at: "2021-04-30 20:02:05.660981"
 container_url: ""
 aliases:
  - "abyss-fixmate"
