+++ _library/bids-validator.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/validator"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/validator/container.yaml"
-updated_at: "2021-04-30 08:37:28.911957"
+updated_at: "2021-04-30 20:02:41.984632"
 container_url: ""
 aliases:
  - "bids-validator"
