+++ _library/bids-mrtrix3_connectome.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/mrtrix3_connectome"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/mrtrix3_connectome/container.yaml"
-updated_at: "2021-04-30 08:37:28.579977"
+updated_at: "2021-04-30 20:02:41.593490"
 container_url: ""
 
 versions:
