### 2021-04-30 #3217 @vsoch
Okay I was able to get it down to around 1000 words, not including the code snippets (and I checked several papers on the current JoSS home and they are of similar length) so I think it's good for another look!

One thing I'm wondering about - I noticed that when I add multiple references, only the first renders, e.g.,:

```
[@Santana-Perez2015-wo, @Boettiger2014-cz, @Wandell2015-yt]
```
And I think this used to work? Is there a different means to add multiple references / citations to a block?
