+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:c7460db8d1fa49b4c72b4670ab7cfe77d27c295f98b32cba8b495432b0ba2d07
+  latest: sha256:6b39ad471c5f85b9fa175bca058016490e541857371e4df43fd2b179e0098668
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
