+++ _library/ghcr.io-autamus-unixodbc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/unixodbc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/unixodbc/container.yaml"
-updated_at: "2021-04-30 20:02:28.748809"
+updated_at: "2021-05-01 15:22:20.945994"
 container_url: ""
 aliases:
  - "odbc_config"
