+++ _library/ghcr.io-autamus-ascent.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/ascent"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/ascent/container.yaml"
-updated_at: "2021-04-30 20:02:32.545516"
+updated_at: "2021-05-01 15:22:25.003267"
 container_url: ""
 
 versions:
