+++ _library/ghcr.io-autamus-wget.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/wget"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/wget/container.yaml"
-updated_at: "2021-04-30 20:02:24.796652"
+updated_at: "2021-05-01 15:22:16.673819"
 container_url: ""
 aliases:
  - "wget"
