+++ _library/nvcr.io-nvidia-hpc-benchmarks.md
@@ -3,7 +3,7 @@ layout: container
 name:  "nvcr.io/nvidia/hpc-benchmarks"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/nvcr.io/nvidia/hpc-benchmarks/container.yaml"
-updated_at: "2021-04-30 20:02:10.691030"
+updated_at: "2021-05-01 15:22:01.493438"
 container_url: ""
 
 versions:
