+++ registry/jupyter/r-notebook/container.yaml
@@ -3,9 +3,9 @@ url: https://hub.docker.com/r/jupyter/r-notebook
 maintainer: '@vsoch'
 description: Jupyter R Notebook from https://github.com/jupyter/docker-stacks
 latest:
-  latest: sha256:9bf2572ea0e19b344ff0576f7ec90ce7f50a78aa1a3fbd8276774989a27e36fc
+  latest: sha256:7062ab1bc3a1ef5272a345beb5cd8308f2265434efd5a983177e5999fa251f37
 tags:
-  latest: sha256:9bf2572ea0e19b344ff0576f7ec90ce7f50a78aa1a3fbd8276774989a27e36fc
+  latest: sha256:7062ab1bc3a1ef5272a345beb5cd8308f2265434efd5a983177e5999fa251f37
 filter:
 - latest
 aliases:
