+++ registry/jupyter/scipy-notebook/container.yaml
@@ -6,8 +6,6 @@ latest:
   latest: sha256:1741d5fc49e0ab6d574870a73ec8de8e972873a69f768c1dc0f0c727c7aa3011
 tags:
   latest: sha256:1741d5fc49e0ab6d574870a73ec8de8e972873a69f768c1dc0f0c727c7aa3011
-filter:
-- latest
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
