+++ registry/jupyter/datascience-notebook/container.yaml
@@ -6,8 +6,6 @@ latest:
   latest: sha256:827f8694f2c887ca971861ec623ca0363f031352199a3e9633b8f0e020be1d33
 tags:
   latest: sha256:827f8694f2c887ca971861ec623ca0363f031352199a3e9633b8f0e020be1d33
-filter:
-- latest
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
