+++ _library/ghcr.io-autamus-bwa.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bwa"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bwa/container.yaml"
-updated_at: "2021-05-01 15:22:28.042409"
+updated_at: "2021-05-02 16:04:38.681101"
 container_url: ""
 aliases:
  - "bwa"
