+++ _library/mysql.md
@@ -3,7 +3,7 @@ layout: container
 name:  "mysql"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/mysql/container.yaml"
-updated_at: "2021-05-01 15:22:36.909107"
+updated_at: "2021-05-02 16:04:47.657392"
 container_url: ""
 aliases:
  - "mysql"
