+++ _library/quay.io-vgteam-vg.md
@@ -3,7 +3,7 @@ layout: container
 name:  "quay.io/vgteam/vg"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/quay.io/vgteam/vg/container.yaml"
-updated_at: "2021-05-01 15:21:53.716722"
+updated_at: "2021-05-02 16:04:03.618140"
 container_url: ""
 aliases:
  - "vg"
