+++ _library/tensorflow-tensorflow.md
@@ -3,7 +3,7 @@ layout: container
 name:  "tensorflow/tensorflow"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/tensorflow/tensorflow/container.yaml"
-updated_at: "2021-05-01 15:22:04.910071"
+updated_at: "2021-05-02 16:04:15.160460"
 container_url: ""
 aliases:
  - "python"
