+++ _library/ghcr.io-autamus-eagle.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/eagle"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/eagle/container.yaml"
-updated_at: "2021-05-01 15:22:27.814455"
+updated_at: "2021-05-02 16:04:38.463450"
 container_url: ""
 aliases:
  - "eagle"
