+++ _library/ghcr.io-autamus-libpng.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/libpng"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/libpng/container.yaml"
-updated_at: "2021-05-01 15:22:17.497144"
+updated_at: "2021-05-02 16:04:27.870141"
 container_url: ""
 aliases:
  - "libpng-config"
