+++ _library/biocontainers-bwa.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/bwa"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/bwa/container.yaml"
-updated_at: "2021-05-02 16:04:06.289733"
+updated_at: "2021-05-03 09:51:14.846412"
 container_url: ""
 aliases:
  - "bwa"
