+++ _library/ghcr.io-autamus-plink.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/plink"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/plink/container.yaml"
-updated_at: "2021-05-02 16:04:25.030136"
+updated_at: "2021-05-03 09:51:31.885852"
 container_url: ""
 aliases:
  - "plink"
