+++ _library/nvcr.io-nvidia-pytorch.md
@@ -3,7 +3,7 @@ layout: container
 name:  "nvcr.io/nvidia/pytorch"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/nvcr.io/nvidia/pytorch/container.yaml"
-updated_at: "2021-05-02 16:04:12.771874"
+updated_at: "2021-05-03 09:51:20.757132"
 container_url: ""
 
 versions:
