+++ registry/jupyter/datascience-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:5e5bf78bfa351c0255cea5269a0461afbf6b50b51e923a7436229208ea8487f9
 tags:
   "4.0": sha256:5e5bf78bfa351c0255cea5269a0461afbf6b50b51e923a7436229208ea8487f9
-  latest: sha256:827f8694f2c887ca971861ec623ca0363f031352199a3e9633b8f0e020be1d33
+  latest: sha256:4ff678f82494b3d90ccd3136ce52eea05204bb8597f0ca79c3741f54fa353fc1
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
