+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:51898986a4e8b79e384fc99cacd190bf2ea0be76eb1f13c3b1dff43df6f3123a
+  latest: sha256:21649696684395d88e98f84bd15bf0dfee98a7468e651cece8de9e1b56377f94
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
