+++ _library/spack-ubuntu-bionic.md
@@ -3,7 +3,7 @@ layout: container
 name:  "spack/ubuntu-bionic"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/spack/ubuntu-bionic/container.yaml"
-updated_at: "2021-05-03 09:51:52.213056"
+updated_at: "2021-05-05 00:27:01.844951"
 container_url: ""
 aliases:
  - "sbang"
