+++ _library/biocontainers-bowtie2.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/bowtie2"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/bowtie2/container.yaml"
-updated_at: "2021-05-03 09:51:15.050294"
+updated_at: "2021-05-05 00:26:24.190386"
 container_url: ""
 aliases:
  - "bowtie2"
