+++ _library/ghcr.io-autamus-grass.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/grass"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/grass/container.yaml"
-updated_at: "2021-05-03 09:51:28.487155"
+updated_at: "2021-05-05 00:26:37.747516"
 container_url: ""
 aliases:
  - "grass78"
