+++ _library/ghcr.io-autamus-htslib.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/htslib"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/htslib/container.yaml"
-updated_at: "2021-05-03 09:51:34.346965"
+updated_at: "2021-05-05 00:26:43.647991"
 container_url: ""
 aliases:
  - "htsfile"
