+++ registry/ghcr.io/autamus/cloc/container.yaml
@@ -4,9 +4,10 @@ maintainer: '@vsoch'
 description: cloc is a command line program that takes file, directory, and/or archive
   names as inputs.
 latest:
-  "1.88": sha256:d83cabface35c70df9484dff3f606a10f13432747a7570e238525e3722061c31
+  "1.90": sha256:3e298409b923bbd4baced50b0b8a6a758119a0b2b693594a956a05994ebb3146
 tags:
   "1.88": sha256:d83cabface35c70df9484dff3f606a10f13432747a7570e238525e3722061c31
-  latest: sha256:d83cabface35c70df9484dff3f606a10f13432747a7570e238525e3722061c31
+  "1.90": sha256:3e298409b923bbd4baced50b0b8a6a758119a0b2b693594a956a05994ebb3146
+  latest: sha256:3e298409b923bbd4baced50b0b8a6a758119a0b2b693594a956a05994ebb3146
 aliases:
   cloc: /opt/view/bin/cloc
