+++ registry/jupyter/minimal-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:6f9d189494fe091700681eb39b429dfe1420ad0717697c2e2293779fcb80cd1e
 tags:
   "4.0": sha256:6f9d189494fe091700681eb39b429dfe1420ad0717697c2e2293779fcb80cd1e
-  latest: sha256:69cf18db122a2dfc849630b7c1d96eef5dd5aa9da820a88442ff6ced9c31e193
+  latest: sha256:13ee26cdee99d17d08a644ba46951c335b3ce09ee4d0b46bd4a13c3198b78f7d
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
