+++ _library/jupyter-r-notebook.md
@@ -3,7 +3,7 @@ layout: container
 name:  "jupyter/r-notebook"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/jupyter/r-notebook/container.yaml"
-updated_at: "2021-05-05 00:26:26.496952"
+updated_at: "2021-05-05 15:43:24.735425"
 container_url: ""
 aliases:
  - "run-notebook"
