+++ _library/ghcr.io-autamus-povray.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/povray"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/povray/container.yaml"
-updated_at: "2021-05-05 00:26:32.570617"
+updated_at: "2021-05-05 15:43:17.180365"
 container_url: ""
 aliases:
  - "povray"
