+++ _library/biocontainers-bedtools.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/bedtools"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/bedtools/container.yaml"
-updated_at: "2021-05-05 00:26:23.579540"
+updated_at: "2021-05-05 15:43:32.433065"
 container_url: ""
 aliases:
  - "bedtools"
