+++ _library/biocontainers-tabix.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/tabix"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/tabix/container.yaml"
-updated_at: "2021-05-05 00:26:25.224504"
+updated_at: "2021-05-05 15:43:31.680683"
 container_url: ""
 aliases:
  - "tabix"
