+++ _library/ghcr.io-autamus-gcc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/gcc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gcc/container.yaml"
-updated_at: "2021-05-05 00:26:36.312106"
+updated_at: "2021-05-05 15:43:04.681274"
 container_url: ""
 aliases:
  - "c++"
