+++ _library/ghcr.io-autamus-openbabel.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/openbabel"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/openbabel/container.yaml"
-updated_at: "2021-05-05 00:26:34.185077"
+updated_at: "2021-05-05 15:43:19.373606"
 container_url: ""
 aliases:
  - "obabel"
