+++ registry/jupyter/pyspark-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:2aa9a0183bcfe80cd9e2a6c0a52d797bd5b67d1e2db9437c7345c2404b5a86c1
 tags:
   "4.0": sha256:2aa9a0183bcfe80cd9e2a6c0a52d797bd5b67d1e2db9437c7345c2404b5a86c1
-  latest: sha256:c879a07561c3efa70d8a8033cad0d2abc607277634a9578027208494c6341fd3
+  latest: sha256:146b72e090cd1910ecca027f10d75d14f4823f0b7b705f929a584db5264d3f40
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
