+++ registry/jupyter/minimal-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:6f9d189494fe091700681eb39b429dfe1420ad0717697c2e2293779fcb80cd1e
 tags:
   "4.0": sha256:6f9d189494fe091700681eb39b429dfe1420ad0717697c2e2293779fcb80cd1e
-  latest: sha256:13ee26cdee99d17d08a644ba46951c335b3ce09ee4d0b46bd4a13c3198b78f7d
+  latest: sha256:01bc441741b46465244bf43153cd9f70961b21d1f2eb28817f92a9dd4d0e5e35
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
