+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:2ccb6490114dfb1e46d5cb4b3e630d50cf4a0686ca3b2b8d96441bc5cca0caa5
+  latest: sha256:f47189f9a68dceb96c4561471afcbd4d4d54d72e4e970c237e70ce5a2c965ddd
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
