+++ _library/ghcr.io-autamus-bcftools.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bcftools"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bcftools/container.yaml"
-updated_at: "2021-05-05 15:43:14.736523"
+updated_at: "2021-05-06 12:52:27.500123"
 container_url: ""
 aliases:
  - "bcftools"
