+++ _library/bids-baracus.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/baracus"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/baracus/container.yaml"
-updated_at: "2021-05-05 15:42:57.251358"
+updated_at: "2021-05-06 12:52:04.376469"
 container_url: ""
 
 versions:
