+++ _library/ghcr.io-autamus-mercurial.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/mercurial"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/mercurial/container.yaml"
-updated_at: "2021-05-05 15:43:14.906280"
+updated_at: "2021-05-06 12:52:27.728531"
 container_url: ""
 aliases:
  - "hg"
