+++ _library/nvcr.io-hpc-lammps.md
@@ -3,7 +3,7 @@ layout: container
 name:  "nvcr.io/hpc/lammps"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/nvcr.io/hpc/lammps/container.yaml"
-updated_at: "2021-05-05 15:43:28.367257"
+updated_at: "2021-05-06 12:52:44.719105"
 container_url: ""
 
 versions:
