+++ _library/ghcr.io-autamus-clingo.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/clingo"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/clingo/container.yaml"
-updated_at: "2021-05-05 15:43:10.446976"
+updated_at: "2021-05-06 12:52:21.746226"
 container_url: ""
 aliases:
  - "clingo"
