+++ registry/rabbitmq/container.yaml
@@ -3,11 +3,11 @@ url: https://hub.docker.com/r/_/rabbitmq
 maintainer: '@vsoch'
 description: RabbitMQ is an open source multi-protocol messaging broker.
 latest:
-  3.8.16-alpine: sha256:01dbc8ecd56e2d19d01ec1d7cce173df43ab1c6cc2d5bc59fd2f5cc35cfa5e80
+  3.8.16-alpine: sha256:db965b78f81597106d9c5d0a3f62fda483ac855a00539f37bb24d25e0db9639e
 tags:
   3.8.15-rc.2-alpine: sha256:56682567544860464b89625edbd28b524ca12aa5cae4e76ee9e5b887a2db7a06
-  3.8.16-alpine: sha256:01dbc8ecd56e2d19d01ec1d7cce173df43ab1c6cc2d5bc59fd2f5cc35cfa5e80
-  latest: sha256:621e15a95c3f5c405363094fd1ffdf1b6611592aeef5efb60a39a909ba3923f9
+  3.8.16-alpine: sha256:db965b78f81597106d9c5d0a3f62fda483ac855a00539f37bb24d25e0db9639e
+  latest: sha256:9a07dca172e10581e7220f773049b42217452733cf5cc167cedc316e463a8553
 aliases:
   rabbitmq-defaults: /opt/rabbitmq/sbin/rabbitmq-defaults
   rabbitmq-diagnostics: /opt/rabbitmq/sbin/rabbitmq-diagnostics
