+++ registry/golang/container.yaml
@@ -6,8 +6,6 @@ latest:
   latest: sha256:07a471a939cd942ed5c23714185dac476df214f94cf073ef2fbfa748d4926a48
 tags:
   latest: sha256:07a471a939cd942ed5c23714185dac476df214f94cf073ef2fbfa748d4926a48
-filter:
-- latest
 aliases:
   go: /usr/local/go/bin/go
   gofmt: /usr/local/go/bin/gofmt
