+++ registry/jupyter/pyspark-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:2aa9a0183bcfe80cd9e2a6c0a52d797bd5b67d1e2db9437c7345c2404b5a86c1
 tags:
   "4.0": sha256:2aa9a0183bcfe80cd9e2a6c0a52d797bd5b67d1e2db9437c7345c2404b5a86c1
-  latest: sha256:146b72e090cd1910ecca027f10d75d14f4823f0b7b705f929a584db5264d3f40
+  latest: sha256:c145d9c3500876c34c54400126481e584f4cdf09587c2a84c97d2afb6aa640e9
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
