+++ _library/ghcr.io-autamus-circos.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/circos"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/circos/container.yaml"
-updated_at: "2021-05-06 12:52:10.032507"
+updated_at: "2021-05-07 08:58:39.552801"
 container_url: ""
 aliases:
  - "circos"
