+++ _library/ghcr.io-autamus-protobuf.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/protobuf"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/protobuf/container.yaml"
-updated_at: "2021-05-06 12:52:10.935365"
+updated_at: "2021-05-07 08:58:40.488169"
 container_url: ""
 aliases:
  - "protoc"
