+++ registry/php/container.yaml
@@ -4,12 +4,13 @@ maintainer: '@vsoch'
 description: While designed for web development, the PHP scripting language also provides
   general-purpose use.
 latest:
-  8.0.5-alpine: sha256:b728c7e60d1bdb28bfba1bbd4ac3328fbde3a70a959a55c318ccae3fc4d170cf
+  8.0.6-alpine: sha256:c6335954fd836e28161b662af8b44fb382f41aeae7090359285c57b207117a21
 tags:
   7.3.27-zts-alpine3.12: sha256:33c32435b162a6aa427a506f4d3fd3616ab1897a0211c5787649e9e07b967223
   8.0.3-alpine: sha256:f32cd843417d9271d1500ed3867922175604a243d8033ad1572d617ad7954171
   8.0.5-alpine: sha256:b728c7e60d1bdb28bfba1bbd4ac3328fbde3a70a959a55c318ccae3fc4d170cf
-  latest: sha256:c3b4938968e6a7b6245c9536203e6eaa5c7552d3673cd2fb92c349772737b313
+  8.0.6-alpine: sha256:c6335954fd836e28161b662af8b44fb382f41aeae7090359285c57b207117a21
+  latest: sha256:286ecf973b3e69e4bcf9a9841ce8c06ded53520d4406ebc76c903f6117b5d4ca
 aliases:
   php: /usr/local/bin/php
   php-cgi: /usr/local/bin/php-cgi
