+++ _library/ghcr.io-autamus-grass.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/grass"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/grass/container.yaml"
-updated_at: "2021-05-07 08:58:45.877365"
+updated_at: "2021-05-08 15:44:42.261735"
 container_url: ""
 aliases:
  - "grass78"
