+++ _library/ghcr.io-autamus-gnuplot.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/gnuplot"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gnuplot/container.yaml"
-updated_at: "2021-05-07 08:59:02.562665"
+updated_at: "2021-05-08 15:44:54.833655"
 container_url: ""
 aliases:
  - "gnuplot"
