+++ _library/ghcr.io-autamus-intel-mkl.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/intel-mkl"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/intel-mkl/container.yaml"
-updated_at: "2021-05-07 08:58:55.103280"
+updated_at: "2021-05-08 15:44:49.191116"
 container_url: ""
 
 versions:
