+++ _library/bids-brainsautoworkup.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/brainsautoworkup"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/brainsautoworkup/container.yaml"
-updated_at: "2021-05-07 08:58:35.372317"
+updated_at: "2021-05-08 15:44:34.307912"
 container_url: ""
 
 versions:
