+++ _library/jupyter-scipy-notebook.md
@@ -3,7 +3,7 @@ layout: container
 name:  "jupyter/scipy-notebook"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/jupyter/scipy-notebook/container.yaml"
-updated_at: "2021-05-07 08:59:10.423275"
+updated_at: "2021-05-08 15:45:00.516990"
 container_url: ""
 aliases:
  - "run-notebook"
