+++ _library/ghcr.io-autamus-bowtie2.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bowtie2"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bowtie2/container.yaml"
-updated_at: "2021-05-07 08:58:43.034333"
+updated_at: "2021-05-08 15:44:40.139436"
 container_url: ""
 aliases:
  - "bowtie2"
