+++ _library/nvcr.io-nvidia-pytorch.md
@@ -3,8 +3,8 @@ layout: container
 name:  "nvcr.io/nvidia/pytorch"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/nvcr.io/nvidia/pytorch/container.yaml"
-updated_at: "2021-05-08 15:45:03.148804"
-container_url: ""
+updated_at: "2021-05-08 17:35:26.675456"
+container_url: "https://ngc.nvidia.com/catalog/containers/nvidia:pytorch/tags"
 
 versions:
  - "21.02-py3"
