+++ _library/biocontainers-samtools.md
@@ -3,8 +3,8 @@ layout: container
 name:  "biocontainers/samtools"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/samtools/container.yaml"
-updated_at: "2021-05-08 15:45:06.363710"
-container_url: ""
+updated_at: "2021-05-08 17:35:30.612241"
+container_url: "https://hub.docker.com/r/biocontainers/samtools"
 aliases:
  - "samtools"
 
