+++ _library/ghcr.io-autamus-ascent.md
@@ -3,8 +3,8 @@ layout: container
 name:  "ghcr.io/autamus/ascent"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/ascent/container.yaml"
-updated_at: "2021-05-08 15:44:56.957986"
-container_url: ""
+updated_at: "2021-05-08 17:35:55.437163"
+container_url: "https://github.com/orgs/autamus/packages/container/package/ascent"
 
 versions:
  - "0.7.0"
