+++ _library/ghcr.io-autamus-qhull.md
@@ -3,8 +3,8 @@ layout: container
 name:  "ghcr.io/autamus/qhull"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/qhull/container.yaml"
-updated_at: "2021-05-08 15:44:44.374779"
-container_url: ""
+updated_at: "2021-05-08 17:35:59.787055"
+container_url: "https://github.com/orgs/autamus/packages/container/package/qhull"
 aliases:
  - "qconvex"
 
