+++ _library/ghcr.io-autamus-fraggenescan.md
@@ -3,8 +3,8 @@ layout: container
 name:  "ghcr.io/autamus/fraggenescan"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/fraggenescan/container.yaml"
-updated_at: "2021-05-08 15:44:48.652594"
-container_url: ""
+updated_at: "2021-05-08 17:36:13.263824"
+container_url: "https://github.com/orgs/autamus/packages/container/package/fraggenescan"
 
 versions:
  - "1.31"
