+++ _library/ghcr.io-autamus-gnuplot.md
@@ -3,8 +3,8 @@ layout: container
 name:  "ghcr.io/autamus/gnuplot"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gnuplot/container.yaml"
-updated_at: "2021-05-08 15:44:54.833655"
-container_url: ""
+updated_at: "2021-05-08 17:35:46.853334"
+container_url: "https://github.com/orgs/autamus/packages/container/package/gnuplot"
 aliases:
  - "gnuplot"
 
