+++ _library/ghcr.io-autamus-gatk.md
@@ -3,8 +3,8 @@ layout: container
 name:  "ghcr.io/autamus/gatk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gatk/container.yaml"
-updated_at: "2021-05-08 15:44:52.054667"
-container_url: ""
+updated_at: "2021-05-08 17:35:57.610438"
+container_url: "https://github.com/orgs/autamus/packages/container/package/gatk"
 aliases:
  - "gatk"
 
