+++ _library/elasticsearch.md
@@ -3,8 +3,8 @@ layout: container
 name:  "elasticsearch"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/elasticsearch/container.yaml"
-updated_at: "2021-05-08 15:44:31.443219"
-container_url: ""
+updated_at: "2021-05-08 17:35:42.312408"
+container_url: "https://hub.docker.com/_/elasticsearch"
 aliases:
  - "elasticsearch"
 
