### 2021-05-08 f4108e3
fixing missing container url in docgen
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
