+++ _library/ghcr.io-autamus-git.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/git"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/git/container.yaml"
-updated_at: "2021-05-08 17:35:51.382770"
+updated_at: "2021-05-08 23:45:50.014376"
 container_url: "https://github.com/orgs/autamus/packages/container/package/git"
 aliases:
  - "git"
