+++ _library/golang.md
@@ -3,7 +3,7 @@ layout: container
 name:  "golang"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/golang/container.yaml"
-updated_at: "2021-05-08 23:46:03.670348"
+updated_at: "2021-05-09 15:35:04.551691"
 container_url: "https://hub.docker.com/r/_/golang"
 aliases:
  - "go"
