+++ _library/biocontainers-talon.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/talon"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/talon/container.yaml"
-updated_at: "2021-05-08 23:46:10.518362"
+updated_at: "2021-05-09 15:35:13.140405"
 container_url: "https://hub.docker.com/r/biocontainers/talon"
 aliases:
  - "talon"
