+++ _library/ghcr.io-autamus-igraph.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/igraph"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/igraph/container.yaml"
-updated_at: "2021-05-09 15:34:47.175009"
+updated_at: "2021-05-10 15:19:13.894806"
 container_url: "https://github.com/orgs/autamus/packages/container/package/igraph"
 
 versions:
