+++ registry/ghcr.io/autamus/beast2/container.yaml
@@ -4,10 +4,11 @@ maintainer: '@vsoch'
 description: BEAST 2 is a cross-platform program for Bayesian phylogenetic analysis
   of molecular sequences.
 latest:
-  2.6.3: sha256:5eae1eabe7b127b3847401fe69e3430f97936d7ef7198287aabe01c5e1a957cf
+  2.6.4: sha256:84a75511810083d5ff5c21dff67e995dc2049c5dd8f569a616164b75baea9b50
 tags:
   2.6.3: sha256:5eae1eabe7b127b3847401fe69e3430f97936d7ef7198287aabe01c5e1a957cf
-  latest: sha256:5eae1eabe7b127b3847401fe69e3430f97936d7ef7198287aabe01c5e1a957cf
+  2.6.4: sha256:84a75511810083d5ff5c21dff67e995dc2049c5dd8f569a616164b75baea9b50
+  latest: sha256:84a75511810083d5ff5c21dff67e995dc2049c5dd8f569a616164b75baea9b50
 aliases:
   beast: /opt/view/bin/beast
   beasti: /opt/view/bin/beasti
