### 2021-05-11 d2b30df
Remove latest filter from mysql
