+++ registry/tomcat/container.yaml
@@ -4,7 +4,7 @@ maintainer: '@vsoch'
 description: Apache Tomcat is an open source implementation of the Java Servlet and
   JavaServer Pages technologies
 latest:
-  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:ae64a2cc72052228acb020f8a188b7cd036c186f668af510ff0532435be2a64f
+  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:a7418f29d3dd7ad20bcf052b3b6dc2777d118344286c6374bc447fb217a97c08
 tags:
-  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:ae64a2cc72052228acb020f8a188b7cd036c186f668af510ff0532435be2a64f
+  10.0.5-jdk11-adoptopenjdk-hotspot: sha256:a7418f29d3dd7ad20bcf052b3b6dc2777d118344286c6374bc447fb217a97c08
   latest: sha256:0509684774ac53d8b152f4a50dd92889218a8633c0feddb4b35d092e55cd215d
