+++ _library/node.md
@@ -3,7 +3,7 @@ layout: container
 name:  "node"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/node/container.yaml"
-updated_at: "2021-05-10 15:19:32.375461"
+updated_at: "2021-05-11 13:31:01.278911"
 container_url: "https://hub.docker.com/r/_/node"
 aliases:
  - "node"
