+++ _library/mysql.md
@@ -3,7 +3,7 @@ layout: container
 name:  "mysql"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/mysql/container.yaml"
-updated_at: "2021-05-10 15:18:57.073379"
+updated_at: "2021-05-11 13:30:26.525124"
 container_url: "https://hub.docker.com/r/_/mysql"
 aliases:
  - "mysql"
