+++ registry/ruby/container.yaml
@@ -4,9 +4,11 @@ maintainer: '@vsoch'
 description: Ruby is a dynamic, reflective, object-oriented, general-purpose, open-source
   programming language.
 latest:
-  latest: sha256:091ee4779c0d90155b6d1a317855ce64714e6485f9db4413c812ddd112df7dc7
+  latest: sha256:0a63efa1831b00397f234668c8cc1257bcc426da965679fabd204b6cc51063b9
 tags:
-  latest: sha256:091ee4779c0d90155b6d1a317855ce64714e6485f9db4413c812ddd112df7dc7
+  latest: sha256:0a63efa1831b00397f234668c8cc1257bcc426da965679fabd204b6cc51063b9
+filter:
+- latest
 aliases:
   bundle: /usr/local/bin/bundle
   bundler: /usr/local/bin/bundler
@@ -20,5 +22,3 @@ aliases:
   ri: /usr/local/bin/ri
   ruby: /usr/local/bin/ruby
   typeprof: /usr/local/bin/typeprof
-filter:
-- latest
