+++ registry/nginx/container.yaml
@@ -8,7 +8,7 @@ latest:
   1.20.0-alpine-perl: sha256:00109feeafcc6b80f2322cf70b406d5e521f53d0b0841acfce22aa2a827115d6
 tags:
   1.20.0-alpine-perl: sha256:00109feeafcc6b80f2322cf70b406d5e521f53d0b0841acfce22aa2a827115d6
-  latest: sha256:75a55d33ecc73c2a242450a9f1cc858499d468f077ea942867e662c247b5e412
+  latest: sha256:bf2ccf1e3f83bbe1c99b2332e7573479dd5c8909f27bc03a712e0750ccbe551d
   stable-alpine-perl: sha256:00109feeafcc6b80f2322cf70b406d5e521f53d0b0841acfce22aa2a827115d6
 aliases:
   nginx: /usr/sbin/nginx
