+++ registry/jupyter/datascience-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:5e5bf78bfa351c0255cea5269a0461afbf6b50b51e923a7436229208ea8487f9
 tags:
   "4.0": sha256:5e5bf78bfa351c0255cea5269a0461afbf6b50b51e923a7436229208ea8487f9
-  latest: sha256:fd9e7676db70e2c625ae1c3fb8c4d1583ce8954f2470fa4fb2c8af153569e8f6
+  latest: sha256:e0063028e83a0d94983f4e07d01f89c0ffd34b7c13e8d0e971b78b17725bf0cb
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
