+++ registry/mongo/container.yaml
@@ -9,7 +9,7 @@ latest:
 tags:
   4.4.5-bionic: sha256:cc8bb8711114fa726ca641f5644791d3d43bdd2c95dac541a5fa2c90a2e6e972
   4.4.6-bionic: sha256:3dbb3b4b8f33c8a61fa72cf1d05398d6cc09c72aa2a1f664b639d8231b0e83ab
-  latest: sha256:4e80da85d20b2893d480fd3c9f5b56fe24cd15ed0a80cda248574b2bf5489c8d
+  latest: sha256:401ae88210c708d54873af22df618b46a5f3d2da1c2676d115326b205ec78d3f
 aliases:
   mongo: /usr/bin/mongo
   mongod: /usr/bin/mongod
