+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:644913ccb1cebfe3ddc46f58577fa9e10c6e07c14c7b781aac688e307222ca46
+  latest: sha256:78f7ce52b4d4505363db8463e7113c71b52a5df2fc299758bd0b60bf5462e867
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack

[…]
