+++ registry/openjdk/container.yaml
@@ -4,9 +4,9 @@ maintainer: '@vsoch'
 description: OpenJDK is an open-source implementation of the Java Platform, Standard
   Edition.
 latest:
-  16.0.1-buster: sha256:a0168916eecee7337948e66604681a40c973715455cdca50ab519b8da3c20760
+  16.0.1-buster: sha256:7bcfb795165585c87725e7a673a39e574ef2feb7063a3fa8f4812f8b90cd9700
 tags:
-  16.0.1-buster: sha256:a0168916eecee7337948e66604681a40c973715455cdca50ab519b8da3c20760
+  16.0.1-buster: sha256:7bcfb795165585c87725e7a673a39e574ef2feb7063a3fa8f4812f8b90cd9700
   latest: sha256:fa5414c6c4f8f836308947ae46daf65f2bce39ac01e89edfbe730cd0703dca06
 aliases:
   jar: /usr/bin/jar
