+++ registry/jupyter/pyspark-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:2aa9a0183bcfe80cd9e2a6c0a52d797bd5b67d1e2db9437c7345c2404b5a86c1
 tags:
   "4.0": sha256:2aa9a0183bcfe80cd9e2a6c0a52d797bd5b67d1e2db9437c7345c2404b5a86c1
-  latest: sha256:c145d9c3500876c34c54400126481e584f4cdf09587c2a84c97d2afb6aa640e9
+  latest: sha256:e81ac846ff83cdd1f06a37e02772525e7f68570fcfd63badb7865f72b44a05bd
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
