+++ _library/openjdk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "openjdk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/openjdk/container.yaml"
-updated_at: "2021-05-11 13:30:26.290780"
+updated_at: "2021-05-12 15:40:06.511592"
 container_url: "https://hub.docker.com/_/openjdk"
 aliases:
  - "jar"
