+++ _library/ghcr.io-autamus-snappy.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/snappy"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/snappy/container.yaml"
-updated_at: "2021-05-11 13:30:42.692325"
+updated_at: "2021-05-12 15:40:19.910611"
 container_url: "https://github.com/orgs/autamus/packages/container/package/snappy"
 
 versions:
