+++ _library/mongo.md
@@ -3,7 +3,7 @@ layout: container
 name:  "mongo"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/mongo/container.yaml"
-updated_at: "2021-05-11 13:31:04.588230"
+updated_at: "2021-05-12 15:40:38.262729"
 container_url: "https://hub.docker.com/r/_/mongo"
 aliases:
  - "mongo"
