+++ registry/tomcat/container.yaml
@@ -7,4 +7,4 @@ latest:
   10.0.5-jdk11-adoptopenjdk-hotspot: sha256:a7418f29d3dd7ad20bcf052b3b6dc2777d118344286c6374bc447fb217a97c08
 tags:
   10.0.5-jdk11-adoptopenjdk-hotspot: sha256:a7418f29d3dd7ad20bcf052b3b6dc2777d118344286c6374bc447fb217a97c08
-  latest: sha256:0509684774ac53d8b152f4a50dd92889218a8633c0feddb4b35d092e55cd215d
+  latest: sha256:906baf28bd9a929544c87f649edcb9780e9739678e833be5a7a62a47f7ff2dd3
