+++ registry/php/container.yaml
@@ -10,7 +10,7 @@ tags:
   8.0.3-alpine: sha256:f32cd843417d9271d1500ed3867922175604a243d8033ad1572d617ad7954171
   8.0.5-alpine: sha256:b728c7e60d1bdb28bfba1bbd4ac3328fbde3a70a959a55c318ccae3fc4d170cf
   8.0.6-alpine: sha256:c6335954fd836e28161b662af8b44fb382f41aeae7090359285c57b207117a21
-  latest: sha256:4481e37e100a811e2e33cd184fb056d1c171a7cc64d054558901ddbb4e110513
+  latest: sha256:1942f4046bbc358c3ad2e157db8341f95fb72673c76fd6b2e1c3fe429eb3dce9
 aliases:
   php: /usr/local/bin/php
   php-cgi: /usr/local/bin/php-cgi
