+++ _library/ghcr.io-autamus-r.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/r"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/r/container.yaml"
-updated_at: "2021-05-12 15:40:23.196641"
+updated_at: "2021-05-13 14:46:31.171480"
 container_url: "https://github.com/orgs/autamus/packages/container/package/r"
 aliases:
  - "R"
