+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:ccf3a52d71ecd91cc9eda767327727f0d1630c65407e39c96752880393b9b8e1
+  latest: sha256:6b9538f524418d17fff060984234e4c1073671e2fac9bf8be393fec5af529941
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
