+++ _library/ghcr.io-autamus-bracken.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bracken"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bracken/container.yaml"
-updated_at: "2021-05-13 14:46:34.529971"
+updated_at: "2021-05-14 14:53:14.391422"
 container_url: "https://github.com/orgs/autamus/packages/container/package/bracken"
 
 versions:
