+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:6b9538f524418d17fff060984234e4c1073671e2fac9bf8be393fec5af529941
+  latest: sha256:5a0c847b4d1f893430ea89369cbc5a34fb4ee3d541da5f737417b8e1126854a9
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
