+++ _library/rocker-tidyverse.md
@@ -3,7 +3,7 @@ layout: container
 name:  "rocker/tidyverse"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/rocker/tidyverse/container.yaml"
-updated_at: "2021-05-14 14:52:56.324269"
+updated_at: "2021-05-15 14:09:47.254767"
 container_url: "https://hub.docker.com/r/rocker/tidyverse"
 aliases:
  - "R"
