+++ _library/ghcr.io-autamus-intel-mkl.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/intel-mkl"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/intel-mkl/container.yaml"
-updated_at: "2021-05-14 14:53:13.097883"
+updated_at: "2021-05-15 14:10:04.581549"
 container_url: "https://github.com/orgs/autamus/packages/container/package/intel-mkl"
 
 versions:
