+++ _library/ghcr.io-autamus-r.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/r"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/r/container.yaml"
-updated_at: "2021-05-15 14:10:03.084989"
+updated_at: "2021-05-16 20:39:32.857702"
 container_url: "https://github.com/orgs/autamus/packages/container/package/r"
 aliases:
  - "R"
