+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:f1f04346e2338829594f14116e0b2189bde86a3968c04f02f861eb68c2b69c3d
+  latest: sha256:ee97ddc153804ca3286775c06cee0c924a6d4a6904078c33bbb305463653f7fc
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
