+++ _library/ncbi-blast.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ncbi/blast"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ncbi/blast/container.yaml"
-updated_at: "2021-05-16 20:39:46.535033"
+updated_at: "2021-05-17 14:28:26.932872"
 container_url: "https://hub.docker.com/r/ncbi/blast"
 
 versions:
