+++ _library/ghcr.io-autamus-accumulo.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/accumulo"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/accumulo/container.yaml"
-updated_at: "2021-05-16 20:39:22.690342"
+updated_at: "2021-05-17 14:27:58.278768"
 container_url: "https://github.com/orgs/autamus/packages/container/package/accumulo"
 
 versions:
