+++ registry/jupyter/tensorflow-notebook/container.yaml
@@ -3,10 +3,10 @@ url: https://hub.docker.com/r/jupyter/tensorflow-notebook
 maintainer: '@vsoch'
 description: Jupyter Tensorflow Notebook from https://github.com/jupyter/docker-stacks
 latest:
-  ubuntu-20.04: sha256:9e35cdea97c6f3919216eeebd647dbea8cf4494cb6ae62efb7ac2c59172c1a25
+  ubuntu-20.04: sha256:55f88c231ad74399b37ada32e0232d429632e88dc01c5a354b5b727144fcd0ee
 tags:
-  latest: sha256:9e35cdea97c6f3919216eeebd647dbea8cf4494cb6ae62efb7ac2c59172c1a25
-  ubuntu-20.04: sha256:9e35cdea97c6f3919216eeebd647dbea8cf4494cb6ae62efb7ac2c59172c1a25
+  latest: sha256:55f88c231ad74399b37ada32e0232d429632e88dc01c5a354b5b727144fcd0ee
+  ubuntu-20.04: sha256:55f88c231ad74399b37ada32e0232d429632e88dc01c5a354b5b727144fcd0ee
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
