+++ _library/consul.md
@@ -3,7 +3,7 @@ layout: container
 name:  "consul"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/consul/container.yaml"
-updated_at: "2021-05-18 14:04:40.833753"
+updated_at: "2021-05-19 12:17:34.137444"
 container_url: "https://hub.docker.com/_/consul"
 aliases:
  - "consul"
