+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:0d1ed6899ff480cf9be5b5a388dfb5363f818fb8595e3cf21d9c6a50a80cc295
+  latest: sha256:c175c7dd27e76c0f14da6fadd3df336dddf32537e30be49e05c357678670f377
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
