+++ _library/ghcr.io-autamus-snappy.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/snappy"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/snappy/container.yaml"
-updated_at: "2021-05-19 12:17:15.151937"
+updated_at: "2021-05-20 15:24:30.971839"
 container_url: "https://github.com/orgs/autamus/packages/container/package/snappy"
 
 versions:
