+++ _library/ghcr.io-autamus-lammps.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/lammps"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/lammps/container.yaml"
-updated_at: "2021-05-19 12:17:19.756114"
+updated_at: "2021-05-20 15:24:36.658368"
 container_url: "https://github.com/orgs/autamus/packages/container/package/lammps"
 
 versions:
