+++ _library/ghcr.io-autamus-sparsehash.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/sparsehash"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/sparsehash/container.yaml"
-updated_at: "2021-05-19 12:17:04.162807"
+updated_at: "2021-05-20 15:24:17.079061"
 container_url: "https://github.com/orgs/autamus/packages/container/package/sparsehash"
 
 versions:
