### 2021-05-20 253d917
update paper to use semicolons to separate references
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
