+++ _library/ghcr.io-autamus-mpc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/mpc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/mpc/container.yaml"
-updated_at: "2021-05-20 15:24:14.150569"
+updated_at: "2021-05-20 18:41:29.952481"
 container_url: "https://github.com/orgs/autamus/packages/container/package/mpc"
 
 versions:
