+++ registry/nvcr.io/nvidia/tensorflow/container.yaml
@@ -6,10 +6,11 @@ description: TensorFlow is an open-source software library for high-performance 
   a variety of platforms (CPUs, GPUs, TPUs), and from desktops to clusters of servers
   to mobile and edge devices.
 latest:
-  21.04-tf1-py3: sha256:fb948e24013e079fbba954dc0ce025d3648d8b12d343cd07d24e7ca14b66ba5b
+  21.05-tf1-py3: sha256:a6a89e9ac5db5e67434785268b52df2a06df2fb437eed2abaf54df06db5e4f78
 tags:
   21.02-tf1-py3: sha256:74f3e5f4a60a312df811ee5104baa189d158697953a91e7768f777895bb1d438
   21.03-tf1-py3: sha256:addbbdccb193849d58083154458318b2aa9b7945d9975284291b08b4336415e1
   21.04-tf1-py3: sha256:fb948e24013e079fbba954dc0ce025d3648d8b12d343cd07d24e7ca14b66ba5b
+  21.05-tf1-py3: sha256:a6a89e9ac5db5e67434785268b52df2a06df2fb437eed2abaf54df06db5e4f78
 filter:
 - 21*
