+++ _library/ghcr.io-autamus-gromacs.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/gromacs"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/gromacs/container.yaml"
-updated_at: "2021-05-20 18:41:30.813581"
+updated_at: "2021-05-21 09:55:13.981817"
 container_url: "https://github.com/orgs/autamus/packages/container/package/gromacs"
 
 versions:
