+++ _library/bids-freesurfer.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/freesurfer"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/freesurfer/container.yaml"
-updated_at: "2021-05-20 18:41:59.849480"
+updated_at: "2021-05-21 09:55:40.321878"
 container_url: "https://hub.docker.com/r/bids/freesurfer"
 
 versions:
