+++ _library/ghcr.io-autamus-lammps.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/lammps"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/lammps/container.yaml"
-updated_at: "2021-05-20 18:41:51.736926"
+updated_at: "2021-05-21 09:55:32.988092"
 container_url: "https://github.com/orgs/autamus/packages/container/package/lammps"
 
 versions:
