### 2021-05-21 #3217 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - journal.pone.0177459 is OK
> - 10.1145/2723872.2723882 is OK
>
> MISSING DOIs
>
> - 10.1145/2063348.2063360 may be a valid DOI for title: Best practices for the deployment and management of production HPC clusters
> - 10.1007/978-1-4842-3012-1_7 may be a valid DOI for title: The Opinionated Jupyter Stacks
> - 10.1007/978-1-4842-3012-1_6 may be a valid DOI for title: Docker Hub
> - 10.1371/journal.pcbi.1005209 may be a valid DOI for title: BIDS apps: Improving ease of use, accessibility, and reproducibility of neuroimaging data analysis methods
>
> INVALID DOIs
>
> - http://dx.doi.org/10.1155/2015/243180 is INVALID because of 'https://doi.org/' prefix
> ```

 @vsoch please check the missing and invalid DOIs and update the paper accordingly.
