### 2021-05-21 #3217 @diehlpk
> Definitely! Where is the Google sheet with the list?

I pinged in the message above all people from the Google sheet having Docker or Singularity mentioned. If you know other folks that would be great.
