+++ registry/jupyter/tensorflow-notebook/container.yaml
@@ -3,10 +3,11 @@ url: https://hub.docker.com/r/jupyter/tensorflow-notebook
 maintainer: '@vsoch'
 description: Jupyter Tensorflow Notebook from https://github.com/jupyter/docker-stacks
 latest:
-  ubuntu-20.04: sha256:99936703ccabdb10f1525fbd37aaf40688c504cd365d21ad04e661bd02d96618
+  ed559b677553: sha256:6c2958b6b738bcbec4d7aaed3af2aabf7dc7278b3bd559213863827279874fb6
 tags:
-  latest: sha256:99936703ccabdb10f1525fbd37aaf40688c504cd365d21ad04e661bd02d96618
-  ubuntu-20.04: sha256:99936703ccabdb10f1525fbd37aaf40688c504cd365d21ad04e661bd02d96618
+  ed559b677553: sha256:6c2958b6b738bcbec4d7aaed3af2aabf7dc7278b3bd559213863827279874fb6
+  latest: sha256:6c2958b6b738bcbec4d7aaed3af2aabf7dc7278b3bd559213863827279874fb6
+  ubuntu-20.04: sha256:6c2958b6b738bcbec4d7aaed3af2aabf7dc7278b3bd559213863827279874fb6
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
