+++ registry/ghcr.io/autamus/ascent/container.yaml
@@ -4,7 +4,8 @@ maintainer: '@vsoch'
 description: An open source many-core capable lightweight in situ visualization and
   analysis infrastructure for multi-physics HPC simulations.
 latest:
-  0.7.0: sha256:c025f72e442b7a4f5626e0520e09e8a7a5e5f5c630565cbd8d0d178ff2585ee6
+  0.7.1: sha256:83c71fb3ac8716336a4c08d200723c6bd7fc1c3cad304f5c391e18cadabc6139
 tags:
   0.7.0: sha256:c025f72e442b7a4f5626e0520e09e8a7a5e5f5c630565cbd8d0d178ff2585ee6
-  latest: sha256:c025f72e442b7a4f5626e0520e09e8a7a5e5f5c630565cbd8d0d178ff2585ee6
+  0.7.1: sha256:83c71fb3ac8716336a4c08d200723c6bd7fc1c3cad304f5c391e18cadabc6139
+  latest: sha256:83c71fb3ac8716336a4c08d200723c6bd7fc1c3cad304f5c391e18cadabc6139
