+++ registry/spack/ubuntu-bionic/container.yaml
@@ -6,7 +6,7 @@ latest:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
-  latest: sha256:c1fedcafdb938084a31d3b2eb8cc2bbb7e452a3034966372a2ffa07943a98127
+  latest: sha256:2cacaf5f6ee2b141912d6df8e6b421ac20ec7a79e6dae77629ac464f80680f89
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
