+++ registry/jupyter/minimal-notebook/container.yaml
@@ -6,7 +6,7 @@ latest:
   "4.0": sha256:6f9d189494fe091700681eb39b429dfe1420ad0717697c2e2293779fcb80cd1e
 tags:
   "4.0": sha256:6f9d189494fe091700681eb39b429dfe1420ad0717697c2e2293779fcb80cd1e
-  latest: sha256:fa135ed685cff884d27c278dcc9fde0afb1b3e05ef50482d94cfa371ed2b0e3c
+  latest: sha256:de3b988c3feba73863865e26a71b8ca8439b5ef40c91e49b76a60caa4e0e00fd
 aliases:
 - name: run-notebook
   command: jupyter notebook --no-browser --port=$(shuf -i 2000-65000 -n 1) --ip 0.0.0.0
