+++ _library/rocker-ml-verse.md
@@ -3,7 +3,7 @@ layout: container
 name:  "rocker/ml-verse"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/rocker/ml-verse/container.yaml"
-updated_at: "2021-05-22 15:46:56.235059"
+updated_at: "2021-05-23 12:57:09.859739"
 container_url: "https://hub.docker.com/r/rocker/ml-verse"
 aliases:
  - "R"
