+++ registry/ghcr.io/autamus/gcc/container.yaml
@@ -3,10 +3,10 @@ url: https://github.com/orgs/autamus/packages/container/package/gcc
 maintainer: '@vsoch'
 description: The GNU Compiler Collection
 latest:
-  10.3.0: sha256:92e25a9d8ae9c9af84376dbd42bd4c65df468aec22266e958de710943d5abba8
+  10.3.0: sha256:099c085f4ab471fd09256baba7bc1cdf73b58df40127a1d91782e005dd41c5c0
 tags:
-  10.3.0: sha256:92e25a9d8ae9c9af84376dbd42bd4c65df468aec22266e958de710943d5abba8
-  latest: sha256:92e25a9d8ae9c9af84376dbd42bd4c65df468aec22266e958de710943d5abba8
+  10.3.0: sha256:099c085f4ab471fd09256baba7bc1cdf73b58df40127a1d91782e005dd41c5c0
+  latest: sha256:099c085f4ab471fd09256baba7bc1cdf73b58df40127a1d91782e005dd41c5c0
 aliases:
   c++: /opt/view/bin/c++
   cpp: /opt/view/bin/cpp
