+++ registry/spack/ubuntu-bionic/container.yaml
@@ -7,7 +7,7 @@ latest:
 tags:
   0.16.1: sha256:8261977ff63fe420446c349f0e3bd4e09a6417ebb1008ab472861041f1edd11b
   0.16.2: sha256:698899684998df4a49f02bce1cffca9aa59644477f94b1909fc26b2adf4c4be4
-  latest: sha256:c8a00d6007bb75ea99850d1926f37328e33457b4dfc3f8239c53d19376c49bfd
+  latest: sha256:ebd1be0cfdcdadfc4a349c35338eac046305a4608c87da04a735c990754e29ce
 aliases:
   sbang: /opt/spack/bin/sbang
   spack: /opt/spack/bin/spack
