### 2021-05-24 #3311 @whedon
[REVIEW]: Collaborative Container Modules with Singularity Registry HPC
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://github.com/singularityhub/singularity-hpc" target ="_blank">https://github.com/singularityhub/singularity-hpc</a>
**Version:** v0.0.31

@rcannood & @zbeekman, @ArangoGutierrez please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
