### 2021-05-24 #3311 @diehlpk
> Failed to discover a `Statement of need` section in paper

@vsoch Could you please add the `Statement of need` because each paper requires one
`
A Statement of Need section that clearly illustrates the research purpose of the software.`
