+++ _library/php.md
@@ -3,7 +3,7 @@ layout: container
 name:  "php"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/php/container.yaml"
-updated_at: "2021-05-23 12:56:38.709014"
+updated_at: "2021-05-24 15:02:48.018352"
 container_url: "https://hub.docker.com/_/php"
 aliases:
  - "php"
