+++ _library/ghcr.io-autamus-argobots.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/argobots"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/argobots/container.yaml"
-updated_at: "2021-05-23 12:56:57.453038"
+updated_at: "2021-05-24 15:03:06.100115"
 container_url: "https://github.com/orgs/autamus/packages/container/package/argobots"
 
 versions:
