+++ _library/ghcr.io-autamus-snappy.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/snappy"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/snappy/container.yaml"
-updated_at: "2021-05-23 12:57:00.172868"
+updated_at: "2021-05-24 15:03:08.698016"
 container_url: "https://github.com/orgs/autamus/packages/container/package/snappy"
 
 versions:
