### 2021-05-24 85d60f0
Update update-binoc.yml
Run binoc once a week
