### 2021-05-24 #3311 @diehlpk
> @diehlpk all set! It was an oversight to not rename the Background section to Statement of Need.

Thanks, the paper is now good for review.
