### 2021-05-27 d78b443
fixing bug with shpc test and copying containers instead of moving them
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
