### 2021-05-27 4cc9f37
adding fix for shpc inspect
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
