+++ shpc/settings.yml
@@ -14,7 +14,7 @@ registry: $root_dir/registry
 module_base: $root_dir/modules
 
 # store containers separately from module files
-container_base: null
+container_base:
 
 # if defined, add to lmod script to load this Singularity module first
 singularity_module:
@@ -23,7 +23,11 @@ singularity_module:
 module_exc_prefix: ''
 
 # string with comma separated list of paths to binds. If set, expored to SINGULARITY_BINDPATH
-bindpaths: null
+bindpaths:
 
 # exported to SINGULARITY_SHELL, defaults to /bin/bash.
 singularity_shell: /bin/bash
+
+# the module namespace you want to install from. E.g., if you have ghcr.io/autamus/clingo
+# and you set the namespace to ghcr.io/autamus, you can just do: shpc install clingo.
+namespace: null
