### 2021-05-28 19eaa58
add tests for show and namespace
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
