### 2021-05-29 ed2d000
Adding namespaces to make module install easier (#366)
* adding namespaces to make module install easier
* add tests for show and namespace

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
