+++ registry/vanessa/salad/container.yaml
@@ -1,0 +1,10 @@
+docker: vanessa/salad
+url: https://hub.docker.com/r/vanessa/salad
+maintainer: '@vsoch'
+description: A container all about fork and spoon puns.
+latest:
+  latest: sha256:e8302da47e3200915c1d3a9406d9446f04da7244e4995b7135afd2b79d4f63db
+tags:
+  latest: sha256:e8302da47e3200915c1d3a9406d9446f04da7244e4995b7135afd2b79d4f63db
+aliases:
+  salad: /code/salad
