### 2021-05-29 d70803e
First round of fixes (#364)
* fixing bug with shpc test and copying containers instead of moving them
* adding fix for shpc inspect

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
