+++ _library/ghcr.io-autamus-zlib.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/zlib"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/zlib/container.yaml"
-updated_at: "2021-05-24 15:22:03.970655"
+updated_at: "2021-05-29 01:53:07.608035"
 container_url: "https://github.com/orgs/autamus/packages/container/package/zlib"
 
 versions:
