+++ _library/bids-spm.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/spm"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/spm/container.yaml"
-updated_at: "2021-05-29 01:53:20.571386"
+updated_at: "2021-05-29 01:53:44.548450"
 container_url: "https://hub.docker.com/r/bids/spm"
 aliases:
  - "spm12"
