+++ _library/golang.md
@@ -3,7 +3,7 @@ layout: container
 name:  "golang"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/golang/container.yaml"
-updated_at: "2021-05-29 01:53:24.176292"
+updated_at: "2021-05-29 01:53:47.342078"
 container_url: "https://hub.docker.com/r/_/golang"
 aliases:
  - "go"
