### 2021-05-29 ab117b7
improve help for add to show sif_path and module_id to close #369
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
