### 2021-05-29 123c821
adding support for container recipe features
The container.yaml recipe files should be agnostic to the container technology,
meaning that instead of an option like --nv specific to singularity, we have gpu: true,
and then the container technology that is loaded can add the flags appropriate for that
feature.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
