### 2021-05-29 bd5783c
default for list will show all unique container:tag
unless the user adds --short, in which case the tags will again be
combined into a single list alongside the container name.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
