+++ registry/vanessa/salad/container.yaml
@@ -8,7 +8,5 @@ tags:
   latest: sha256:e8302da47e3200915c1d3a9406d9446f04da7244e4995b7135afd2b79d4f63db
 aliases:
   salad: /code/salad
-features:
-  gpu: true
 env:
   maintainer: vsoch
