### 2021-05-29 aa7ca98
allowing for the user to provide a container:tag instead of container/tag
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
