### 2021-05-29 f6bfc9e
fixing bug with inspect
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
