+++ shpc/main/schemas.py
@@ -112,7 +112,7 @@ settingsProperties = {
     "bindpaths": {"type": ["string", "null"]},
     "updated_at": {"type": "string"},
     "environment_file": {"type": "string"},
-    "container_tech": {"type": "string", "enum": ["singularity"]},
+    "container_tech": {"type": "string", "enum": ["singularity", "podman"]},
     "singularity_shell": {"type": "string", "enum": ["/bin/bash", "/bin/sh"]},
     "module_sys": {"type": "string", "enum": ["lmod", "tcl", None]},
     "container_features": container_features,
