+++ registry/ghcr.io/autamus/wget/container.yaml
@@ -4,9 +4,9 @@ maintainer: '@vsoch'
 description: GNU Wget is a free software package for retrieving files using HTTP,
   HTTPS, FTP and FTPS, the most widely used Internet protocols.
 latest:
-  1.21.1: sha256:118febd78a52f4a7b7d85271008dbc8ff944e48d4d3e1025787af2e09cf3c734
+  1.21.1: sha256:486b3df7b1f8971243b0257aa1f2ab223c2378d7f4b4d8314c1962c0bb018ec9
 tags:
-  1.21.1: sha256:118febd78a52f4a7b7d85271008dbc8ff944e48d4d3e1025787af2e09cf3c734
-  latest: sha256:118febd78a52f4a7b7d85271008dbc8ff944e48d4d3e1025787af2e09cf3c734
+  1.21.1: sha256:486b3df7b1f8971243b0257aa1f2ab223c2378d7f4b4d8314c1962c0bb018ec9
+  latest: sha256:486b3df7b1f8971243b0257aa1f2ab223c2378d7f4b4d8314c1962c0bb018ec9
 aliases:
   wget: /opt/view/bin/wget
