+++ registry/ghcr.io/autamus/lmod/container.yaml
@@ -5,9 +5,9 @@ description: The Persistence of Vision Ray Tracer, most commonly acronymed as PO
   is a cross-platform ray-tracing program that generates images from a text-based
   scene description.
 latest:
-  "8.3": sha256:7f4e27462257d3059f79da144d6797655dac8d41769985bf527e78e632e79794
+  "8.3": sha256:617368ca58439b0b8e1c94f9e8dc01189a3abfbcdbf75593686ea6eb6d75a122
 tags:
-  "8.3": sha256:7f4e27462257d3059f79da144d6797655dac8d41769985bf527e78e632e79794
-  latest: sha256:7f4e27462257d3059f79da144d6797655dac8d41769985bf527e78e632e79794
+  "8.3": sha256:617368ca58439b0b8e1c94f9e8dc01189a3abfbcdbf75593686ea6eb6d75a122
+  latest: sha256:617368ca58439b0b8e1c94f9e8dc01189a3abfbcdbf75593686ea6eb6d75a122
 aliases:
   module: . /opt/view/lmod/8.3/init/profile && module
