+++ registry/ghcr.io/autamus/julia/container.yaml
@@ -3,10 +3,10 @@ url: https://github.com/orgs/autamus/packages/container/package/julia
 maintainer: '@vsoch'
 description: Julia is a high-level, high-performance, dynamic programming language.
 latest:
-  1.6.1: sha256:61513de14ad2e3534d07ba5b79b0b1c17cbc3c046025fe136d7a1ebb0712c03c
+  1.6.1: sha256:03cf73f08ed622a2e51adb5196f42962ae8e971fa5b5c598c60db79e157c464e
 tags:
   1.6.0: sha256:e898f6e5faced775bca53c9e8cd28f56f1b6d42a719d06678a7012ab9f21f32e
-  1.6.1: sha256:61513de14ad2e3534d07ba5b79b0b1c17cbc3c046025fe136d7a1ebb0712c03c
-  latest: sha256:61513de14ad2e3534d07ba5b79b0b1c17cbc3c046025fe136d7a1ebb0712c03c
+  1.6.1: sha256:03cf73f08ed622a2e51adb5196f42962ae8e971fa5b5c598c60db79e157c464e
+  latest: sha256:03cf73f08ed622a2e51adb5196f42962ae8e971fa5b5c598c60db79e157c464e
 aliases:
   julia: /opt/view/bin/julia
