### 2021-05-30 1849784
refactoring organization to better support different container arhitectures
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
