+++ _library/ghcr.io-autamus-bowtie2.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bowtie2"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bowtie2/container.yaml"
-updated_at: "2021-05-29 01:53:22.584695"
+updated_at: "2021-06-05 18:56:42.134015"
 container_url: "https://github.com/orgs/autamus/packages/container/package/bowtie2"
 aliases:
  - "bowtie2"
@@ -63,33 +63,45 @@ You can use tab for auto-completion of module names or commands that are provide
 
 ### Commands
 
-When you install this module, you'll be able to load it to make the following commands accessible:
+When you install this module, you'll be able to load it to make the following commands accessible.
+Examples for both Singularity and Podman (container technologies supported) are included.
 
-#### ghcr.io-autamus-bowtie2-run:
+#### -run:
 
 ```bash
 $ singularity run <container>
+$ podman run --rm  -v ${PWD} -w ${PWD} <container>
 ```
 
-#### ghcr.io-autamus-bowtie2-shell:
+#### -shell:
 
 ```bash
-$ singularity shell -s /bin/bash <container>
+$ singularity shell -s /bin/sh <container>
+$ podman run --it --rm --entrypoint /bin/sh  -v ${PWD} -w ${PWD} <container>
 ```
 
-#### ghcr.io-autamus-bowtie2-exec:
+#### -exec:
 
 ```bash
-$ singularity exec -s /bin/bash <container> "$@"
+$ singularity exec -s /bin/sh <container> "$@"
+$ podman run --it --rm --entrypoint ""  -v ${PWD} -w ${PWD} <container> "$@"
 ```
 
-#### ghcr.io-autamus-bowtie2-inspect-runscript:
+#### -inspect:
+
+Podman only has one inspect type.
+
+```bash
+$ podman inspect <container>
+```
+
+#### -inspect-runscript:
 
 ```bash
 $ singularity inspect -r <container>
 ```
 
-#### ghcr.io-autamus-bowtie2-inspect-deffile:
+#### -inspect-deffile:
 
 ```bash
 $ singularity inspect -d <container>
@@ -100,6 +112,7 @@ $ singularity inspect -d <container>
        
 ```bash
 $ singularity exec <container> /opt/view/bin/bowtie2
+$ podman run --it --rm --entrypoint /opt/view/bin/bowtie2   -v ${PWD} -w ${PWD} <container> -c " $@"
 ```
 
 
@@ -107,6 +120,7 @@ $ singularity exec <container> /opt/view/bin/bowtie2
        
 ```bash
 $ singularity exec <container> /opt/view/bin/bowtie2-align-l
+$ podman run --it --rm --entrypoint /opt/view/bin/bowtie2-align-l   -v ${PWD} -w ${PWD} <container> -c " $@"
 ```
 
 
@@ -114,6 +128,7 @@ $ singularity exec <container> /opt/view/bin/bowtie2-align-l
        
 ```bash
 $ singularity exec <container> /opt/view/bin/bowtie2-align-s
+$ podman run --it --rm --entrypoint /opt/view/bin/bowtie2-align-s   -v ${PWD} -w ${PWD} <container> -c " $@"
 ```
 
 
@@ -121,6 +136,7 @@ $ singularity exec <container> /opt/view/bin/bowtie2-align-s
        
 ```bash
 $ singularity exec <container> /opt/view/bin/bowtie2-build
+$ podman run --it --rm --entrypoint /opt/view/bin/bowtie2-build   -v ${PWD} -w ${PWD} <container> -c " $@"
 ```
 
 
[…]
