### 2021-06-05 2c0901a
version bump to 0.0.25
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
