+++ _library/rocker-ml-verse.md
@@ -3,7 +3,7 @@ layout: container
 name:  "rocker/ml-verse"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/rocker/ml-verse/container.yaml"
-updated_at: "2021-06-05 18:57:01.758375"
+updated_at: "2021-06-05 19:02:18.365341"
 container_url: "https://hub.docker.com/r/rocker/ml-verse"
 aliases:
  - "R"
