+++ _library/ghcr.io-autamus-dyninst.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/dyninst"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/dyninst/container.yaml"
-updated_at: "2021-06-05 18:56:38.424991"
+updated_at: "2021-06-05 19:01:53.003642"
 container_url: "https://github.com/orgs/autamus/packages/container/package/dyninst"
 
 versions:
