+++ _library/biocontainers-picard-tools.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/picard-tools"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/picard-tools/container.yaml"
-updated_at: "2021-06-05 18:57:13.167983"
+updated_at: "2021-06-05 19:02:30.675093"
 container_url: "https://hub.docker.com/r/biocontainers/picard-tools/tags"
 
 versions:
