### 2021-06-05 80aa756
add note about --container-tech to getting started docs
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
