+++ _library/ghcr.io-autamus-boost.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/boost"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/boost/container.yaml"
-updated_at: "2021-06-05 19:06:50.043832"
+updated_at: "2021-06-05 19:27:39.191655"
 container_url: "https://github.com/orgs/autamus/packages/container/package/boost"
 
 versions:
