+++ _library/ghcr.io-autamus-fraggenescan.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/fraggenescan"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/fraggenescan/container.yaml"
-updated_at: "2021-06-05 19:27:27.791463"
+updated_at: "2021-06-05 19:31:25.900896"
 container_url: "https://github.com/orgs/autamus/packages/container/package/fraggenescan"
 
 versions:
