+++ _library/vanessa-salad.md
@@ -3,7 +3,7 @@ layout: container
 name:  "vanessa/salad"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/vanessa/salad/container.yaml"
-updated_at: "2021-06-05 19:28:01.761384"
+updated_at: "2021-06-05 19:32:05.028888"
 container_url: "https://hub.docker.com/r/vanessa/salad"
 aliases:
  - "salad"
