+++ _library/consul.md
@@ -3,7 +3,7 @@ layout: container
 name:  "consul"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/consul/container.yaml"
-updated_at: "2021-06-05 19:28:03.509894"
+updated_at: "2021-06-05 19:32:06.999978"
 container_url: "https://hub.docker.com/_/consul"
 aliases:
  - "consul"
