+++ _library/rocker-rstudio.md
@@ -3,7 +3,7 @@ layout: container
 name:  "rocker/rstudio"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/rocker/rstudio/container.yaml"
-updated_at: "2021-06-05 19:27:48.356808"
+updated_at: "2021-06-05 19:31:49.541709"
 container_url: "https://hub.docker.com/r/rocker/rstudio"
 aliases:
  - "R"
