### 2021-06-05 683bf7b
adding conainer tech section in docs
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
