### 2021-06-05 975d4c9
add developer guide and user guide on main sidebar
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
