+++ _library/bids-aa.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/aa"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/aa/container.yaml"
-updated_at: "2021-06-05 19:34:10.114490"
+updated_at: "2021-06-05 21:27:07.095062"
 container_url: "https://hub.docker.com/r/bids/aa"
 
 versions:
