+++ _library/redis.md
@@ -3,7 +3,7 @@ layout: container
 name:  "redis"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/redis/container.yaml"
-updated_at: "2021-06-05 19:34:23.843009"
+updated_at: "2021-06-05 21:27:19.547692"
 container_url: "https://hub.docker.com/r/_/redis"
 aliases:
  - "redis-benchmark"
