+++ _library/ghcr.io-autamus-dyninst.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/dyninst"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/dyninst/container.yaml"
-updated_at: "2021-06-05 19:33:40.569603"
+updated_at: "2021-06-05 21:26:40.164440"
 container_url: "https://github.com/orgs/autamus/packages/container/package/dyninst"
 
 versions:
