### 2021-06-05 08ac2d6
adding support for Docker as a new container technology
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
