+++ registry/ghcr.io/autamus/fraggenescan/container.yaml
@@ -3,7 +3,7 @@ url: https://github.com/orgs/autamus/packages/container/package/fraggenescan
 maintainer: '@vsoch'
 description: An application for finding (fragmented) genes in short reads.
 latest:
-  "1.31": sha256:5fff7294a5b59369c06b0e29ab23ee55edbba7f15131ad33ea1eb737ee0e9691
+  "1.31": sha256:3a546cbb97d03693bceec485691d794c26378236ff653c9f12672a837a1a16fc
 tags:
-  "1.31": sha256:5fff7294a5b59369c06b0e29ab23ee55edbba7f15131ad33ea1eb737ee0e9691
-  latest: sha256:5fff7294a5b59369c06b0e29ab23ee55edbba7f15131ad33ea1eb737ee0e9691
+  "1.31": sha256:3a546cbb97d03693bceec485691d794c26378236ff653c9f12672a837a1a16fc
+  latest: sha256:3a546cbb97d03693bceec485691d794c26378236ff653c9f12672a837a1a16fc
