+++ registry/ghcr.io/autamus/astral/container.yaml
@@ -4,9 +4,9 @@ maintainer: '@vsoch'
 description: ASTRAL is a tool for estimating an unrooted species tree given a set
   of unrooted gene trees.
 latest:
-  5.7.1: sha256:071aaa91502149e02f31b7adf0a563da0a7adbbb7ea3ac4d780bb65a18847108
+  5.7.1: sha256:e03ad784171ebc1bdd52f1e0664a938adf08eb00df9962a9c8e766e69e008e8f
 tags:
-  5.7.1: sha256:071aaa91502149e02f31b7adf0a563da0a7adbbb7ea3ac4d780bb65a18847108
-  latest: sha256:071aaa91502149e02f31b7adf0a563da0a7adbbb7ea3ac4d780bb65a18847108
+  5.7.1: sha256:e03ad784171ebc1bdd52f1e0664a938adf08eb00df9962a9c8e766e69e008e8f
+  latest: sha256:e03ad784171ebc1bdd52f1e0664a938adf08eb00df9962a9c8e766e69e008e8f
 aliases:
   astral: /opt/view/bin/astral
