### 2021-06-06 5fdb424
Cite 'Lmod' with the accurate character case
The Lmod software and project is defined as "Lmod", not "LMOD".
