+++ docs/getting_started/index.rst
@@ -5,7 +5,7 @@ Getting Started
 ===============
 
 Singularity Registry (HPC) is a tool that makes it easy to install containers as
-LMOD modules. You can create your own registry entries (e.g., a specification
+Lmod modules. You can create your own registry entries (e.g., a specification
 to pull a particular container and expose some number of entrypoints) or
 the library also provides you with a community set.
 
