### 2021-06-06 5fea69d
ci: merge lmod/tcl module test script
Merge the *Run python module tests* script for Lmod and Environment
Modules as a single script. It will help to see the initialization
differences between the two module flavors. It will also be easier to
apply changes to this test script or to add support for other module
implementation.

Explicit the use of the bash shell for this script, thus use the
bash-specific initialisation script for module.
