### 2021-06-06 0101153
ci: add step to run python module tests under tcsh
Add test step to check the generated modules under a shell of the Csh
family.
