+++ .github/workflows/test.yml
@@ -15,7 +15,7 @@ jobs:
 
       - name: Install Dependencies
         run: |
-          sudo apt-get install lua5.2 liblua5.2-dev lua-filesystem lua-posix tcl tcl-dev
+          sudo apt-get install lua5.2 liblua5.2-dev lua-filesystem lua-posix tcl tcl-dev tcsh
 
       - name: Install Environment Modules
         if: ${{ matrix.module == 'tcl' }}
@@ -54,7 +54,7 @@ jobs:
           source activate shpc
           pip install -e .
 
-      - name: Run python module tests
+      - name: Run python module tests (bash)
         shell: bash
         run: |
           case "${{ matrix.module }}" in
@@ -70,6 +70,40 @@ jobs:
 
           export PATH="/usr/share/miniconda/bin:$PATH"
           source activate shpc
+          shpc config set container_tech:${{ matrix.container_tech }}
+          shpc config set module_sys:${{ matrix.module }}
+          shpc config set enable_tty:false
+          shpc install python:3.9.5-alpine
+
+          module use ./modules
+          module load python/3.9.5-alpine
+          alias
+
+          printf "\n\nmodule show ============================================\n"
+          module show python/3.9.5-alpine
+
+          printf "\n\nmodule whatis ==========================================\n"
+          module whatis python/3.9.5-alpine
+
+          printf "\n\nmodule help ============================================\n"
+          module help python/3.9.5-alpine
+          python-exec echo donuts
+          python-run python --version
+
+      - name: Run python module tests (tcsh)
+        shell: tcsh -e {0}
+        run: |
+          switch ("${{ matrix.module }}")
+            case lmod:
+              source /usr/share/lmod/lmod/init/tcsh
+              breaksw
+            case tcl:
+              source /usr/local/Modules/init/tcsh
+              breaksw
+          endsw
+
+          source /usr/share/miniconda/etc/profile.d/conda.csh
+          conda activate shpc
           shpc config set container_tech:${{ matrix.container_tech }}
           shpc config set module_sys:${{ matrix.module }}
           shpc config set enable_tty:false
