### 2021-06-06 6af9361
ci: check expected output is obtained on module tests
Verify that the expected output is obtained when testing run/exec. Helps
to determine if arguments are correctly passed over the shell alias or
function defined when loading module.
