+++ shpc/main/container/podman.py
@@ -3,10 +3,7 @@ __copyright__ = "Copyright 2021, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
 
-from shpc.logger import logger
 from .docker import DockerContainer
-import shpc.main.templates
-import shpc.utils
 
 import os
 
@@ -19,11 +16,6 @@ class PodmanContainer(DockerContainer):
     # The module technology adds extensions here
     templatefile = "docker"
     command = "podman"
-
-    def __init__(self):
-        if shpc.utils.which("podman")["return_code"] != 0:
-            logger.exit("podman is required to use the podman base.")
-        super(PodmanContainer, self).__init__()
 
     def shell(self, image):
         """
