+++ _library/bids-niak.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/niak"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/niak/container.yaml"
-updated_at: "2021-06-06 14:33:33.500099"
+updated_at: "2021-06-06 20:43:48.397810"
 container_url: "https://hub.docker.com/r/bids/niak"
 
 versions:
