+++ _library/openjdk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "openjdk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/openjdk/container.yaml"
-updated_at: "2021-06-06 14:33:29.477612"
+updated_at: "2021-06-06 20:43:44.968494"
 container_url: "https://hub.docker.com/_/openjdk"
 aliases:
  - "jar"
