+++ _library/bids-broccoli.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/broccoli"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/broccoli/container.yaml"
-updated_at: "2021-06-06 14:33:32.161590"
+updated_at: "2021-06-06 20:43:47.273760"
 container_url: "https://hub.docker.com/r/bids/broccoli"
 
 versions:
