### 2021-06-07 #3311 @rcannood
Almost finished! I just need to verify the functional claims of the paper.

I can already say that, overall, I think the manuscript is well-written and logically structured. I don't think many suggestions or comments will be needed before publication.

I'm hoping to finish my review this week; I'll keep you posted.
