### 2021-06-07 f34149b
[bot] binoc/update-2021-06-06 (#385)
* Automated deployment to update containers 2021-06-06
* Limit container to just ubuntu to avoid commit tags

Co-authored-by: github-actions <github-actions@users.noreply.github.com>
