+++ registry/ghcr.io/autamus/valgrind/container.yaml
@@ -3,10 +3,10 @@ url: https://github.com/orgs/autamus/packages/container/package/valgrind
 maintainer: '@vsoch'
 description: 'A suite of tools for debugging and profiling. '
 latest:
-  3.17.0: sha256:9c2298a3a3fae58b1c6a5d4b4fc6e323d23a8855a687535060a2c344511faf1b
+  3.17.0: sha256:0301861f6d2aea217f4e8640187fc90baa7f0a9f8efb8aebfb0e059e43bebeb8
 tags:
-  3.17.0: sha256:9c2298a3a3fae58b1c6a5d4b4fc6e323d23a8855a687535060a2c344511faf1b
-  latest: sha256:9c2298a3a3fae58b1c6a5d4b4fc6e323d23a8855a687535060a2c344511faf1b
+  3.17.0: sha256:0301861f6d2aea217f4e8640187fc90baa7f0a9f8efb8aebfb0e059e43bebeb8
+  latest: sha256:0301861f6d2aea217f4e8640187fc90baa7f0a9f8efb8aebfb0e059e43bebeb8
 aliases:
   valgrind: /opt/view/bin/valgrind
   valgrind-di-server: /opt/view/bin/valgrind-di-server
