+++ registry/ghcr.io/autamus/accumulo/container.yaml
@@ -4,7 +4,7 @@ maintainer: '@vsoch'
 description: Apache Accumulo is a sorted, distributed key/value store that provides
   robust, scalable data storage and retrieval.
 latest:
-  2.0.1: sha256:5ad3905517cb8ae1dede4b3f80f9fc46eae9456abf681d91236474c326fb463c
+  2.0.1: sha256:aeef479dc7232f57278befecf024aa7a33aa88971792b7ed15152d826a486f0e
 tags:
-  2.0.1: sha256:5ad3905517cb8ae1dede4b3f80f9fc46eae9456abf681d91236474c326fb463c
-  latest: sha256:5ad3905517cb8ae1dede4b3f80f9fc46eae9456abf681d91236474c326fb463c
+  2.0.1: sha256:aeef479dc7232f57278befecf024aa7a33aa88971792b7ed15152d826a486f0e
+  latest: sha256:aeef479dc7232f57278befecf024aa7a33aa88971792b7ed15152d826a486f0e
