+++ registry/ghcr.io/autamus/corset/container.yaml
@@ -4,10 +4,10 @@ maintainer: '@vsoch'
 description: Corset is a command-line program to go from a de novo transcriptome assembly
   to gene-level counts.
 latest:
-  "1.09": sha256:bbb486ecf8eecfec2840f78f845110cfcbe61ca23f3b932e4005eee5de83ee45
+  "1.09": sha256:ae6cf76aee7bd57e7e1acd554bb66e7b2ba8c115c7c1fbd7e6aaea577738eb27
 tags:
-  "1.09": sha256:bbb486ecf8eecfec2840f78f845110cfcbe61ca23f3b932e4005eee5de83ee45
-  latest: sha256:bbb486ecf8eecfec2840f78f845110cfcbe61ca23f3b932e4005eee5de83ee45
+  "1.09": sha256:ae6cf76aee7bd57e7e1acd554bb66e7b2ba8c115c7c1fbd7e6aaea577738eb27
+  latest: sha256:ae6cf76aee7bd57e7e1acd554bb66e7b2ba8c115c7c1fbd7e6aaea577738eb27
 aliases:
   corset: /opt/view/bin/corset
   corset_fasta_ID_changer: /opt/view/bin/corset_fasta_ID_changer
