+++ registry/ghcr.io/autamus/bbmap/container.yaml
@@ -4,7 +4,7 @@ maintainer: '@vsoch'
 description: A suite of fast, multithreaded bioinformatics tools designed for analysis
   of DNA and RNA sequence data.
 latest:
-  "38.63": sha256:e9187056011f6000c68b52016659d888a89e109976d2966fab5edee281990272
+  "38.63": sha256:1ca6d41b57170c8ef19218830bb1c99400719a3825b7a69a3d6406568deab7f0
 tags:
-  "38.63": sha256:e9187056011f6000c68b52016659d888a89e109976d2966fab5edee281990272
-  latest: sha256:e9187056011f6000c68b52016659d888a89e109976d2966fab5edee281990272
+  "38.63": sha256:1ca6d41b57170c8ef19218830bb1c99400719a3825b7a69a3d6406568deab7f0
+  latest: sha256:1ca6d41b57170c8ef19218830bb1c99400719a3825b7a69a3d6406568deab7f0

[…]
