+++ _library/bids-brainsautoworkup.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/brainsautoworkup"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/brainsautoworkup/container.yaml"
-updated_at: "2021-06-06 20:44:28.249190"
+updated_at: "2021-06-07 16:22:47.285631"
 container_url: "https://hub.docker.com/r/bids/brainsautoworkup"
 
 versions:
