### 2021-06-07 #3311 @vsoch
@zbeekman @rcannood since the review started I had new centers request additional container technologies (Podman, Docker) for shpc, and they are now released. Would you like me to update the paper to reflect this?
