### 2021-06-07 e94498f
updates to several container and module bases
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
