### 2021-06-08 5e2b88a
adding back PODMAN_OPTS and restoring conflict of alias.name to tcl
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
