+++ shpc/settings.yml
@@ -21,7 +21,7 @@ registry: $root_dir/registry
 module_base: $root_dir/modules
 
 # Format string for module commands exec,run,shell (not aliases) (can also include:
-# {{registry}} , {{ namespace }}, {{tool}}, {{version}}
+# {{registry}} , {{ repository }}, {{tool}}, {{version}}
 # This is where you might add a prefix to your module names, if desired.
 module_name: '{{ tool }}'
 
