### 2021-06-08 d129dad
several updates to config and environment settings
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
