+++ shpc/main/container/docker.py
@@ -60,9 +60,11 @@ class DockerContainer(ContainerTechnology):
         if pull_type != "docker":
             logger.exit("%s only supports Docker (oci registry) pulls." % self.command)
 
-        # Podman doesn't keep a record of digest->tag, so we use tag
-        uri = "%s:%s" % (self.add_registry(config.docker), tag.name)
-        return self.pull(uri)
+        tag_uri = "%s:%s" % (self.add_registry(config.docker), tag.name)
+        tag_digest = "%s@%s" % (self.add_registry(config.docker), tag.digest)
+        self.pull(tag_digest)
+        # Podman doesn't keep a record of digest->tag, so we tag after
+        return self.tag(tag_digest, tag_uri)
 
     def pull(self, uri):
         """
@@ -72,6 +74,15 @@ class DockerContainer(ContainerTechnology):
         if res["return_code"] != 0:
             logger.exit("There was an issue pulling %s" % uri)
         return uri
+
+    def tag(self, image, tag_as):
+        """
+        Given a container URI, tag as something else.
+        """
+        res = shpc.utils.run_command([self.command, "tag", image, tag_as])
+        if res["return_code"] != 0:
+            logger.exit("There was an issue tagging %s as %s" % (image, tag_as))
+        return tag_as
 
     def inspect(self, image):
         """
@@ -118,7 +129,7 @@ class DockerContainer(ContainerTechnology):
         """
         image = self.get(image)
         if self.exists(image):
-            shpc.utils.run_command([self.command, "rmi", image])
+            shpc.utils.run_command([self.command, "rmi", "--force", image])
 
     def check(self, module_name, config):
         """
