### 2021-06-09 #3311 @zbeekman
> @zbeekman @rcannood since the paper was submit I had new centers request additional container technologies (Podman, Docker) for shpc, and they are now released. Would you like me to update the paper to reflect this?

I have a minor preference for updating the paper to reflect the new capabilities, but this is absolutely *NOT* a deal breaker for me either way, and I think you should probably do what makes the most sense for you. Thanks for the heads up though!!! Easier for me to play around with if it includes docker too!
