### 2021-06-10 eb43118
temporary fix of moving registry entiry to end (so no comments are eaten) and printing warning
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
