### 2021-06-10 e83ce03
ensuring that multiple podman images are deleted if no tag is provided
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
