### 2021-06-10 fdc3b1f
fix for ruamel deleting comments - we need to preserve its type for the list
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
