### 2021-06-10 67517dc
fixing typo for shpc uninstall
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
