### 2021-06-11 c66051b
try doing uninstall after test, and looking at docker/podman images before and after
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
