+++ .github/workflows/test.yml
@@ -101,6 +101,7 @@ jobs:
           grep --quiet 'Python 3.9.5' test_output
           rm test_output
           shpc uninstall --force python:3.9.5-alpine 
+
         
       - name: Run python module tests (tcsh)
         shell: tcsh -e {0}
