### 2021-06-11 7387d95
Several updates to config and environment settings (#403)
* several updates to config and environment settings
* temporary fix of moving registry entiry to end (so no comments are eaten) and printing warning
* ensuring that multiple podman images are deleted if no tag is provided
* fix for ruamel deleting comments - we need to preserve its type for the list
* fixing typo for shpc uninstall
* try doing uninstall after test, and looking at docker/podman images before and after

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
