+++ _library/bids-baracus.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/baracus"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/baracus/container.yaml"
-updated_at: "2021-06-08 15:06:53.102923"
+updated_at: "2021-06-11 05:47:05.998596"
 container_url: "https://hub.docker.com/r/bids/baracus"
 
 versions:
