+++ _library/ghcr.io-autamus-igraph.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/igraph"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/igraph/container.yaml"
-updated_at: "2021-06-08 15:06:24.868280"
+updated_at: "2021-06-11 05:46:39.648083"
 container_url: "https://github.com/orgs/autamus/packages/container/package/igraph"
 
 versions:
