### 2021-06-11 6c64d36
properly adding subparser help strings.
We need to use description and not help

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
