+++ _library/vanessa-salad.md
@@ -3,7 +3,7 @@ layout: container
 name:  "vanessa/salad"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/vanessa/salad/container.yaml"
-updated_at: "2021-06-11 05:47:12.723168"
+updated_at: "2021-06-11 05:58:33.822191"
 container_url: "https://hub.docker.com/r/vanessa/salad"
 aliases:
  - "salad"
