+++ _library/mariadb.md
@@ -3,7 +3,7 @@ layout: container
 name:  "mariadb"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/mariadb/container.yaml"
-updated_at: "2021-06-11 05:47:14.324823"
+updated_at: "2021-06-11 05:58:35.161917"
 container_url: "https://hub.docker.com/r/_/mariadb"
 aliases:
  - "mariabackup"
