+++ _library/ghcr.io-autamus-cfitsio.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/cfitsio"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/cfitsio/container.yaml"
-updated_at: "2021-06-11 05:46:48.544559"
+updated_at: "2021-06-11 05:58:08.981813"
 container_url: "https://github.com/orgs/autamus/packages/container/package/cfitsio"
 
 versions:
