+++ _library/postgres.md
@@ -3,7 +3,7 @@ layout: container
 name:  "postgres"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/postgres/container.yaml"
-updated_at: "2021-06-11 05:47:09.003452"
+updated_at: "2021-06-11 05:58:30.222108"
 container_url: "https://hub.docker.com/r/_/postgres"
 aliases:
  - "clusterdb"
