+++ registry/rocker/tidyverse/container.yaml
@@ -3,9 +3,11 @@ url: https://hub.docker.com/r/rocker/tidyverse
 maintainer: '@vsoch'
 description: 'Version-stable build of R, rstudio, and R packages '
 latest:
-  latest: sha256:f9671fa9329160cc57f76ec822896467f420f4b8d55bf3a811293b9f94283a3e
+  latest: sha256:cdf639fd99b81f5b219a689d489489540e9a0d139336003b015cde4b047a294d
 tags:
-  latest: sha256:f9671fa9329160cc57f76ec822896467f420f4b8d55bf3a811293b9f94283a3e
+  latest: sha256:cdf639fd99b81f5b219a689d489489540e9a0d139336003b015cde4b047a294d
+filter:
+- latest
 aliases:
   R: /usr/local/bin/R
   Rscript: /usr/local/bin/Rscript
@@ -15,5 +17,3 @@ aliases:
   rserver-pam: /usr/lib/rstudio-server/bin/rserver-pam
   rsession: /usr/lib/rstudio-server/bin/rsession
   rstudio-server: /usr/lib/rstudio-server/bin/rstudio-server
-filter:
-- latest
