+++ _library/bids-brainsautoworkup.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/brainsautoworkup"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/brainsautoworkup/container.yaml"
-updated_at: "2021-06-11 05:58:25.674335"
+updated_at: "2021-06-13 00:50:25.985576"
 container_url: "https://hub.docker.com/r/bids/brainsautoworkup"
 
 versions:
