+++ _library/ghcr.io-autamus-git.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/git"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/git/container.yaml"
-updated_at: "2021-06-11 05:58:02.967576"
+updated_at: "2021-06-13 00:50:00.487983"
 container_url: "https://github.com/orgs/autamus/packages/container/package/git"
 aliases:
  - "git"
