+++ _library/biocontainers-talon.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/talon"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/talon/container.yaml"
-updated_at: "2021-06-11 05:58:31.928561"
+updated_at: "2021-06-13 00:50:33.349085"
 container_url: "https://hub.docker.com/r/biocontainers/talon"
 aliases:
  - "talon"
