+++ _library/solr.md
@@ -3,7 +3,7 @@ layout: container
 name:  "solr"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/solr/container.yaml"
-updated_at: "2021-06-11 05:58:35.735532"
+updated_at: "2021-06-13 00:50:37.607983"
 container_url: "https://hub.docker.com/_/solr"
 aliases:
  - "post"
