+++ registry/ghcr.io/autamus/dyninst/container.yaml
@@ -3,6 +3,7 @@ url: https://github.com/orgs/autamus/packages/container/package/dyninst
 maintainer: '@vsoch'
 description: DyninstAPI - Tools for binary instrumentation, analysis, and modification.
 latest:
-  11.0.0: sha256:c684f2bffc592e8ae8f82e977b93b6d8306483d5ca71dda81fa8828663122e3f
+  11.0.1: sha256:feb4f8079cc23407013b8c61709d5d40133496fbb9c26c58559928d5d2dee747
 tags:
   11.0.0: sha256:c684f2bffc592e8ae8f82e977b93b6d8306483d5ca71dda81fa8828663122e3f
+  11.0.1: sha256:feb4f8079cc23407013b8c61709d5d40133496fbb9c26c58559928d5d2dee747
