+++ registry/quay.io/biocontainers/trinity/container.yaml
@@ -1,0 +1,12 @@
+docker: quay.io/biocontainers/trinity
+latest:
+ "2.12.0": "sha256:2736617a96c30bce93a149fe5f47b4f4a290a7d21c1503619430db619d9bc3a0"
+tags:
+ "2.12.0": "sha256:2736617a96c30bce93a149fe5f47b4f4a290a7d21c1503619430db619d9bc3a0"
+ "2.11.0": "sha256:6d35d716aa12ba7b0c715fa8a30359b43ba9151e854dc4407b949e7b57c3a50a"
+maintainer: "@audreystott"
+description: "Trinity, developed at the Broad Institute and the Hebrew University of Jerusalem, represents a novel method for the efficient and robust de novo reconstruction of transcriptomes from RNA-seq data."
+url: https://quay.io/repository/biocontainers/trinity
+aliases:
+  trinity: /usr/local/bin/Trinity
+  Trinity: /usr/local/bin/Trinity
