+++ registry/quay.io/biocontainers/samtools/container.yaml
@@ -1,0 +1,12 @@
+docker: quay.io/biocontainers/samtools
+latest:
+ "1.12": "sha256:5fd5f0937adf8a24b5bf7655110e501df78ae51588547c8617f17c3291a723e1"
+tags:
+ "1.12": "sha256:5fd5f0937adf8a24b5bf7655110e501df78ae51588547c8617f17c3291a723e1"
+ "1.11": "sha256:141120f19f849b79e05ae2fac981383988445c373b8b5db7f3dd221179af382b"
+ "1.10": "sha256:84a8d0c0acec87448a47cefa60c4f4a545887239fcd7984a58b48e7a6ac86390"
+maintainer: "@audreystott"
+description: "Tools for reading/writing/editing/indexing/viewing SAM/BAM/CRAM format."
+url: https://quay.io/repository/biocontainers/samtools
+aliases:
+  samtools: /usr/local/bin/samtools
