### 2021-06-22 6001e7a
Update registry/quay.io/biocontainers/spades/container.yaml
Co-authored-by: Vanessasaurus <814322+vsoch@users.noreply.github.com>
