### 2021-06-22 85e5116
added quay.io/biocontainers tools (#409)
* added quay.io/biocontainers tools
* Update registry/quay.io/biocontainers/spades/container.yaml

Co-authored-by: Vanessasaurus <814322+vsoch@users.noreply.github.com>
Co-authored-by: audreystott <audrey.stott@pawsey.org.au>
