+++ registry/quay.io/biocontainers/spades/container.yaml
@@ -1,0 +1,11 @@
+docker: quay.io/biocontainers/spades
+latest:
+ "3.15.2": "sha256:250657af71d37fafdff81af644e92bf84485d4d9b6894f58ff84efdc1a6ec9eb"
+tags:
+ "3.15.2": "sha256:250657af71d37fafdff81af644e92bf84485d4d9b6894f58ff84efdc1a6ec9eb"
+ "3.15.0": "sha256:4cc7e41f53d3effb25b59afedc3fbd2d3e08fff449a507ecfa6731ff6d81ff79"
+maintainer: "@audreystott"
+description: "SPAdes – St. Petersburg genome assembler – is an assembly toolkit containing various assembly pipelines."
+url: https://quay.io/repository/biocontainers/spades
+aliases:
+  spades.py: /usr/local/bin/spades.py
