+++ _library/ghcr.io-autamus-bbmap.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bbmap"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bbmap/container.yaml"
-updated_at: "2021-06-20 01:56:57.507607"
+updated_at: "2021-06-22 03:45:30.940197"
 container_url: "https://github.com/orgs/autamus/packages/container/package/bbmap"
 
 versions:
