+++ registry/ghcr.io/autamus/scr/container.yaml
@@ -1,0 +1,46 @@
+docker: ghcr.io/autamus/scr
+url: https://github.com/orgs/autamus/packages/container/package/scr
+maintainer: '@vsoch'
+description: SCR caches checkpoint data in storage on the compute nodes of a Linux cluster to provide a fast, scalable checkpoint/restart capability for MPI codes.
+latest:
+  "3.0.rc.1": sha256:86b2b8c8b530e17b01714be42e37aaa85061015aefd7683fa3332cd0aa61e958
+tags:
+  "3.0.rc.1": sha256:86b2b8c8b530e17b01714be42e37aaa85061015aefd7683fa3332cd0aa61e958s
+alieses:
+  scancel: /opt/view/bin/scancel
+  scontrol: /opt/view/bin/scontrol
+  scp: /opt/view/bin/scp
+  scr_check_node: /opt/view/bin/scr_check_node
+  scr_ckpt_interval.py: /opt/view/bin/scr_ckpt_interval.py
+  scr_copy: /opt/view/bin/scr_copy
+  scr_crc32: /opt/view/bin/scr_crc32
+  scr_env: /opt/view/bin/scr_env
+  scr_flush_file: /opt/view/bin/scr_flush_file
+  scr_get_jobstep_id: /opt/view/bin/scr_get_jobstep_id
+  scr_glob_hosts: /opt/view/bin/scr_glob_hosts
+  scr_halt: /opt/view/bin/scr_halt
+  scr_halt_cntl: /opt/view/bin/scr_halt_cntl
+  scr_index: /opt/view/bin/scr_index
+  scr_inspect: /opt/view/bin/scr_inspect
+  scr_kill_jobstep: /opt/view/bin/scr_kill_jobstep
+  scr_list_dir: /opt/view/bin/scr_list_dir
+  scr_list_down_nodes: /opt/view/bin/scr_list_down_nodes
+  scr_log_event: /opt/view/bin/scr_log_event
+  scr_log_transfer: /opt/view/bin/scr_log_transfer
+  scr_nodes_file: /opt/view/bin/scr_nodes_file
+  scr_postrun: /opt/view/bin/scr_postrun
+  scr_prefix: /opt/view/bin/scr_prefix
+  scr_prerun: /opt/view/bin/scr_prerun
+  scr_print: /opt/view/bin/scr_print
+  scr_rebuild_partner: /opt/view/bin/scr_rebuild_partner
+  scr_rebuild_rs: /opt/view/bin/scr_rebuild_rs
+  scr_rebuild_xor: /opt/view/bin/scr_rebuild_xor
+  scr_retries_halt: /opt/view/bin/scr_retries_halt
+  scr_run: /opt/view/bin/scr_run
+  scr_scavenge: /opt/view/bin/scr_scavenge
+  scr_srun: /opt/view/bin/scr_srun
+  scr_test_datemanip: /opt/view/bin/scr_test_datemanip
+  scr_test_runtime: /opt/view/bin/scr_test_runtime
+  scr_watchdog: /opt/view/bin/scr_watchdog
+  scrlog.py: /opt/view/bin/scrlog.py
+  scrontab: /opt/view/bin/scrontab
