+++ registry/ghcr.io/autamus/unifyfs/container.yaml
@@ -1,0 +1,12 @@
+docker: ghcr.io/autamus/unifyfs
+url: https://github.com/orgs/autamus/packages/container/package/unifyfs
+maintainer: '@vsoch'
+description: User level file system that enables applications to use node-local storage as burst buffers for shared files.
+latest:
+  "0.9.2": sha256:401d1041f3daf9b1fb52306fce6fbc58df48a64ba225da752f8d3b16da85e65a
+tags:
+  "0.9.2": sha256:401d1041f3daf9b1fb52306fce6fbc58df48a64ba225da752f8d3b16da85e65a
+aliases:
+  unifyfs: /opt/view/bin/unifyfs
+  unifyfs-config: /opt/view/bin/unifyfs-config
+  unifyfsd: /opt/view/bin/unifyfsd
