### 2021-06-23 f41d5d0
retest with public images
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
