### 2021-06-23 7bff5be
First page of new autamus packages (#411)
* first page of new autamus packages
* retest with public images
* update docs

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>
