+++ registry/ghcr.io/autamus/turbine/container.yaml
@@ -1,0 +1,13 @@
+docker: ghcr.io/autamus/turbine
+url: https://github.com/orgs/autamus/packages/container/package/turbine
+maintainer: '@vsoch'
+description: Turbine is the Swift/T runtime 
+latest:
+  1.3.0: sha256:076739ee1f6754f6cbf43dcdaedf021410529c195a671ca099b3a3e11160b611 
+tags:
+  1.3.0: sha256:076739ee1f6754f6cbf43dcdaedf021410529c195a671ca099b3a3e11160b611 
+aliases:
+  turbine: /opt/view/bin/turbine
+  turbine-pilot: /opt/view/bin/turbine-pilot
+  turbine-read-doubles: /opt/view/bin/turbine-read-doubles
+  turbine-write-doubles: /opt/view/bin/turbine-write-doubles
