+++ registry/ghcr.io/autamus/rsync/container.yaml
@@ -1,0 +1,11 @@
+docker: ghcr.io/autamus/rsync
+url: https://github.com/orgs/autamus/packages/container/package/rsync
+maintainer: '@vsoch'
+description: An open source utility that provides fast incremental file transfer. https://rsync.samba.org/
+latest:
+  "3.2.2": sha256:99c4780710b1d844ca8998dea4c59b2d3b54c28ac206dc3ee580ca2c34f455d1
+tags:
+  "3.2.2": sha256:99c4780710b1d844ca8998dea4c59b2d3b54c28ac206dc3ee580ca2c34f455d1
+aliases:
+  rsync: /opt/view/bin/rsync
+  rsync-ssl: /opt/view/bin/rsync-ssl
