+++ .github/workflows/test-container.yml
@@ -5,15 +5,54 @@ on:
     types: [labeled]
 
 jobs:
-  test-registry:
+  changes:
     if: ${{ github.event.label.name == 'container-recipe' }}
     runs-on: ubuntu-latest
+    outputs:
+      changed_recipes: ${{ steps.files.outputs.added_modified }}
     steps:
+      - name: Find changed recipes
+        id: files
+        uses: jitterbit/get-changed-files@b17fbb00bdc0c0f63fcf166580804b4d2cdc2a42
+        with:
+          format: 'json'
+  
+  test-recipes:
+    needs:
+      - changes
+    runs-on: ubuntu-latest
+    strategy:
+      # Keep going on other deployments if anything bloops
+      fail-fast: false
+      matrix:
+        changed_recipe: ${{ fromJson(needs.changes.outputs.changed_recipes) }}
+
+    name: Check ${{ matrix.changed_recipe }}
+    steps:
+      - name: Continue
+        run: |
+          # Continue if we have a changed recipe file
+          if [[ "${{ matrix.changed_recipe }}" = registry* ]]; then
+            echo "keepgoing=true" >> $GITHUB_ENV
+          fi          
+
+      - name: Checkout
+        if: ${{ env.keepgoing == 'true' }}
+        uses: actions/checkout@v2        
+
+      - name: Make Space For Build
+        if: ${{ env.keepgoing == 'true' }}
+        run: |
+          sudo rm -rf /usr/share/dotnet
+          sudo rm -rf /opt/ghc
+
       - name: Install Dependencies
+        if: ${{ env.keepgoing == 'true' }}
         run: |
           sudo apt-get install lua5.2 liblua5.2-dev lua-filesystem lua-posix tcl tcl-dev
 
       - name: Install Lmod
+        if: ${{ env.keepgoing == 'true' }}
         run: |
           PKG_VERSION=8.4.27
           PKG_URL="https://github.com/TACC/Lmod/archive/${PKG_VERSION}.tar.gz"
@@ -23,61 +62,33 @@ jobs:
           ./configure --prefix=/usr/share && make && make install
 
       - uses: eWaterCycle/setup-singularity@v6
+        if: ${{ env.keepgoing == 'true' }}
         name: Install Singularity
         with:
           singularity-version: 3.6.4
 
-      - name: Derive number of commits
-        run: |
-            cd /opt
-            branch="${{ github.head_ref }}"
-            git config --global user.email "github-actions"
-            git config --global user.name "github-actions@noreply.github.users.com"
-            git clone -b main https://github.com/singularityhub/singularity-hpc
-            cd singularity-hpc
-            git checkout -b $branch
-            git pull origin $branch
-            commits=$(git log --oneline "$branch" ^main | wc -l)
-            echo "Found $commits since main"
-            cd registry
-            touch changes.txt
-            # Note that you can also do: git diff --name-only $branch..main .
-            # We show the count of commits as a sanity check to the developer
-            for container_yaml in $(git diff --name-only HEAD~$commits..HEAD .); do
-                echo $container_yaml >> changes.txt
-            done            
-            cat changes.txt
-
-      - name: Save changelist
-        uses: actions/upload-artifact@v2
-        with:
-          name: changes.txt
-          path: /opt/singularity-hpc/registry/changes.txt
-
       - name: Create conda environment
+        if: ${{ env.keepgoing == 'true' }}
         run: conda create --quiet -c conda-forge --name shpc spython
 
       - uses: actions/checkout@v2
+        if: ${{ env.keepgoing == 'true' }}
+
       - name: Install shpc
+        if: ${{ env.keepgoing == 'true' }}
         run: |
           export PATH="/usr/share/miniconda/bin:$PATH"
           source activate shpc
           pip install -e .
 
[…]
