+++ registry/ghcr.io/autamus/emacs/container.yaml
@@ -1,0 +1,13 @@
+docker: ghcr.io/autamus/emacs
+url: https://github.com/orgs/autamus/packages/container/package/emacs
+maintainer: "@vsoch"
+description: The Emacs programmable text editor. https://www.gnu.org/software/emacs
+latest:
+  "27.2": sha256:a1fa851f9dbd3ba0474fd8f0365f23ba5cedc46095c5f6dedfa5fd433652a84a
+tags:
+  "27.2": sha256:a1fa851f9dbd3ba0474fd8f0365f23ba5cedc46095c5f6dedfa5fd433652a84a
+aliases:
+  ebrowse: /opt/view/bin/ebrowse
+  emacs: /opt/view/bin/emacs
+  emacsclient: /opt/view/bin/emacsclient
+  etags: /opt/view/bin/etags
