+++ registry/ghcr.io/autamus/umap/container.yaml
@@ -1,0 +1,8 @@
+docker: ghcr.io/autamus/umap
+url: https://github.com/orgs/autamus/packages/container/package/umap
+maintainer: '@vsoch'
+description: Umap is a library that provides an mmap()-like interface to a simple, user-space page fault handler based on the userfaultfd Linux feature (starting with 4.3 linux kernel).
+latest:
+  2.1.0: sha256:b81554c3aba70bf5d093ff8a9c76d504757a70aa9c9415c219cfe831c46a8edd
+tags:
+  2.1.0: sha256:b81554c3aba70bf5d093ff8a9c76d504757a70aa9c9415c219cfe831c46a8edd
