+++ _library/ghcr.io-autamus-bracken.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bracken"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bracken/container.yaml"
-updated_at: "2021-06-22 03:45:38.216949"
+updated_at: "2021-06-23 01:49:23.195816"
 container_url: "https://github.com/orgs/autamus/packages/container/package/bracken"
 
 versions:
