+++ registry/ghcr.io/autamus/superlu/container.yaml
@@ -1,0 +1,8 @@
+docker: ghcr.io/autamus/superlu
+url: https://github.com/orgs/autamus/packages/container/package/superlu
+maintainer: '@vsoch'
+description: SuperLU is a general purpose library for the direct solution of large, sparse, nonsymmetric systems of linear equations on high performance machines.
+latest:
+  "5.2.2": sha256:c0498deeb4c5bc53139ec41e441cff30c85f323b1551e569ed2176b0bb7f6039
+tags:
+  "5.2.2": sha256:c0498deeb4c5bc53139ec41e441cff30c85f323b1551e569ed2176b0bb7f6039
