+++ registry/ghcr.io/autamus/libquo/container.yaml
@@ -1,0 +1,10 @@
+docker: ghcr.io/autamus/libquo
+url: https://github.com/orgs/autamus/packages/container/package/libquo
+maintainer: '@vsoch'
+description: QUO (as in status quo) is a runtime library that aids in accommodating thread-level heterogeneity in dynamic, phased MPI+X applications comprising single- and multi-threaded libraries. 
+latest:
+  "1.3.1": sha256:ad579fc629317a0a881c982b953899737ac618f83dcfbfa4905966cac08845e6
+tags:
+  "1.3.1": sha256:ad579fc629317a0a881c982b953899737ac618f83dcfbfa4905966cac08845e6
+aliases:
+  quo-info: /opt/view/bin/quo-info
