+++ registry/ghcr.io/autamus/legion/container.yaml
@@ -1,0 +1,12 @@
+docker: ghcr.io/autamus/legion
+url: https://github.com/orgs/autamus/packages/container/package/legion
+maintainer: '@vsoch'
+description: Legion is a data-centric parallel programming system for writing portable high performance programs targeted at distributed heterogeneous architectures.
+latest:
+  "21.03.0": sha256:466efaad695294bd3ec28082c74d141b96af2032392121df432a09fcc39d1340
+tags:
+  "21.03.0": sha256:466efaad695294bd3ec28082c74d141b96af2032392121df432a09fcc39d1340
+aliases:
+  legion_prof.py: /opt/view/bin/legion_prof.py
+  legion_serializer.py: /opt/view/bin/legion_serializer.py
+  legion_spy.py: /opt/view/bin/legion_spy.py
