### 2021-06-23 2ef7ce7
adding last page of autamus new packages
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
