+++ registry/ghcr.io/autamus/conduit/container.yaml
@@ -1,0 +1,16 @@
+docker: ghcr.io/autamus/conduit
+url: https://github.com/orgs/autamus/packages/container/package/conduit
+maintainer: '@vsoch'
+description: Conduit is an open source project from Lawrence Livermore National Laboratory that provides an intuitive model for describing hierarchical scientific data in C++, C, Fortran, and Python. 
+latest:
+  "0.7.2": sha256:6f8933a27f2f71f52787a5f27fa6f3b84a5f850e4c5cf26a79e4e0a360954c74
+tags:
+  "0.7.2": sha256:6f8933a27f2f71f52787a5f27fa6f3b84a5f850e4c5cf26a79e4e0a360954c74
+aliases:
+  conduit_blueprint_verify: /opt/view/bin/conduit_blueprint_verify
+  conduit_relay_entangle.py: /opt/view/bin/conduit_relay_entangle.py
+  conduit_relay_io_convert: /opt/view/bin/conduit_relay_io_convert
+  conduit_relay_io_ls: /opt/view/bin/conduit_relay_io_ls
+  conduit_relay_node_viewer: /opt/view/bin/conduit_relay_node_viewer
+  conduit_staging: /opt/view/bin/conduit_staging
+  conduit_staging.sh: /opt/view/bin/conduit_staging.sh
