+++ registry/ghcr.io/autamus/ginkgo/container.yaml
@@ -1,0 +1,8 @@
+docker: ghcr.io/autamus/ginkgo
+url: https://github.com/orgs/autamus/packages/container/package/ginkgo
+maintainer: '@vsoch'
+description: High-performance linear algebra library for manycore systems, with a focus on sparse solution of linear systems. 
+latest:
+  "1.3.0": sha256:cd43879f08cf9e5cd3f2b76a03aafdaeeb2f6944128893eaa0e3b4cabd9039c1
+tags:
+  "1.3.0": sha256:cd43879f08cf9e5cd3f2b76a03aafdaeeb2f6944128893eaa0e3b4cabd9039c1
