+++ _library/ghcr.io-autamus-libquo.md
@@ -1,0 +1,131 @@
+---
+layout: container
+name:  "ghcr.io/autamus/libquo"
+maintainer: "@vsoch"
+github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/libquo/container.yaml"
+updated_at: "2021-06-23 03:27:39.608011"
+container_url: "https://github.com/orgs/autamus/packages/container/package/libquo"
+aliases:
+ - "quo-info"
+
+versions:
+ - "1.3.1"
+description: "QUO (as in status quo) is a runtime library that aids in accommodating thread-level heterogeneity in dynamic, phased MPI+X applications comprising single- and multi-threaded libraries."
+---
+
+This module is a singularity container wrapper for ghcr.io/autamus/libquo.
+QUO (as in status quo) is a runtime library that aids in accommodating thread-level heterogeneity in dynamic, phased MPI+X applications comprising single- and multi-threaded libraries.
+After [installing shpc](#install) you will want to install this container module:
+
+```bash
+$ shpc install ghcr.io/autamus/libquo
+```
+
+Or a specific version:
+
+```bash
+$ shpc install ghcr.io/autamus/libquo:1.3.1
+```
+
+And then you can tell lmod about your modules folder:
+
+```bash
+$ module use ./modules
+```
+
+And load the module, and ask for help, or similar.
+
+```bash
+$ module load ghcr.io/autamus/libquo/1.3.1
+$ module help ghcr.io/autamus/libquo/1.3.1
+```
+
+You can use tab for auto-completion of module names or commands that are provided.
+
+<br>
+
+### Commands
+
+When you install this module, you'll be able to load it to make the following commands accessible.
+Examples for both Singularity, Podman, and Docker (container technologies supported) are included.
+
+#### -run:
+
+```bash
+$ singularity run <container>
+$ podman run --rm  -v ${PWD} -w ${PWD} <container>
+$ docker run --rm  -v ${PWD} -w ${PWD} <container>
+```
+
+#### -shell:
+
+```bash
+$ singularity shell -s /bin/sh <container>
+$ podman run --it --rm --entrypoint /bin/sh  -v ${PWD} -w ${PWD} <container>
+$ docker run --it --rm --entrypoint /bin/sh  -v ${PWD} -w ${PWD} <container>
+```
+
+#### -exec:
+
+```bash
+$ singularity exec -s /bin/sh <container> "$@"
+$ podman run --it --rm --entrypoint ""  -v ${PWD} -w ${PWD} <container> "$@"
+$ docker run --it --rm --entrypoint ""  -v ${PWD} -w ${PWD} <container> "$@"
+```
+
+#### -inspect:
+
+Podman and Docker only have one inspect type.
+
+```bash
+$ podman inspect <container>
+$ docker inspect <container>
+```
+
+#### -inspect-runscript:
+
+```bash
+$ singularity inspect -r <container>
+```
+
+#### -inspect-deffile:
+
+```bash
+$ singularity inspect -d <container>
+```
+
+
+#### quo-info
+       
+```bash
+$ singularity exec <container> /opt/view/bin/quo-info
+$ podman run --it --rm --entrypoint /opt/view/bin/quo-info   -v ${PWD} -w ${PWD} <container> -c " $@"
+$ docker run --it --rm --entrypoint /opt/view/bin/quo-info   -v ${PWD} -w ${PWD} <container> -c " $@"
+```
+
+
+
+In the above, the `<container>` directive will reference an actual container provided
+by the module, for the version you have chosen to load. An environment file in the
+module folder will also be bound. Note that although a container
+might provide custom commands, every container exposes unique exec, shell, run, and
[…]
