+++ _library/ghcr.io-autamus-tau.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/tau"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/tau/container.yaml"
-updated_at: "2021-06-23 01:49:08.169443"
+updated_at: "2021-06-23 03:27:42.200679"
 container_url: "https://github.com/orgs/autamus/packages/container/package/tau"
 
 versions:
