+++ registry/ghcr.io/autamus/mpc/container.yaml
@@ -4,7 +4,7 @@ maintainer: '@vsoch'
 description: 'mithi/mpc: A software pipeline using the Model Predictive Control method
   to drive a car around a virtual track.'
 latest:
-  1.1.0: sha256:dd66715bc94695d9638cd2ee70b562b8c2ab171adb46d0c6f0862338732dc7aa
+  1.1.0: sha256:497deb4c62bc1e7c6a0571eac110234cf49e97927771148042cd6218d674b3c3
 tags:
-  1.1.0: sha256:dd66715bc94695d9638cd2ee70b562b8c2ab171adb46d0c6f0862338732dc7aa
-  latest: sha256:dd66715bc94695d9638cd2ee70b562b8c2ab171adb46d0c6f0862338732dc7aa
+  1.1.0: sha256:497deb4c62bc1e7c6a0571eac110234cf49e97927771148042cd6218d674b3c3
+  latest: sha256:497deb4c62bc1e7c6a0571eac110234cf49e97927771148042cd6218d674b3c3
