+++ registry/ghcr.io/autamus/mercury/container.yaml
@@ -1,8 +1,8 @@
 docker: ghcr.io/autamus/mercury
 url: https://github.com/orgs/autamus/packages/container/package/mercury
 maintainer: '@vsoch'
-description: Mercury is a C library for implementing RPC, optimized for HPC 
+description: Mercury is a C library for implementing RPC, optimized for HPC
 latest:
-  "2.0.1": sha256:0f5161fc2a16dceacb037bbf60d030940827dfaad0c9177212b4edeb873d3f9e
+  2.0.1: sha256:97737711d7033eaa759643a1d093ea8071ab490d3577bd85b883d8b32d0494e4
 tags:
-  "2.0.1": sha256:0f5161fc2a16dceacb037bbf60d030940827dfaad0c9177212b4edeb873d3f9e
+  2.0.1: sha256:97737711d7033eaa759643a1d093ea8071ab490d3577bd85b883d8b32d0494e4
