+++ registry/mongo/container.yaml
@@ -8,9 +8,9 @@ latest:
   5.0.0-rc1-focal: sha256:238ca9b76e3034377a38be3dac927bd3920634c3af68053be000739c15106cec
 tags:
   4.4.5-bionic: sha256:cc8bb8711114fa726ca641f5644791d3d43bdd2c95dac541a5fa2c90a2e6e972
-  4.4.6-bionic: sha256:ee8a1b33bdaeb890aa9e806b5e69161614b6bd067d57a939be752895b37c5b12
+  4.4.6-bionic: sha256:bcb400212394146860409d89c18dd4231d7e144b60783869c40e082561f52599
   5.0.0-rc1-focal: sha256:238ca9b76e3034377a38be3dac927bd3920634c3af68053be000739c15106cec
-  latest: sha256:be0b7f6b4e83c39d598696d18e5264368bb2cd8230dcb1e2546d025ed5a3dab9
+  latest: sha256:aea63042abdf9a81319d6e9ffc6a381a6af53ae81a849c59d6269c3f233028a3
 aliases:
   mongo: /usr/bin/mongo
   mongod: /usr/bin/mongod
