+++ registry/ghcr.io/autamus/adios2/container.yaml
@@ -1,11 +1,12 @@
 docker: ghcr.io/autamus/adios2
 url: https://github.com/orgs/autamus/packages/container/package/adios2
 maintainer: '@vsoch'
-description: The Adaptable Input Output System version 2, developed in the Exascale Computing Program 
+description: The Adaptable Input Output System version 2, developed in the Exascale
+  Computing Program
 latest:
-  "2.7.1": sha256:38eb228d479fe61d558e8cbe6d5297ebf2950aa7aa146cb7fed9861c412f7263
+  2.7.1: sha256:f74621b5c0b4451093f720e69e29c209ebdfe75e167512f66c3a42b1cb2a0a49
 tags:
-  "2.7.1": sha256:38eb228d479fe61d558e8cbe6d5297ebf2950aa7aa146cb7fed9861c412f7263
+  2.7.1: sha256:f74621b5c0b4451093f720e69e29c209ebdfe75e167512f66c3a42b1cb2a0a49
 aliases:
   adios2-config: /opt/view/bin/adios2-config
   adios2_deactivate_bp: /opt/view/bin/adios2_deactivate_bp
