+++ registry/ghcr.io/autamus/slepc/container.yaml
@@ -3,6 +3,6 @@ url: https://github.com/orgs/autamus/packages/container/package/slepc
 maintainer: '@vsoch'
 description: Scalable Library for Eigenvalue Problem Computations.
 latest:
-  "3.15.1": sha256:80351f0d9f487abf208ba81ff4990e654d0a446e2d74fa8105449996092583cf
+  3.15.1: sha256:0b73bb63f73eb552682c3aa49ae07983d35b98c170b1ba76b46301bce4e4cc58
 tags:
-  "3.15.1": sha256:80351f0d9f487abf208ba81ff4990e654d0a446e2d74fa8105449996092583cf
+  3.15.1: sha256:0b73bb63f73eb552682c3aa49ae07983d35b98c170b1ba76b46301bce4e4cc58
