+++ registry/quay.io/biocontainers/star/container.yaml
@@ -3,10 +3,11 @@ url: https://quay.io/repository/biocontainers/star
 maintainer: '@audreystott'
 description: Spliced Transcripts Alignment to a Reference.
 latest:
-  2.7.9a--h9ee0642_0: sha256:d80ae2e0354e890d55ad0009e9ad9667d1113a679659071e6e02d50e46c8bba1
+  2.7.9a: sha256:d80ae2e0354e890d55ad0009e9ad9667d1113a679659071e6e02d50e46c8bba1
+filter:
+ - "^((?!--).)*$"
 tags:
   2.7.8a: sha256:b30f079606cb90aa8c4c8ebf96f4e7988e8e4c475bb842d75a76311b3d39cacd
   2.7.9a: sha256:d80ae2e0354e890d55ad0009e9ad9667d1113a679659071e6e02d50e46c8bba1
-  2.7.9a--h9ee0642_0: sha256:d80ae2e0354e890d55ad0009e9ad9667d1113a679659071e6e02d50e46c8bba1
 aliases:
   STAR: /usr/local/bin/STAR
