+++ registry/quay.io/biocontainers/trinity/container.yaml
@@ -5,11 +5,12 @@ description: Trinity, developed at the Broad Institute and the Hebrew University
   Jerusalem, represents a novel method for the efficient and robust de novo reconstruction
   of transcriptomes from RNA-seq data.
 latest:
-  2.12.0--h5ef6573_0: sha256:4045560dad56fee946c9bcccd2ba0e373292a5adf8a2f7ba09917cbaeaa74323
+  2.12.0: sha256:2736617a96c30bce93a149fe5f47b4f4a290a7d21c1503619430db619d9bc3a0
+filter:
+ - "^((?!--).)*$"
 tags:
   2.11.0: sha256:6d35d716aa12ba7b0c715fa8a30359b43ba9151e854dc4407b949e7b57c3a50a
   2.12.0: sha256:2736617a96c30bce93a149fe5f47b4f4a290a7d21c1503619430db619d9bc3a0
-  2.12.0--h5ef6573_0: sha256:4045560dad56fee946c9bcccd2ba0e373292a5adf8a2f7ba09917cbaeaa74323
 aliases:
   Trinity: /usr/local/bin/Trinity
   trinity: /usr/local/bin/Trinity
