+++ registry/quay.io/biocontainers/vcftools/container.yaml
@@ -3,11 +3,12 @@ url: https://quay.io/repository/biocontainers/vcftools
 maintainer: '@audreystott'
 description: A set of tools written in Perl and C++ for working with VCF files, such
   as those generated by the 1000 Genomes Project.
+filter:
+ - "^((?!--).)*$"
 latest:
-  0.1.16--he941832_2: sha256:55843b6aca7bb4fb78adee214755ce6289b34bdb49c128d1279225d8374ba420
+  0.1.16: sha256:91fac9c61535846675356d0d79abb9b9f8ecd0fb8b46411feaa891e652c8d51e
 tags:
   0.1.15: sha256:8af7cefca50bfdcf184f8a4faff336ae436e04f7a16dcd9b18fa0687b0ae975e
   0.1.16: sha256:91fac9c61535846675356d0d79abb9b9f8ecd0fb8b46411feaa891e652c8d51e
-  0.1.16--he941832_2: sha256:55843b6aca7bb4fb78adee214755ce6289b34bdb49c128d1279225d8374ba420
 aliases:
   vcftools: /usr/local/bin/vcftools
