+++ registry/quay.io/biocontainers/velvet/container.yaml
@@ -2,11 +2,12 @@ docker: quay.io/biocontainers/velvet
 url: https://quay.io/repository/biocontainers/velvet
 maintainer: '@audreystott'
 description: Short read de novo assembler using de Bruijn graphs
+filter:
+ - "^((?!--).)*$"
 latest:
-  1.2.10--h470a237_2: sha256:eb63751dc77dd0d3fbf2beb87eb48dc4d939c4b426676016f92dd4b202b87176
+  1.2.10: sha256:7fc2606a1431883dcd0acf830abcfeddb975677733d110a085da0f07782f5a27
 tags:
   1.2.10: sha256:7fc2606a1431883dcd0acf830abcfeddb975677733d110a085da0f07782f5a27
-  1.2.10--h470a237_2: sha256:eb63751dc77dd0d3fbf2beb87eb48dc4d939c4b426676016f92dd4b202b87176
 aliases:
   velvetg: /usr/local/bin/velvetg
   velveth: /usr/local/bin/velveth
