### 2021-06-27 4c01a65
[bot] binoc/update-2021-06-27 (#414)
* Automated deployment to update containers 2021-06-27
* Update biocontainers to not match tags with --, which are tag names combined with a commit (redundant, we just want the release tags)
Co-authored-by: github-actions <github-actions@users.noreply.github.com>
Co-authored-by: Vanessasaurus <814322+vsoch@users.noreply.github.com>
