+++ registry/rocker/tidyverse/container.yaml
@@ -3,9 +3,9 @@ url: https://hub.docker.com/r/rocker/tidyverse
 maintainer: '@vsoch'
 description: 'Version-stable build of R, rstudio, and R packages '
 latest:
-  latest: sha256:33a0bc1f76aa8db226970c461f34536c4169a7e731477ea9eb9b2b2b78a989a3
+  latest: sha256:37fd15f9e9795a700b3aa412cde7c7da7f89096a13dfc2323f6050cf834113db
 tags:
-  latest: sha256:33a0bc1f76aa8db226970c461f34536c4169a7e731477ea9eb9b2b2b78a989a3
+  latest: sha256:37fd15f9e9795a700b3aa412cde7c7da7f89096a13dfc2323f6050cf834113db
 filter:
 - latest
 aliases:
