+++ registry/ghcr.io/autamus/clhep/container.yaml
@@ -5,10 +5,10 @@ description: CLHEP is a C++ library that provides utility classes for general nu
   programming, vector arithmetic, geometry, pseudorandom number generation, and linear
   algebra, specifically targeted for high energy physics simulation and analysis software.
 latest:
-  2.4.4.0: sha256:8f5874189273e85511b0b26bd3734b0372764afb78f6171927081508f6fbfc26
+  2.4.4.0: sha256:6649a0d1b46a713ea8caa42df5390d186952229bfa8738cf7f3b01b615f060ce
 tags:
-  2.4.4.0: sha256:8f5874189273e85511b0b26bd3734b0372764afb78f6171927081508f6fbfc26
-  latest: sha256:8f5874189273e85511b0b26bd3734b0372764afb78f6171927081508f6fbfc26
+  2.4.4.0: sha256:6649a0d1b46a713ea8caa42df5390d186952229bfa8738cf7f3b01b615f060ce
+  latest: sha256:6649a0d1b46a713ea8caa42df5390d186952229bfa8738cf7f3b01b615f060ce
 aliases:
   Cast-config: /opt/view/bin/Cast-config
   Evaluator-config: /opt/view/bin/Evaluator-config
