+++ _library/openjdk.md
@@ -3,7 +3,7 @@ layout: container
 name:  "openjdk"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/openjdk/container.yaml"
-updated_at: "2021-06-23 03:36:26.205045"
+updated_at: "2021-06-27 03:46:59.085534"
 container_url: "https://hub.docker.com/_/openjdk"
 aliases:
  - "jar"
