+++ _library/gradle.md
@@ -3,7 +3,7 @@ layout: container
 name:  "gradle"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/gradle/container.yaml"
-updated_at: "2021-06-23 03:36:25.130330"
+updated_at: "2021-06-27 03:46:58.011735"
 container_url: "https://hub.docker.com/_/gradle"
 aliases:
  - "gradle"
