+++ _library/ghcr.io-autamus-bbmap.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bbmap"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bbmap/container.yaml"
-updated_at: "2021-06-23 03:35:53.752756"
+updated_at: "2021-06-27 03:46:26.634566"
 container_url: "https://github.com/orgs/autamus/packages/container/package/bbmap"
 
 versions:
