### 2021-06-28 singularityhub/singularity-hpc#415 @zbeekman
JOSS paper state of the field & references
This issue is part of the JOSS paper review over at https://github.com/openjournals/joss-reviews/issues/3311

Great paper, and great tool/ecosystem! I need to double check that you haven't added further content to the paper since I last read it, and in general I think it's great! Nice work!

Regarding state of the field/prior art and references: I'm wondering if it would make sense to reference both the [e4s](https://e4s-project.github.io/) project and [Pantheon Science](http://pantheonscience.org/) as they are both projects that have some similarities to singularity-hpc and ties to DOE labs as well. Pantheon Science is focused on reproducibility and E4S makes use of containers, Spack, and (at least theoretically) well defined software environments. In particular, I think highlighting some of the differences and similarities could be illustrative. I also wonder if there might be an opportunity for a future collaboration.  I have yet to completely dig my teeth into the singularity-hpc documentation and functionality verification process, but one advantage of singularity-hpc vs, e.g., Pantheon Science, is that the target audience might be smaller research groups & projects instead of very large and complex research projects and groups.

In the interest of full transparency: To me this isn't a blocker issue necessarily, and I don't think it would be fair for me to require you to cite/discuss e4s etc. since I do have some nascent connections and vested interest in that project through my employer and University of Oregon colleagues. However, I do think a brief discussion of the context in which a researcher might embrace one eco system (e.g. singularity-hpc vs e4s and optionally PantheonScience) over the other has the potential to add value to the paper. (Or some other comparison if you don't see them as direct competitors--I don't think they are necessarily.) And if you don't think adding citation(s) & a comparison to the paper is worthwhile, a very brief explanation of why you feel that way would be sufficient.

Once again, the JOSS article is looking quite good and a very interesting read about a very interesting project. I may have spotted one or two typos or awkward wording/sentence construction issues in the last draft I read--I'll double check that it's the most recent and highlight them/provide suggestions in a separate issue or PR whatever your preference is. (PR might be clearer, you can make the edits manually and reject the PR etc. but I'm open to whatever is easiest for you.) If for some reason you feel that the latest draft addresses these comments, and I simply haven't read it yet, feel free to say so! I'm going to double check first thing in the morning that I'm up to date with the latest version of the paper.
