### 2021-06-28 #3311 @zbeekman
OK, looking at the dates on this thread, I think my comments ☝️ are still up to date, and that I'm working from the latest draft. The issue I opened isn't anything major, in general I think it's a great paper and a great project, but I'm still going through the documentation and verifying functionality. The issues I opened isn't a blocker or dealbreaker, but I'd like to hear what you have to say in response before checking off the remaining checkboxes for the paper/manuscript portion of the review.

Once again great work, I've really enjoyed reading the paper and documentation so far and am excited to continue my test drive and functionality claims verification!
