### 2021-06-28 singularityhub/singularity-hpc#415 @vsoch
Thank you @zbeekman ! Here are some quick points for discussion.

> Regarding state of the field/prior art and references: I'm wondering if it would make sense to reference both the e4s project and Pantheon Science as they are both projects that have some similarities to singularity-hpc and ties to DOE labs as well.

Just one quick clarification - singularity-hpc has no tie to a DOE lab. I am employed by a national lab but this is a personal project I've been working on in my spare time, inspired and helped by random people reaching out to me that have nothing to do with my employment.

There is an interesting (and indirect) tie between shpc and e4s, and I can share and we can discuss if it's meaningful  to add to the paper or not (just as a side note, I had to cut out a lot of content from the paper off the bat because the editor said the paper was too long so we'd want to double check). So most of the e4s stack -> https://e4s-project.github.io/Resources/ProductInfo.html comes from spack. spack can build containers with containerize, but I believe the stack they are referencing here are just the binaries from the build cache. https://oaciss.uoregon.edu/e4s/inventory.html. The interesting connection is that Singularity Registry HPC uses a lot of containers from autamus -> https://autamus.io, and in fact I think most (if not all of) the e4s packages are provided there, in container form. Since all of autamus is in shpc, this means that shpc is (not intentionally) serving the same stack. So that's the unintended relation. But in terms of similarity, e4s itself is more of a large scale, collaborative project, and shpc is just a small piece of software (that could even be part of the project). So I think it would feel strange to list it as prior art because (other than providing software) I don't see them as comparable things. The comparable things might be other tools (like LMOD for example) that are tools themselves like shpc. I'm not familiar with pantheon, but akin to e4s it seems to be more of a project with a goal to provide an umbrella of projects. And it's explicitly for workflows, which again is unlike shpc.

I had included more detail on spack (and the differences) in the original draft but had to remove it because of length, but I was able to keep the reference to spack at the beginning of the paper. In short, spack is indeed comparable to shpc because it's a means to install and then provide modules (I don't do the comparison as a package manager because shpc is not a package manager) but both do provide modules! The main difference is with respect to the container detail - with spack you still need to install everything from source, and with shpc you just pull a container (much faster, but not ideal for everyone).

>  I also wonder if there might be an opportunity for a future collaboration. I have yet to completely dig my teeth into the singularity-hpc documentation and functionality verification process, but one advantage of singularity-hpc vs, e.g., Pantheon Science, is that the target audience might be smaller research groups & projects instead of very large and complex research projects and groups.

Oh collaboration is always wonderful! To be honest I haven't given much thought to how shpc fits into these larger, collaborative projects. For its inception and early use cases it's been helping single users (e.g., "I want to run rust quickly on my cluster" and HPC admins "I want to provide a bunch of containers as software to my users."). I think what you are getting at is could this provide more of a framework that would ensure something like a reproducible workflow, and the answer is that I don't know.

> (e.g. singularity-hpc vs e4s and optionally PantheonScience) over the other has the potential to add value to the paper.

I think I don't think these are mutually exclusive? I might benefit from either e4s or PantheonScience and still have use cases for shpc. I think the assumption here is that there must be one underlying "project or tool to rule them all" and I wouldn't want to place that burden on any particular piece of software (although maybe e4s/Pantheon don't mind taking on that huge challenge!) I think since you are experienced with these projects, if you want me to add a few sentences to the paper (which I'm happy to do) I'd need to understand why they are mutually exclusive, or viewed as competitors? To reiterate what i said earlier, I don't think I consider a little tool like SHPC to be on the same level, or to have the same aspirations as some of these huge, multi-institutional, and multi-software projects!

> Once again, the JOSS article is looking quite good and a very interesting read about a very interesting project. I may have spotted one or two typos or awkward wording/sentence construction issues in the last draft I read--I'll double check that it's the most recent and highlight them/provide suggestions in a separate issue or PR whatever your preference is. (PR might be clearer, you can make the edits manually and reject the PR etc. but I'm open to whatever is easiest for you.)

Personally, I'd like to give credit to you for your fixes, but if it's easier for you to just type them up somewhere (or otherwise not do a PR) I am good with that too!

So that's quite a bit of thinking about the above - let me know your thoughts! I can definitely add clarification / reference to the paper and I want to make sure we discuss what makes sense to add first.
