### 2021-06-29 8545fdd
adding more verbository about installing module software
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
