+++ _library/ghcr.io-autamus-slate.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/slate"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/slate/container.yaml"
-updated_at: "2021-06-27 03:46:49.729370"
+updated_at: "2021-06-29 14:59:29.723086"
 container_url: "https://github.com/orgs/autamus/packages/container/package/slate"
 
 versions:
