+++ _library/biocontainers-talon.md
@@ -3,7 +3,7 @@ layout: container
 name:  "biocontainers/talon"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/biocontainers/talon/container.yaml"
-updated_at: "2021-06-27 03:45:57.016990"
+updated_at: "2021-06-29 14:59:41.085432"
 container_url: "https://hub.docker.com/r/biocontainers/talon"
 aliases:
  - "talon"
