+++ _library/ghcr.io-autamus-mpc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/mpc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/mpc/container.yaml"
-updated_at: "2021-06-27 03:46:53.454468"
+updated_at: "2021-06-29 14:59:30.915576"
 container_url: "https://github.com/orgs/autamus/packages/container/package/mpc"
 
 versions:
