+++ _library/golang.md
@@ -3,7 +3,7 @@ layout: container
 name:  "golang"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/golang/container.yaml"
-updated_at: "2021-06-27 03:46:02.002736"
+updated_at: "2021-06-29 14:59:50.131458"
 container_url: "https://hub.docker.com/r/_/golang"
 aliases:
  - "go"
