### 2021-07-02 60bb75b
Adding new registry recipes
 On branch registryupdate
 Changes to be committed:
	new file:   canu.yml
	new file:   clustalo.yml
	new file:   cutadapt.yml
	new file:   diamond.yml
	new file:   fastqc.yml
	new file:   gatk4.yml
	new file:   maker.yml
	new file:   mrbayes.yml
	new file:   mummer.yml
