+++ fastqc.yml
@@ -1,0 +1,12 @@
+docker: quay.io/biocontainers/fastqc
+latest:
+    "0.11.9--hdfd78af_1":
+        "sha256:0c60406af11b0723339df05b10a592aa3f8c9a4d2ec8f213cbe11051e7264a25"
+tags:
+    "0.11.7--hdfd78af_7":
+        "sha256:cd90a6413e43865cacfa9e2217169d0acd38f3f6d6bee621cc7eac6f68cae823"
+maintainer: "@sarahbeecroft"
+description: "A quality control tool for high throughput sequence data"
+url: https://quay.io/repository/biocontainers/fastqc
+aliases:
+    fastqc: /usr/local/bin/fastqc
