+++ registry/ghcr.io/autamus/clhep/container.yaml
@@ -5,10 +5,10 @@ description: CLHEP is a C++ library that provides utility classes for general nu
   programming, vector arithmetic, geometry, pseudorandom number generation, and linear
   algebra, specifically targeted for high energy physics simulation and analysis software.
 latest:
-  2.4.4.0: sha256:6649a0d1b46a713ea8caa42df5390d186952229bfa8738cf7f3b01b615f060ce
+  2.4.4.0: sha256:87252d126a499e96349b97451a50ad2c1eaef1423e6dded074276542611006a8
 tags:
-  2.4.4.0: sha256:6649a0d1b46a713ea8caa42df5390d186952229bfa8738cf7f3b01b615f060ce
-  latest: sha256:6649a0d1b46a713ea8caa42df5390d186952229bfa8738cf7f3b01b615f060ce
+  2.4.4.0: sha256:87252d126a499e96349b97451a50ad2c1eaef1423e6dded074276542611006a8
+  latest: sha256:87252d126a499e96349b97451a50ad2c1eaef1423e6dded074276542611006a8
 aliases:
   Cast-config: /opt/view/bin/Cast-config
   Evaluator-config: /opt/view/bin/Evaluator-config

[…]
