### 2021-07-05 d123090
Update and rename quay.io/biocontainers/clustalo.yaml to registry/quay.io/biocontainers/clustalo/container.yaml
