### 2021-07-05 991a215
Update registry/quay.io/biocontainers/mummer/container.yaml
Co-authored-by: Vanessasaurus <814322+vsoch@users.noreply.github.com>
