### 2021-07-05 66d7a72
Adding new registry recipes (#417)
* Adding new registry recipes

 On branch registryupdate
 Changes to be committed:
	new file:   canu.yml
	new file:   clustalo.yml
	new file:   cutadapt.yml
	new file:   diamond.yml
	new file:   fastqc.yml
	new file:   gatk4.yml
	new file:   maker.yml
	new file:   mrbayes.yml
	new file:   mummer.yml

* Rename canu.yml to quay.io/biocontainers/can/container.yaml

* Rename clustalo.yml to quay.io/biocontainers/clustalo.yaml

* Rename quay.io/biocontainers/can/container.yaml to registry/quay.io/biocontainers/canu/container.yaml

* Update and rename quay.io/bioco
[…]
