+++ registry/quay.io/biocontainers/maker/container.yaml
@@ -1,0 +1,12 @@
+docker: quay.io/biocontainers/maker
+latest:
+    "3.01.03--pl5262h8f1cd36_1":
+        "sha256:1a588e177ae87eee1bf54edebe245b41ef2fbd7f0261762398c3b62883bbcd60"
+tags:
+    "2.31.11--pl526h61907ee_0":
+        "sha256:b9bd2979ddc5f74effd67c03d56a7f4f2640b36e242c54e2337ade9c5b6dca98"
+description: "A portable and easily configurable genome annotation pipeline"
+maintainer: "@sarahbeecroft"
+url: https://quay.io/repository/biocontainers/maker
+aliases:
+    maker: /usr/local/bin/maker
