+++ _library/quay.io-biocontainers-mrbayes.md
@@ -1,0 +1,132 @@
+---
+layout: container
+name:  "quay.io/biocontainers/mrbayes"
+maintainer: "@vsoch"
+github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/quay.io/biocontainers/mrbayes/container.yaml"
+updated_at: "2021-07-05 07:27:19.637007"
+container_url: "https://quay.io/repository/biocontainers/mrbayes"
+aliases:
+ - "mb"
+
+versions:
+ - "3.2.6--0"
+ - "3.2.7--h19cf415_2"
+description: "A program for Bayesian inference and model choice across a wide range of phylogenetic and evolutionary models"
+---
+
+This module is a singularity container wrapper for quay.io/biocontainers/mrbayes.
+A program for Bayesian inference and model choice across a wide range of phylogenetic and evolutionary models
+After [installing shpc](#install) you will want to install this container module:
+
+```bash
+$ shpc install quay.io/biocontainers/mrbayes
+```
+
+Or a specific version:
+
+```bash
+$ shpc install quay.io/biocontainers/mrbayes:3.2.6--0
+```
+
+And then you can tell lmod about your modules folder:
+
+```bash
+$ module use ./modules
+```
+
+And load the module, and ask for help, or similar.
+
+```bash
+$ module load quay.io/biocontainers/mrbayes/3.2.6--0
+$ module help quay.io/biocontainers/mrbayes/3.2.6--0
+```
+
+You can use tab for auto-completion of module names or commands that are provided.
+
+<br>
+
+### Commands
+
+When you install this module, you'll be able to load it to make the following commands accessible.
+Examples for both Singularity, Podman, and Docker (container technologies supported) are included.
+
+#### -run:
+
+```bash
+$ singularity run <container>
+$ podman run --rm  -v ${PWD} -w ${PWD} <container>
+$ docker run --rm  -v ${PWD} -w ${PWD} <container>
+```
+
+#### -shell:
+
+```bash
+$ singularity shell -s /bin/sh <container>
+$ podman run --it --rm --entrypoint /bin/sh  -v ${PWD} -w ${PWD} <container>
+$ docker run --it --rm --entrypoint /bin/sh  -v ${PWD} -w ${PWD} <container>
+```
+
+#### -exec:
+
+```bash
+$ singularity exec -s /bin/sh <container> "$@"
+$ podman run --it --rm --entrypoint ""  -v ${PWD} -w ${PWD} <container> "$@"
+$ docker run --it --rm --entrypoint ""  -v ${PWD} -w ${PWD} <container> "$@"
+```
+
+#### -inspect:
+
+Podman and Docker only have one inspect type.
+
+```bash
+$ podman inspect <container>
+$ docker inspect <container>
+```
+
+#### -inspect-runscript:
+
+```bash
+$ singularity inspect -r <container>
+```
+
+#### -inspect-deffile:
+
+```bash
+$ singularity inspect -d <container>
+```
+
+
+#### mb
+       
+```bash
+$ singularity exec <container> /usr/local/bin/mb
+$ podman run --it --rm --entrypoint /usr/local/bin/mb   -v ${PWD} -w ${PWD} <container> -c " $@"
+$ docker run --it --rm --entrypoint /usr/local/bin/mb   -v ${PWD} -w ${PWD} <container> -c " $@"
+```
+
+
+
+In the above, the `<container>` directive will reference an actual container provided
+by the module, for the version you have chosen to load. An environment file in the
+module folder will also be bound. Note that although a container
+might provide custom commands, every container exposes unique exec, shell, run, and
+inspect aliases. For anycommands above, you can export:
+
[…]
