+++ _library/bids-broccoli.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/broccoli"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/broccoli/container.yaml"
-updated_at: "2021-06-29 14:59:46.887351"
+updated_at: "2021-07-05 07:28:09.558779"
 container_url: "https://hub.docker.com/r/bids/broccoli"
 
 versions:
