+++ registry/quay.io/biocontainers/diamond/container.yaml
@@ -1,12 +1,11 @@
 docker: quay.io/biocontainers/diamond
-latest:
-    "2.0.9--hdcc8f71_0":
-        "sha256:3e6ab3300dbbad7d86a0827813cc840d5454e762f10c325c3e28b61b0c2bfc72"
-tags:
-    "0.8.26--h2e03b76_4":
-        "sha256:1bdf8de97dfdfb795a14cecf22cc7b249801ac9886e187533c3b1cf3c9cb7613"
-maintainer: "@sarahbeecroft"
-description: "Accelerated blast compatible local sequence aligner"
 url: https://quay.io/repository/biocontainers/diamond
+maintainer: '@sarahbeecroft'
+description: Accelerated blast compatible local sequence aligner
+latest:
+  2.0.11--hdcc8f71_0: sha256:cb3a52f7f7d745e3c023b40faebd5099afceddb290bd65ebd892ddabef41463a
+tags:
+  0.8.26--h2e03b76_4: sha256:1bdf8de97dfdfb795a14cecf22cc7b249801ac9886e187533c3b1cf3c9cb7613
+  2.0.11--hdcc8f71_0: sha256:cb3a52f7f7d745e3c023b40faebd5099afceddb290bd65ebd892ddabef41463a
 aliases:
-    diamond: /usr/local/bin/diamond
+  diamond: /usr/local/bin/diamond
