+++ registry/ghcr.io/autamus/py-libensemble/container.yaml
@@ -3,6 +3,6 @@ url: https://github.com/orgs/autamus/packages/container/package/py-libensemble
 maintainer: '@vsoch'
 description: Library for managing ensemble-like collections of computations.
 latest:
-  0.7.2: sha256:410bb2de1f0990e15b2761964068574ed9731972b4a2a13b5fa683dafc6880ea
+  0.7.2: sha256:655fc3e1a63186074b11d3e1d46bd2728ad550805f867523dd7c5f8f23076c4e
 tags:
-  0.7.2: sha256:410bb2de1f0990e15b2761964068574ed9731972b4a2a13b5fa683dafc6880ea
+  0.7.2: sha256:655fc3e1a63186074b11d3e1d46bd2728ad550805f867523dd7c5f8f23076c4e
