+++ registry/ghcr.io/autamus/py-petsc4py/container.yaml
@@ -3,6 +3,6 @@ url: https://github.com/orgs/autamus/packages/container/package/py-petsc4py
 maintainer: '@vsoch'
 description: This package provides Python bindings for the PETSc package.
 latest:
-  3.15.1: sha256:679bcb8e6455bdd11bca93c8813d33cbcc9cb7ab6f34d2436c65ba1062351f41
+  3.15.1: sha256:678b37e3805ea09bc9c0f591f787b67f5e38024dffa745c37a6db047ae4f4cd4
 tags:
-  3.15.1: sha256:679bcb8e6455bdd11bca93c8813d33cbcc9cb7ab6f34d2436c65ba1062351f41
+  3.15.1: sha256:678b37e3805ea09bc9c0f591f787b67f5e38024dffa745c37a6db047ae4f4cd4
