+++ registry/ghcr.io/autamus/octave/container.yaml
@@ -4,10 +4,11 @@ maintainer: '@vsoch'
 description: GNU Octave is a high-level interpreted language, primarily intended for
   numerical computations (like an open source Matlab)
 latest:
-  6.2.0: sha256:455604046653c2d7066b58adf40d1211151a30b15d12f0e73f5b47de3cfe0433
+  6.3.0: sha256:72137fca3a63a012f377273c04d284c9008dc0c99231c9eb1a3b2c5d9686cc1f
 tags:
-  6.2.0: sha256:455604046653c2d7066b58adf40d1211151a30b15d12f0e73f5b47de3cfe0433
-  latest: sha256:455604046653c2d7066b58adf40d1211151a30b15d12f0e73f5b47de3cfe0433
+  6.2.0: sha256:26a7d5c22a201b4acaa3e4c4d7d9c1aac0009fae4ea71cd73282762b46f39184
+  6.3.0: sha256:72137fca3a63a012f377273c04d284c9008dc0c99231c9eb1a3b2c5d9686cc1f
+  latest: sha256:72137fca3a63a012f377273c04d284c9008dc0c99231c9eb1a3b2c5d9686cc1f
 aliases:
   octave: /opt/view/bin/octave
   octave-6.2.0: /opt/view/bin/octave-6.2.0
