+++ registry/ghcr.io/autamus/turbine/container.yaml
@@ -3,9 +3,9 @@ url: https://github.com/orgs/autamus/packages/container/package/turbine
 maintainer: '@vsoch'
 description: Turbine is the Swift/T runtime
 latest:
-  1.3.0: sha256:fee6911aeb014b2343437788f20128368712d800dc3dce33373b19000c11b033
+  1.3.0: sha256:d4b512191d96e9b37c3a65eaf42c8f5ef8d4bac28b48878fa2a2326c12f27ac8
 tags:
-  1.3.0: sha256:fee6911aeb014b2343437788f20128368712d800dc3dce33373b19000c11b033
+  1.3.0: sha256:d4b512191d96e9b37c3a65eaf42c8f5ef8d4bac28b48878fa2a2326c12f27ac8
 aliases:
   turbine: /opt/view/bin/turbine
   turbine-pilot: /opt/view/bin/turbine-pilot
