+++ registry/ghcr.io/autamus/clingo/container.yaml
@@ -3,11 +3,11 @@ url: https://github.com/orgs/autamus/packages/container/package/clingo
 maintainer: '@vsoch'
 description: An ASP system to ground and solve logic programs.
 latest:
-  5.5.0: sha256:4d67898635786430cf4c7db8c2b319801cc0126eab8977ff874dd9a0c3ea6215
+  5.5.0: sha256:75bf5cb51fffa4c62ed166379d91f429b8bb5b1e98442238e7e56ee5eec097b6
 tags:
   5.4.1: sha256:c98a480f85a68063dfb9c4cb6eedfbe0cf84a1af737b73ef2d94213fffd91ff5
-  5.5.0: sha256:4d67898635786430cf4c7db8c2b319801cc0126eab8977ff874dd9a0c3ea6215
-  latest: sha256:4d67898635786430cf4c7db8c2b319801cc0126eab8977ff874dd9a0c3ea6215
+  5.5.0: sha256:75bf5cb51fffa4c62ed166379d91f429b8bb5b1e98442238e7e56ee5eec097b6
+  latest: sha256:75bf5cb51fffa4c62ed166379d91f429b8bb5b1e98442238e7e56ee5eec097b6
 aliases:
   clingo: /opt/view/bin/clingo
   gringo: /opt/view/bin/gringo
