+++ registry/ghcr.io/autamus/caliper/container.yaml
@@ -3,9 +3,9 @@ url: https://github.com/orgs/autamus/packages/container/package/caliper
 maintainer: '@vsoch'
 description: Caliper is a program instrumentation and performance measurement framework.
 latest:
-  2.6.0: sha256:066ff2868b6824a4ba3e505842b0b1986bfd2c625f813ee9655625ddef6cccfb
+  2.6.0: sha256:f291b44f6b323191c88a54985b14131f4bf6f6ca5d11ede92e872080c06720e2
 tags:
-  2.6.0: sha256:066ff2868b6824a4ba3e505842b0b1986bfd2c625f813ee9655625ddef6cccfb
+  2.6.0: sha256:f291b44f6b323191c88a54985b14131f4bf6f6ca5d11ede92e872080c06720e2
 aliases:
   cali-query: /opt/view/bin/cali-query
   cali-stat: /opt/view/bin/cali-stat
