+++ registry/tomcat/container.yaml
@@ -4,9 +4,10 @@ maintainer: '@vsoch'
 description: Apache Tomcat is an open source implementation of the Java Servlet and
   JavaServer Pages technologies
 latest:
-  10.0.7-jdk11-adoptopenjdk-hotspot: sha256:d75c50123194e5533dd8b397175fab79c9dff69ed5c0bada70be5dff6d8fcc6d
+  10.0.8-jdk11-adoptopenjdk-hotspot: sha256:f6cdb1b05f9b3b1ff7bfd0982cd880ccb4be433409e740792cc37690ec801717
 tags:
   10.0.5-jdk11-adoptopenjdk-hotspot: sha256:a7418f29d3dd7ad20bcf052b3b6dc2777d118344286c6374bc447fb217a97c08
   10.0.6-jdk11-adoptopenjdk-hotspot: sha256:c0019c8254bc1017f64e6ffd1612e25b50abca78d98d25f9ff8023e8999f0384
   10.0.7-jdk11-adoptopenjdk-hotspot: sha256:d75c50123194e5533dd8b397175fab79c9dff69ed5c0bada70be5dff6d8fcc6d
-  latest: sha256:d0e769cea1684c065bb6a2e52f13cd5d7138a19a5ef3c16828f32e9bd8323f56
+  10.0.8-jdk11-adoptopenjdk-hotspot: sha256:f6cdb1b05f9b3b1ff7bfd0982cd880ccb4be433409e740792cc37690ec801717
+  latest: sha256:a5abf192aceed45620dbb2e09f8abdec2b96108b86365a988c85e753c28cd36b
