+++ registry/ghcr.io/autamus/tasmanian/container.yaml
@@ -5,8 +5,8 @@ description: The Toolkit for Adaptive Stochastic Modeling and Non-Intrusive Appr
   is a robust library for high dimensional integration and interpolation as well as
   parameter calibration.  https://tasmanian.ornl.gov
 latest:
-  "7.5": sha256:1aa5c31e66629c319e30eaaed94f5df3f689efcf023a2b54a9a530148f81c689
+  "7.5": sha256:1a282fc0d7b5d59e0ad9c8f7f2d1cceec75f2086515fe3a02968a49bb73aa054
 tags:
-  "7.5": sha256:1aa5c31e66629c319e30eaaed94f5df3f689efcf023a2b54a9a530148f81c689
+  "7.5": sha256:1a282fc0d7b5d59e0ad9c8f7f2d1cceec75f2086515fe3a02968a49bb73aa054
 aliases:
   tasgrid: /opt/view/bin/tasgrid
