+++ _library/ghcr.io-autamus-xnnpack.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/xnnpack"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/xnnpack/container.yaml"
-updated_at: "2021-07-18 00:53:35.084629"
+updated_at: "2021-07-18 00:55:42.711279"
 container_url: "https://github.com/orgs/autamus/packages/container/package/xnnpack"
 
 versions:
