+++ _library/ghcr.io-autamus-mpc.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/mpc"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/mpc/container.yaml"
-updated_at: "2021-07-18 00:53:26.005140"
+updated_at: "2021-07-18 00:55:31.255830"
 container_url: "https://github.com/orgs/autamus/packages/container/package/mpc"
 
 versions:
