### 2021-07-19 #3311 @ArangoGutierrez
> Sorry for holding this up, I didn’t intend to become the dreaded “reviewer 3” scream but work and life have placed some unexpected demands on my time these past few weeks. This is a great project and paper and I do want to get this wrapped up ASAP so it can get published. I need to re-read the paper in light of the discussion/comments on the project repository issue I opened—I may have one or two follow up questions but possibly not. I’m nearly done reviewing the documentation & functionality. The plan is to wrap up tomorrow (Monday) morning ahead of a work call I have.
>
> Thanks, and sorry once again for being “reviewer 3” and holding this up.

Hey sorry, I was on PTO and outside my country so was hard to get back to this, will prioritize this today
