### 2021-07-19 #3311 @ArangoGutierrez
> > > Hi @ArangoGutierrez how is your review going?
> >
> >
> > I went over my checklist and I think this qualifies for publication
>
> @ArangoGutierrez, I think you are good with your review. Just two boxes were not checked and I checked them for you, since I assumed you finished the review. If not, please uncheck them.

I am I left this https://github.com/openjournals/joss-reviews/issues/3311#issuecomment-876677672 , but then I think I simply forgot those 2 points, I think the doc has good value for the community
