### 2021-07-19 singularityhub/singularity-hpc#415 @zbeekman
> Just one quick clarification - singularity-hpc has no tie to a DOE lab. I am employed by a national lab but this is a personal project I've been working on in my spare time, inspired and helped by random people reaching out to me that have nothing to do with my employment.

My mistake, my comment was based solely off your affiliation.

> The interesting connection is that Singularity Registry HPC uses a lot of containers from autamus -> https://autamus.io, and in fact I think most (if not all of) the e4s packages are provided there, in container form. Since all of autamus is in shpc, this means that shpc is (not intentionally) serving the same stack. So that's the unintended relation. But in terms of similarity, e4s itself is more of a large scale, collaborative project, and shpc is just a small piece of software (that could even be part of the project). So I think it would feel strange to list it as prior art because (other than providing software) I don't see them as comparable things. The comparable things might be other tools (like LMOD for example) that are tools themselves like shpc. I'm not familiar with pantheon, but akin to e4s it seems to be more of a project with a goal to provide an umbrella of projects. And it's explicitly for workflows, which again is unlike shpc.

All good points!

> I had included more detail on spack (and the differences) in the original draft but had to remove it because of length, but I was able to keep the reference to spack at the beginning of the paper. In short, spack is indeed comparable to shpc because it's a means to install and then provide modules (I don't do the comparison as a package manager because shpc is not a package manager) but both do provide modules! The main difference is with respect to the container detail - with spack you still need to install everything from source, and with shpc you just pull a container (much faster, but not ideal for everyone).

Yes, it certainly is tough to squeeze in everything one wishes to discuss with the short page limit. I guess part of my ignorance or confusion is: how does one categorize tools like singularity-hpc and E4S? They're both providing curated software to improve reproducibility and ease developer build/install pain points with hashes and (optionally in the case of E4S) containerized software.

> For its inception and early use cases it's been helping single users (e.g., "I want to run rust quickly on my cluster" and HPC admins "I want to provide a bunch of containers as software to my users.").

It seems to me that this is one of the chief distinctions of singularity-hpc, and distinguishes it from other offerings, even though they do target single users too. I recall when Linuxbrew (now officially part of Homebrew) was a last resort that I would use out of desperation to build an entire software stack (including compilers, glibc, etc!) on HPC systems where I couldn't easily get an admin to install what I needed. (Spack/easy_build either didn't exist yet or I didn't know about them.) Singularity-hpc seems like a modern, polished and--at least relative to other similar offerings, which, in fact, do a good/decent job--path of least resistance. I.e. it excels at supporting the "lone researcher" or rapid iterations outside of a large team. I think E4S aspires to this as well, but their main focus might be more medium & large projects & industry groups. And as you point out singularity-hpc is comparatively small project--funding/team-wise at least. (However it's incredible what has been accomplished!)

> I think what you are getting at is could this provide more of a framework that would ensure something like a reproducible workflow, and the answer is that I don't know.

Well, for it to be like Pantheon yes, but I was mostly just trying to figure out in my mind when I (or when someone) would want to use singularity-hpc vs something like E4S. There is some overlap in the use cases but they are different tools with different strengths and weaknesses. As a relative newcomer to both my challenge is "which one is good for me?" or "which one is good for my project". I think the barrier to entry of contributing software to singularity-hpc seems lower, so perhaps that's one big strength. (This is a large reason Homebrew became so widely adopted on macOS.)

> > (e.g. singularity-hpc vs e4s and optionally PantheonScience) over the other has the potential to add value to the paper.
>
> I think I don't think these are mutually exclusive? I might benefit from either e4s or PantheonScience and still have use cases for shpc. I think the assumption here is that there must be one underlying "project or tool to rule them all" and I wouldn't want to place that burden on any particular piece of software (although maybe e4s/Pantheon don't mind taking on that huge challenge!) I think since you are experienced with these projects, if you want me to add a few sentences to the paper (which I'm happy to do) I'd need to understand why they are mutually exclusive, or viewed as competitors? To reiterate what i said earlier, I don't think I consider a little tool like SHPC to be on the same level, or to have the same aspirations as some of these huge, multi-institutional, and multi-software projects!

Yes, I absolutely 💯 percent agree with you that they're not mutually exclusive and that there shouldn't be a "one package (manager) or project to rule them all" mentality. I did not mean to imply any sort of competition or "why would I use singularity-hpc when E4S, or other tool X exists." I just found the degree of similarity of a number of aspects of E4S and singularity-hpc quite high, but the overlap is not 100%.

However, I guess there's no great way of guiding users (by providing more detailed use cases) or performing a comparison of multiple tools without it being construed of as competitive. (Unless of course you're authoring a review paper or meta analysis, etc.) So I guess I ultimately agree with your decision to write the paper the way you did. It doesn't completely give me all the information *I* want as a new user about when I should pick singularity-hpc over something else, but there's probably not a great diplomatic way to do that 1) without being an expert with the other tool(s) and 2) while being the author of one of the tools.

> So that's quite a bit of thinking about the above - let me know your thoughts! I can definitely add clarification / reference to the paper and I want to make sure we discuss what makes sense to add first.

After reading your comments and thinking about this some more I don't think you should change the paper unless you can find a diplomatic and relevant way of highlighting the differences with other psuedo-competitor projects. (And without becoming an expert in all the other projects!) I just can't think of good way to provide a list of use cases and say "in situation A use singularity-hpc, in situation B use singularity-hpc, in situation C E4S is more convenient," etc. that is accurate, unbiased, diplomatic, and doesn't take up a ton of space. This, to me, is the biggest unanswered question I have after reading the paper and knowing (a little) about E4S. If the software stack in both is nearly identical and suits my needs which one would be better for my hypothetical project? However, it might not be an easy question to answer, and it certainly would be hard to put a dent in with the JOSS page limit.

I will note that the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#an-important-note-about-novel-software-and-citations-of-relevant-work) state:

> __An important note about ‘novel’ software and citations of relevant work__
> Submissions that implement solutions already solved in other software packages are accepted into JOSS provided that they meet the criteria listed above and cite prior similar work. Reviewers should point out relevant published work which is not yet cited.

Whether or not E4S is a "solution already solved in other software packages" is a bit of a gray area to me, and I'm content whether or not you decide to cite it, however, I would be inclined to cite it if I were the author of your paper. (But this is not a blocking issue for me!)

Finally, my apologies if my prose turned out as word vomit--I don't want to hold up publication any longer, and I'm not the most natural writer/communicator. I'll close this unless you think it should remain open.
