### 2021-07-19 singularityhub/singularity-hpc#415 @vsoch
No worries! I appreciate the detailed feedback. I will have to think more about how E4s relates to shpc (I still don't think they are birds of a feather!) but probably won't be able to add to the paper because I don't have that understanding. I think the related / most relevant work is the other container modules library, and then environment modules, all of which are cited. E4s in my mind is something much bigger (which ironically given my current employment I haven't used yet!)

Thank you for the learnings / discussion @zbeekman !
