### 2021-07-19 de35ae7
removing lmod_base and other unused defaults
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
