### 2021-07-19 singularityhub/singularity-hpc#415 @zbeekman
> I still don't think they are birds of a feather!

I still wonder why you say this, but, admittedly my understanding of shpc and e4s is incomplete and imperfect. Obviously E4S is a much bigger project with much more funding, but a lot of shpc features appear to give it a run for its money, and it seems to me like E4S can do a lot of what shpc does, albeit perhaps not as easily. For my own curiosity and education, if you have the time and don't mind indulging me, a (short, and incomplete) list of "things" (capabilities, strengths, qualities) shpc has that E4S does not would help me understand the unique capabilities of shpc. But you have absolutely no obligation to spend your time doing this, JOSS review is finished 🎉
