### 2021-07-19 141499f
removing lmod_base and other unused defaults (#422)
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>
