+++ _library/adminer.md
@@ -3,7 +3,7 @@ layout: container
 name:  "adminer"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/adminer/container.yaml"
-updated_at: "2021-07-18 00:55:55.896050"
+updated_at: "2021-07-19 23:53:35.559154"
 container_url: "https://hub.docker.com/_/adminer"
 
 versions:
