+++ _library/ghcr.io-autamus-libnrm.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/libnrm"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/libnrm/container.yaml"
-updated_at: "2021-07-18 00:55:19.733289"
+updated_at: "2021-07-19 23:52:54.296206"
 container_url: "https://github.com/orgs/autamus/packages/container/package/libnrm"
 
 versions:
