### 2021-07-20 9d0c51d
Escape `$` for TCL modules (don't deref in TCL)
TCL modules will dereference (expand) things that look like bash
variables and "new-style" command substitutions.
