### 2021-07-20 78c9f2a
fixing bug that envars need to be escaped for tcl
Also updating recipes for jupyter containers to create a temporary directory to bind to /var/run

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
