### 2021-07-20 #3311 @diehlpk
@vsoch I recently was the editor for the following paper

https://joss.theoj.org/papers/10.21105/joss.02403

and I was wondering how your paper is different, and I was wondering if this paper should be cited?
