### 2021-07-20 #3311 @vsoch
@diehlpk all set!  https://github.com/singularityhub/singularity-hpc/commit/060fe7a684d0fc97cf8659dcc66f49776016b4ac

That work appears to be for a workflow manager (for cloud workflows) which is a totally different animal. [There are a to of them](https://github.com/pditommaso/awesome-pipeline) and shpc is not in that family. Probably the fact that "reproducible" has become a buzzword (and common in papers for computational tools) is the only overlap.
