+++ _library/ghcr.io-autamus-libunwind.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/libunwind"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/libunwind/container.yaml"
-updated_at: "2021-07-19 23:52:43.048010"
+updated_at: "2021-07-20 16:01:37.967834"
 container_url: "https://github.com/orgs/autamus/packages/container/package/libunwind"
 
 versions:
