+++ _library/quay.io-biocontainers-velvet.md
@@ -3,7 +3,7 @@ layout: container
 name:  "quay.io/biocontainers/velvet"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/quay.io/biocontainers/velvet/container.yaml"
-updated_at: "2021-07-19 23:52:29.294371"
+updated_at: "2021-07-20 16:01:28.107536"
 container_url: "https://quay.io/repository/biocontainers/velvet"
 aliases:
  - "velvetg"
