+++ _library/quay.io-biocontainers-mrbayes.md
@@ -3,7 +3,7 @@ layout: container
 name:  "quay.io/biocontainers/mrbayes"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/quay.io/biocontainers/mrbayes/container.yaml"
-updated_at: "2021-07-20 16:01:26.347643"
+updated_at: "2021-07-20 16:15:55.627422"
 container_url: "https://quay.io/repository/biocontainers/mrbayes"
 aliases:
  - "mb"
