+++ _library/ghcr.io-autamus-zlib.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/zlib"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/zlib/container.yaml"
-updated_at: "2021-07-20 16:16:23.473781"
+updated_at: "2021-07-20 16:19:34.268172"
 container_url: "https://github.com/orgs/autamus/packages/container/package/zlib"
 
 versions:
