+++ _library/bids-broccoli.md
@@ -3,7 +3,7 @@ layout: container
 name:  "bids/broccoli"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/bids/broccoli/container.yaml"
-updated_at: "2021-07-20 16:16:39.902537"
+updated_at: "2021-07-20 16:19:51.813516"
 container_url: "https://hub.docker.com/r/bids/broccoli"
 
 versions:
