### 2021-07-20 #3311 @diehlpk
@vsoch Not yet, the title of the above item should read as

Collaborative Container Modules with Singularity1
Registry HPC

and not read as

singularityhub/singularity-hpc: Singularity Registry HPC v0.0.31
