### 2021-07-20 #3311 @vsoch
It looks like we are getting here:

https://github.com/openjournals/whedon/blob/919a44c1cae3a44f4475db2cd3bbc17b3b3ff721/lib/whedon/bibtex_parser.rb#L82

And the value that we are trying to call encode on is nil.
