### 2021-07-20 #3311 @danielskatz
yes, looks like a problem in a DOI in a citation - is there a bibtex entry with a DOI = {} field?
