### 2021-07-20 #3311 @danielskatz
try changing
>month     =  "11~" # may,

to
>month     =  may,
