### 2021-07-20 #3311 @danielskatz
Also probably remove
>  abstract    = "Binoc is a Lookout Bot that updates containers in Autamus -
                 autamus/binoc",

This isn't causing the error but will confuse the XML that is submitted to crossref
