### 2021-07-20 #3311 @danielskatz
similarly, remove the abstract from Wandell2015-yt
