### 2021-07-20 20e2f30
changes in bib to try to fix current and future issues
