+++ paper/paper.bib
@@ -1,28 +1,13 @@
 @ARTICLE{Kurtzer2017-xj,
   title     = "Singularity: Scientific containers for mobility of compute",
   author    = "Kurtzer, Gregory M and Sochat, Vanessa and Bauer, Michael W",
-  abstract  = "Here we present Singularity, software developed to bring
-               containers and reproducibility to scientific computing. Using
-               Singularity containers, developers can work in reproducible
-               environments of their choosing and design, and these complete
-               environments can easily be copied and executed on other
-               platforms. Singularity is an open source initiative that
-               harnesses the expertise of system and software engineers and
-               researchers alike, and integrates seamlessly into common
-               workflows for both of these groups. As its primary use case,
-               Singularity brings mobility of computing to both users and HPC
-               centers, providing a secure means to capture and distribute
-               software and compute environments. This ability to create and
-               deploy reproducible environments across these centers, a
-               previously unmet need, makes Singularity a game changing
-               development for computational science.",
   journal   = "PLoS One",
   doi       = "journal.pone.0177459",
   publisher = "Public Library of Science",
   volume    =  12,
   number    =  5,
   pages     = "e0177459",
-  month     =  "11~" # may,
+  month     =  may,
   year      =  2017
 }
 
@@ -43,25 +28,6 @@
   booktitle = "State of the Practice Reports",
   author    = "McLay, Robert and Schulz, Karl W and Barth, William L and
                Minyard, Tommy",
-  abstract  = "Commodity-based Linux HPC clusters dominate the scientific
-               computing landscape in both academia and industry ranging from
-               small research clusters to petascale supercomputers supporting
-               thousands of users. To support broad user communities and manage
-               a user-friendly environment, end-user sites must combine a range
-               of low-level system software with multiple compiler chains,
-               support libraries, and a suite of 3rd party applications. In
-               addition, large systems require bare metal provisioning and a
-               flexible software management strategy to maintain consistency
-               and upgrade-ability across thousands of compute nodes. This
-               report documents a Linux operating system framework, (LosF),
-               which has evolved over the last seven years to provide an
-               integrated strategy for the deployment of multiple HPC systems
-               at the Texas Advanced Computing Center. Documented within this
-               effort is the high-level cluster configuration options and
-               definitions, bare-metal provisioning, hierarchical HPC software
-               stack design, package-management, user environment management
-               tools, user account synchronization, and local customization
-               configurations.",
   publisher = "Association for Computing Machinery",
   number    = "Article 9",
   pages     = "1--11",
@@ -95,7 +61,7 @@
   abstract    = "Build and deploy Singularity containers to GitHub releases,
                  and pull with the singularity-hpc client -
                  singularityhub/singularity-deploy",
-  institution = "Github",
+  institution = "GitHub",
   year      =  2021
 }
 
@@ -119,8 +85,6 @@
   title       = "binoc",
   author      = "Scott, Alec",
   howpublished = "\url{https://github.com/autamus/binoc}",
-  abstract    = "Binoc is a Lookout Bot that updates containers in Autamus -
-                 autamus/binoc",
   institution = "Github",
   year        = 2021
 }
@@ -153,9 +117,9 @@
 @incollection{cook2017opinionated,
   title={The Opinionated Jupyter Stacks},
   author={Cook, Joshua},
-  booktitle={Docker for Data Science},
+  booktitle={{Docker for Data Science}},
   pages={119--135},
-  year={2017},
+  year=2017,
   publisher={Springer},
   doi="10.1007/978-1-4842-3012-1_7"
 }
@@ -177,7 +141,7 @@
[…]
