### 2021-07-20 #3311 @danielskatz
ok, try merging https://github.com/singularityhub/singularity-hpc/pull/427 and rebuilding
