+++ _library/ghcr.io-autamus-dyninst.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/dyninst"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/dyninst/container.yaml"
-updated_at: "2021-07-20 16:19:37.446281"
+updated_at: "2021-07-20 19:27:15.606975"
 container_url: "https://github.com/orgs/autamus/packages/container/package/dyninst"
 
 versions:
