+++ _library/rocker-tidyverse.md
@@ -3,7 +3,7 @@ layout: container
 name:  "rocker/tidyverse"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/rocker/tidyverse/container.yaml"
-updated_at: "2021-07-20 16:18:58.204469"
+updated_at: "2021-07-20 19:26:35.766795"
 container_url: "https://hub.docker.com/r/rocker/tidyverse"
 aliases:
  - "R"
