+++ _library/ghcr.io-autamus-bbmap.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/bbmap"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/bbmap/container.yaml"
-updated_at: "2021-07-20 16:19:39.952628"
+updated_at: "2021-07-20 19:27:18.262475"
 container_url: "https://github.com/orgs/autamus/packages/container/package/bbmap"
 
 versions:
