### 2021-07-20 #3311 @vsoch
I'm close to being able to debug locally but I'm missing a style file?

```bash
$ bundle exec whedon compile 3311
/home/vanessa/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/gems/json-1.8.6/lib/json/common.rb:155: warning: Using the last argument as keyword parameters is deprecated
/home/vanessa/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/gems/json-1.8.6/lib/json/common.rb:155: warning: Using the last argument as keyword parameters is deprecated
/home/vanessa/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/gems/json-1.8.6/lib/json/common.rb:155: warning: Using the last argument as keyword parameters is deprecated
/home/vanessa/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/gems/json-1.8.6/lib/json/common.rb:155: warning: Using the last argument as keyword parameters is deprecated
Error producing PDF.
! LaTeX Error: File `biblatex.sty' not found.

Type X to quit or <RETURN> to proceed,
or enter new name. (Default extension: sty)

Enter file name:
! Emergency stop.
<read *>

l.36 \bibliography

Looks like we failed to compile the PDF
```
I'll see if I can find it anywhere.
