+++ paper/paper.bib
@@ -2,7 +2,7 @@
   title     = "Singularity: Scientific containers for mobility of compute",
   author    = "Kurtzer, Gregory M and Sochat, Vanessa and Bauer, Michael W",
   journal   = "PLoS One",
-  doi       = "journal.pone.0177459",
+  doi       = "10.1371/journal.pone.0177459",
   publisher = "Public Library of Science",
   volume    =  12,
   number    =  5,
