### 2021-07-20 #3311 @danielskatz
see https://github.com/singularityhub/singularity-hpc/pull/428 - the first DOI was wrong (incomplete)
