### 2021-07-20 #3311 @vsoch
I think what's happening is the DOI is not matching either of the regular expressions.
