### 2021-07-20 #3311 @danielskatz
right - see my PR for why
DOIs should be x/y
this was just y
