+++ _library/postgres.md
@@ -3,7 +3,7 @@ layout: container
 name:  "postgres"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/postgres/container.yaml"
-updated_at: "2021-07-20 19:27:33.341627"
+updated_at: "2021-07-20 19:37:37.616562"
 container_url: "https://hub.docker.com/r/_/postgres"
 aliases:
  - "clusterdb"
