### 2021-07-20 #3311 @danielskatz
see https://github.com/singularityhub/singularity-hpc/pull/429 (and note the two comments/questions in the PR) and https://github.com/singularityhub/singularity-hpc/pull/430
