### 2021-07-20 5625218
Update paper/paper.md
Co-authored-by: Vanessasaurus <814322+vsoch@users.noreply.github.com>
