+++ _library/ghcr.io-autamus-mercury.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/mercury"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/mercury/container.yaml"
-updated_at: "2021-07-20 19:37:02.560139"
+updated_at: "2021-07-20 19:57:11.673251"
 container_url: "https://github.com/orgs/autamus/packages/container/package/mercury"
 
 versions:
