### 2021-07-20 339177d
paper edits (#429)
* paper edits
* Update paper/paper.md

Co-authored-by: Vanessasaurus <814322+vsoch@users.noreply.github.com>
