+++ _library/ghcr.io-autamus-omega-h.md
@@ -3,7 +3,7 @@ layout: container
 name:  "ghcr.io/autamus/omega-h"
 maintainer: "@vsoch"
 github: "https://github.com/singularityhub/singularity-hpc/blob/main/registry/ghcr.io/autamus/omega-h/container.yaml"
-updated_at: "2021-07-20 19:57:03.581263"
+updated_at: "2021-07-20 20:00:00.032133"
 container_url: "https://github.com/orgs/autamus/packages/container/package/omega-h"
 
 versions:
